Switch the default dataset for mcns_* functions

choose_mcns_dataset This sets the default dataset used by all mcns_* functions. It is the recommended way to access malecns snapshots. Unlike choose_mcns it does not permanently change the default dataset used when callers use functions from the malevnc package (such as malevnc::manc_xyz2bodyid) directly.

Usage

choose_mcns_dataset(dataset = "male-cns:v0.9")

Arguments

dataset

The name of the dataset as reported in Clio e.g. CNS, etc

See also

Examples

if (FALSE) { # \dontrun{
# use the v0.9 snapshot for the rest of this R session
choose_mcns_dataset("male-cns:v0.9")
# use production for the rest of this R session
choose_mcns_dataset("CNS")
} # }