malecns: Access to the latest 'Janelia FlyEM' datasets

Provides access to the latest 'Janelia FlyEM' datasets by providing a thin wrapper around the 'malevnc' package.

Package Options

malecns.dataset This is to used to keep track of the active dataset. At present it is only for expert use and it is not recommended that you change this manually.

Bridging registrations

Philipp Schlegel has made bridging registrations using bigwarp and the presynapse predictions for the male half brain and male cns. See slack for details.

See also

Author

Maintainer: Gregory Jefferis jefferis@gmail.com (ORCID)

Examples

# \donttest{
options()[grepl("^malecns", names(options()))]
#> $malecns.dataset
#> [1] "CNS"
#> 
# }
if (FALSE) {
dr_malecns()
}

# \donttest{
library(nat.templatebrains)
xform_brain(cbind(443344, 225172, 44920), sample = 'FAFB14',
  reference = 'malecns')
#>          [,1]     [,2]     [,3]
#> [1,] 302807.7 215023.5 117948.1

mcflm=system.file("landmarks/maleCNS_brain_FAFB_landmarks_um.csv", package = 'malecns')
mcflm=read.csv(mcflm)
head(mcflm)
#>   X   fafb_x    fafb_y    fafb_z    cns_x     cns_y    cns_z
#> 1 0 522.3291 183.04947  95.20722 388.0265 161.15978 167.5336
#> 2 1 454.3915  88.80129  95.36000 316.4307  61.91385 170.2850
#> 3 2 597.4928  98.03833  83.20296 448.9167  60.41804 171.1397
#> 4 3 520.0082 210.97573 125.01538 386.9356 193.88847 208.3717
#> 5 4 414.5413 169.81426 155.26571 274.9307 152.79578 222.6996
#> 6 5 645.5330 176.75863 134.32295 501.8497 141.52257 219.8995
if (FALSE) {
library(nat.jrcbrains)
da1.hb=neuprintr::neuprint_read_neurons('/DA1.*lPN',
  conn=neuprintr::neuprint_login(server='https://neuprint.janelia.org',
    dataset = 'hemibrain:v1.2.1'))
# nb hemibrain neurons comes in 8nm raw voxel coordinates not microns
da1.hb.mcns=xform_brain(da1.hb*(8/1000), sample='JRCFIB2018F', reference='malecns')
# read in a male cns DA1 neuron
da1.1=read_mcns_meshes(11996, units='nm')
nclear3d()
plot3d(da1.hb.mcns, col='cyan')
plot3d(da1.1, col='red')
plot3d(malecns.surf, alpha=.1)
}
# }