Surface models of the malecns brain (in nm)

malecns_shell.surf is a surface model of the cortex (rind) of the malecns brain (in nm). It was downloaded from a link provided by Stuart Berg (of FlyEM).

malecns.surf is a surface model of the malecns neuropil generated by Philipp Schlegel in May 2022 based on the location of T bars in the dataset. See slack for details.

JRCFIB2022M.surf and malecns.surf are aliases

malecnsvnc_shell.surf is a surface model of the cortex (rind) of the malecns ventral nerve cord (in nm). It was downloaded from a link provided by Stuart Berg (of FlyEM) on 2023-11-18.

Usage

malecns_shell.surf

malecns.surf

JRCFIB2022M.surf

malecnsvnc_shell.surf

Format

An object of class hxsurf (inherits from list) of length 4.

An object of class hxsurf (inherits from list) of length 4.

An object of class hxsurf (inherits from list) of length 4.

An object of class hxsurf (inherits from list) of length 4.

Examples

# \donttest{
library(nat)
boundingbox(malecns_shell.surf)
#>          [,1]     [,2]     [,3]
#> [1,]  16898.1  31849.1  26324.4
#> [2,] 750590.0 441841.0 460357.0
#> attr(,"class")
#> [1] "boundingbox"
# }
if (FALSE) { # \dontrun{
plot3d(malecns_shell.surf, alpha=.2, col='grey')
# in microns
plot3d(malecns_shell.surf/1e3, alpha=.2, col='grey')
# in raw voxels
plot3d(malecns_shell.surf*(8/1e3), alpha=.2, col='grey')
# mesh3d gives more flexibility
plot3d(as.mesh3d(malecns_shell.surf), alpha=.2, col='grey', type='wire', add=T)
} # }
# \donttest{
library(nat)
boundingbox(malecns.surf)
#>          [,1]     [,2]     [,3]
#> [1,]  45319.4  38146.7  79817.4
#> [2,] 726917.9 417784.6 341814.2
#> attr(,"class")
#> [1] "boundingbox"
# }
if (FALSE) { # \dontrun{
plot3d(malecns.surf, alpha=.2, col='grey')
# in microns
plot3d(malecns.surf/1e3, alpha=.2, col='grey')
# in raw voxels
plot3d(malecns.surf*(8/1e3), alpha=.2, col='grey')
# mesh3d gives more flexibility
plot3d(as.mesh3d(malecns.surf), alpha=.2, col='grey', type='wire', add=T)
} # }
# \donttest{
library(nat)
boundingbox(malecnsvnc_shell.surf)
#>        [,1]   [,2]    [,3]
#> [1,] 236909 307526  415488
#> [2,] 551724 574158 1076990
#> attr(,"class")
#> [1] "boundingbox"
# }
if (FALSE) { # \dontrun{
nclear3d()
plot3d(malecns.surf, col='blue')
wire3d(malecns_shell.surf, col='grey')
plot3d(malecnsvnc_shell.surf, alpha=.2, col='green')
} # }