Connectivity query for CNS neurons

Connectivity query for CNS neurons

Usage

mcns_connection_table(
  ids,
  partners = c("inputs", "outputs"),
  moredetails = c("group", "class", "somaSide"),
  summary = FALSE,
  threshold = 1L,
  conn = mcns_neuprint(),
  ...
)

Arguments

ids

A set of body ids (see manc_ids for a range of ways to specify these).

partners

Either inputs or outputs. Redundant with prepost, but probably clearer.

moredetails

Either a logical (to add all fields when TRUE) or a character vector naming additional fields returned by mcns_neuprint_meta that will be added to the results data.frame.

summary

Whether to summarise results per partner when giving multiple query neurons

threshold

Only return partners >= to an integer value. Default of 1 returns all partners. This threshold will be applied to the ROI weight when the roi argument is specified, otherwise to the whole neuron.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

...

additional arguments passed to neuprint_connection_table

Value

A data.frame

Examples

# \donttest{
joffrey.id=mcns_xyz2bodyid(cbind(24590, 13816, 26102)+4096, node = 'neuprint')
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
joffrey.us=mcns_connection_table(joffrey.id, partners = 'in')
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
joffrey.uss=mcns_connection_table(joffrey.id, partners = 'in', summary=TRUE)
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
# }
if (FALSE) { # \dontrun{
# open top 10 partners in neuroglancer,
# NB segmentation / meshes to match neuprint
mcns_scene(joffrey.uss$partner[1:10], open = TRUE, node='neuprint')
} # }