Predict the group of neurons using instance or type information

Predict the group of neurons using instance or type information

Usage

mcns_predict_group(
  ids,
  method = c("auto", "group", "instance", "type", "all"),
  badtypes = c(NA, "", "Lamina_R1-R6", "Descending", "KC", "ER", "LC", "PB",
    "Ascending Interneuron", "Delta", "P1_L candidate", "LT", "MeMe", "PFGs", "Mi", "VT",
    "ML", "EL", "FB", "Dm", "DNp", "FC", "OL", "T", "Y")
)

Arguments

ids

Body ids in any form understood by mcns_ids. If you have a metadata dataframe as returned by mcns_neuprint_meta then this is ideal as that function is called under the hood.

method

A string specifying which of 3 methods to use to identify the group. "all" means to return all 3, while "auto" (the default) means to look at each method in turn successively filling in missing group values.

badtypes

Values of the type column which should be ignored for the purposes of defining cell type groups. This will be because they contain bad values or because the types are too broad to be very useful.

Value

For method="all" a dataframe as returned by

mcns_neuprint_meta with additional columns

instance_group and type_group. Otherwise a numeric vector.

Details

Grouping information for neurons in the male cns is presently scattered in several locations. These include the numeric group field, the type field or the instance field. If the type field has the same value, the neurons should form a group. However there are some values

See also

Examples

# \donttest{
library(dplyr)
# return all body ids with a group type or instance
tig_ids=mcns_ids('where:exists(n.group) OR exists(n.type) OR exists (n.instance)')
allg=mcns_predict_group(tig_ids, method = 'all')
# neurons where the recorded group and instance group disagree
allg %>% filter(!is.na(group) & !is.na(instance_group) & group!=instance_group)
#>    bodyid post  pre downstream upstream synweight   class  group hemibrainType
#> 1   24495 1652  369       2347     1652      3999    <NA>      0          <NA>
#> 2   29418 1591  333       2013     1591      3604    <NA>      0          <NA>
#> 3  111782 1120  340       2510     1120      3630    <NA>      0          <NA>
#> 4   82845  492  156       1312      492      1804    <NA>      0          <NA>
#> 5  104734  344  153       1315      344      1659    <NA>      0          <NA>
#> 6   27325 1908  265       1661     1908      3569    <NA>      0          <NA>
#> 7  134842  471  133        950      471      1421    <NA>      0          <NA>
#> 8  547498  390  133       1018      390      1408    <NA>      0          <NA>
#> 9   16872 1134  372       2515     1134      3649    <NA>      0          <NA>
#> 10 104089  495  172       1260      495      1755    <NA>      0          <NA>
#> 11  20666 1778  598       5150     1778      6928    <NA>      0          <NA>
#> 12  29577 1154  406       2991     1154      4145    <NA>      0          <NA>
#> 13  30984  909  240       1588      909      2497    <NA>      0          <NA>
#> 14  22752 2304  470       3977     2304      6281    <NA>      0          <NA>
#> 15  84191  740  237       2033      740      2773    <NA>      0          <NA>
#> 16 127917  526  192       1317      526      1843    <NA>      0          <NA>
#> 17  31509 1789  405       2592     1789      4381    <NA>      0          <NA>
#> 18  19113  717  729       6918      717      7635 central      0         CL258
#> 19 107544  930  223       1459      930      2389    <NA>      0          <NA>
#> 20 109887  404  181       1265      404      1669    <NA>      0          <NA>
#> 21  42774  827  198       1257      827      2084    <NA>      0          <NA>
#> 22  60026  448  257        942      448      1390    <NA>      0          <NA>
#> 23  49620  579  214       1762      579      2341    <NA>      0          <NA>
#> 24  62916  497  307       2453      497      2950    <NA>      0          <NA>
#> 25  22523 1847  507       4510     1847      6357    <NA>      0          <NA>
#> 26 201131  436  405       3499      436      3935    <NA>      0          <NA>
#> 27  54757  357  215        765      357      1122    <NA>      0          <NA>
#> 28  18646  654  254       1824      654      2478    <NA>      0          <NA>
#> 29 130029  841  153       1107      841      1948    <NA>      0          <NA>
#> 30  93468  456  170       1486      456      1942    <NA>      0          <NA>
#> 31 161331  360   99        654      360      1014    <NA>      0          <NA>
#> 32  54333  647  249       1764      647      2411    <NA>      0          <NA>
#> 33  81265  574  428       3512      574      4086    <NA>      0          <NA>
#> 34 131000  364  184       1409      364      1773    <NA>      0          <NA>
#> 35  67228  439  108        717      439      1156    <NA>      0          <NA>
#> 36  95690  246  163       1372      246      1618    <NA>      0          <NA>
#> 37 109830  544  246       2140      544      2684    <NA>      0          <NA>
#> 38  61578  537  219       1659      537      2196    <NA>      0          <NA>
#> 39  19127 2561  658       5480     2561      8041    <NA>      0          <NA>
#> 40  69966  750  865       8211      750      8961 central      0         CL258
#> 41 240951  642  211       1390      642      2032    <NA>      0          <NA>
#> 42  81760  715  218       1550      715      2265    <NA>      0          <NA>
#> 43  53587  484  136        888      484      1372    <NA>      0          <NA>
#> 44 107894  408  127        870      408      1278    <NA>      0          <NA>
#> 45 156523  377  126        692      377      1069    <NA>      0          <NA>
#> 46  84895  547  423       3550      547      4097    <NA>      0          <NA>
#> 47 122986  631  180       1342      631      1973    <NA>      0          <NA>
#> 48 126806  478  412       3492      478      3970    <NA>      0          <NA>
#> 49 124940  809  166       1059      809      1868    <NA>      0          <NA>
#> 50  15899  763  807       7575      763      8338 central      0         CL258
#> 51  71977  508  215       1349      508      1857    <NA>      0          <NA>
#> 52  34350 1654  318       2528     1654      4182    <NA>      0          <NA>
#> 53 187933  407  151       1233      407      1640    <NA>      0          <NA>
#> 54 103476  454  217       1513      454      1967    <NA>      0          <NA>
#> 55 165612  660  264       1939      660      2599    <NA>      0          <NA>
#> 56  75501  650  168        994      650      1644    <NA>      0          <NA>
#> 57 119700  880  151       1049      880      1929    <NA>      0          <NA>
#> 58 132353  324  177       1261      324      1585    <NA>      0          <NA>
#> 59 536253  428  159       1037      428      1465    <NA>      0          <NA>
#> 60 132840  410  204       1473      410      1883    <NA>      0          <NA>
#> 61 169402  345  149       1325      345      1670 central      0          <NA>
#> 62 104203  419  401       3397      419      3816    <NA>      0          <NA>
#> 63 902249  432  110        712      432      1144    <NA>      0          <NA>
#> 64 543906  446  175       1240      446      1686    <NA>      0          <NA>
#> 65 154952  300  119        922      300      1222    <NA>      0          <NA>
#> 66 127111  433  178       1180      433      1613    <NA>      0          <NA>
#> 67 140612  422  129        973      422      1395    <NA>      0          <NA>
#> 68 133153  466  404       3424      466      3890    <NA>      0          <NA>
#> 69 531400  740  691       6343      740      7083 central      0         CL258
#> 70 523568  667  206       1334      667      2001    <NA>      0          <NA>
#> 71 124421  747  234       1651      747      2398    <NA>      0          <NA>
#> 72  70921  426  125        877      426      1303    <NA>      0          <NA>
#> 73 513970 2239 1415      10805     2239     13044    <NA>  12906          <NA>
#> 74 513506 3960  645       4766     3960      8726    <NA>  12980          <NA>
#> 75 513117 3742  672       4894     3742      8636    <NA>  12980          <NA>
#> 76  17153 4331  618       4685     4331      9016    <NA>  12980          <NA>
#> 77  45643 3095  783       5041     3095      8136    <NA>  13160          <NA>
#> 78  13240 2709  812       5693     2709      8402    <NA>  13240          <NA>
#> 79 524005 3477  813       5619     3477      9096    <NA>  13240          <NA>
#> 80 514252 3614 1095       7456     3614     11070 central  13264       AVLP488
#> 81  14036 4386 1166       7836     4386     12222 central  13264       AVLP488
#> 82  26331 2346  724       6918     2346      9264 central  13983       PVLP118
#> 83  20488 1757  779       6996     1757      8753 central  13983       PVLP118
#> 84 518549 1688  256       2282     1688      3970    <NA>  14751          <NA>
#> 85  99043 2823  438       2944     2823      5767    <NA>  17436          <NA>
#> 86 525548  879  345       2543      879      3422    <NA>  18772          <NA>
#> 87  34160  667  370       2435      667      3102    <NA>  18772          <NA>
#> 88  22315 1763  414       3402     1763      5165    <NA>  20447          <NA>
#> 89  23278  813  309       2067      813      2880    <NA>  33024          <NA>
#> 90 551544  930  318       2133      930      3063    <NA>  33024          <NA>
#> 91  36117 1387  368       2714     1387      4101    <NA>  36177          <NA>
#> 92  39333 1423  342       2434     1423      3857    <NA>  36177          <NA>
#> 93 100167  697  313       2748      697      3445    <NA>  50866          <NA>
#> 94  10474 1724  668       4165     1724      5889    <NA> 109931           DGI
#> 95  11383 1589  574       3432     1589      5021    <NA> 109931           DGI
#>        name mancBodyid somaSide           statusLabel    type      somaLocation
#> 1   16872_L         NA        L Prelim Roughly traced    <NA> 71947,25189,18444
#> 2   16872_L         NA        L Prelim Roughly traced    <NA> 72880,23008,18746
#> 3  111782_R         NA        R        Primary Anchor    <NA> 23510,22998,20768
#> 4   82845_R         NA        R        Roughly traced    <NA> 35336,22210,12972
#> 5   82845_R         NA        R        Roughly traced    <NA> 34944,21388,13006
#> 6   16872_R         NA        R Prelim Roughly traced    <NA> 22622,22486,21732
#> 7   82845_L         NA        L        Roughly traced    <NA> 65298,19670,14782
#> 8   82845_L         NA        L        Roughly traced    <NA> 63224,19614,14370
#> 9   16872_L         NA        L Prelim Roughly traced    <NA> 73152,24306,18560
#> 10  42774_R         NA        R        Roughly traced    <NA> 32410,19278,17168
#> 11  18267_L         NA        L Prelim Roughly traced    <NA> 68300,35690,24796
#> 12  29577_R         NA        R Prelim Roughly traced    <NA> 32922,18784,36854
#> 13  30984_R         NA        R Prelim Roughly traced    <NA> 41206,21989,37912
#> 14  10021_R         NA        R Prelim Roughly traced    <NA> 29638,36896,25030
#> 15  49620_L         NA        L Prelim Roughly traced    <NA> 65952,18837,16578
#> 16  42774_L         NA        L Prelim Roughly traced    <NA> 65656,17934,16495
#> 17  16872_R         NA        R Prelim Roughly traced    <NA> 22468,23226,21268
#> 18  15899_R         NA        R        Roughly traced   CL258 37175,30477,37911
#> 19  42774_R         NA        R        Roughly traced    <NA> 32604,19206,17540
#> 20  18646_R         NA        R        Roughly traced    <NA> 31952,19320,17179
#> 21  42774_R         NA        R        Roughly traced    <NA> 31002,18496,17600
#> 22  54757_R         NA        R Prelim Roughly traced    <NA> 27264,19628,20480
#> 23  49620_L         NA        L Prelim Roughly traced    <NA> 65428,19010,16586
#> 24  62916_R         NA        R Prelim Roughly traced    <NA> 31835,19626,17507
#> 25  18267_R         NA        R Prelim Roughly traced    <NA> 29869,35927,24087
#> 26  84895_L         NA        L Prelim Roughly traced    <NA> 66324,18552,16910
#> 27  54757_R         NA        R Prelim Roughly traced    <NA> 30114,19870,18613
#> 28  18646_R         NA        R        Roughly traced    <NA> 31398,19404,17322
#> 29 119700_R         NA        R        Roughly traced    <NA> 36720,21152,13300
#> 30  82845_R         NA        R        Roughly traced    <NA> 35176,20478,13764
#> 31  53587_L         NA        L        Roughly traced    <NA> 64660,21352,13282
#> 32  18646_L         NA        L        Roughly traced    <NA> 64542,19174,15526
#> 33  30900_R         NA        R Prelim Roughly traced    <NA> 32498,20354,16520
#> 34  18646_L         NA        L Prelim Roughly traced    <NA> 64634,18284,16768
#> 35  53587_R         NA        R        Roughly traced    <NA> 34859,20224,14088
#> 36  82845_R         NA        R        Roughly traced    <NA> 36421,22180,13023
#> 37  49620_R         NA        R Prelim Roughly traced    <NA> 31890,21501,15937
#> 38  18646_R         NA        R        Roughly traced    <NA> 32298,19658,17404
#> 39  18267_R         NA        R Prelim Roughly traced    <NA> 30538,36466,24606
#> 40  15899_L         NA        L        Roughly traced   CL258 58065,22436,38063
#> 41  75501_R         NA        R        Roughly traced    <NA> 35346,21310,13252
#> 42  18646_R         NA        R        Roughly traced    <NA> 32490,18632,16784
#> 43  53587_L         NA        L        Roughly traced    <NA> 65628,20616,14554
#> 44  53587_R         NA        R        Roughly traced    <NA> 35132,20648,13396
#> 45  53587_L         NA        L        Roughly traced    <NA> 64839,20223,14979
#> 46  84895_R         NA        R Prelim Roughly traced    <NA> 32820,21208,14590
#> 47 119700_L         NA        L Prelim Roughly traced    <NA> 63012,20356,13214
#> 48  84895_R         NA        R Prelim Roughly traced    <NA> 33154,20522,15582
#> 49 119700_R         NA        R        Roughly traced    <NA> 38710,21048,14070
#> 50  15899_L         NA        L        Roughly traced   CL258 57586,22601,38255
#> 51  71977_R         NA        R        Roughly traced    <NA> 32860,18488,17022
#> 52  34350_R         NA        R        Roughly traced    <NA> 33543,20006,15633
#> 53  82845_L         NA        L        Roughly traced    <NA> 64976,20602,13712
#> 54  18646_L         NA        L        Roughly traced    <NA> 64444,18830,16526
#> 55  42774_L         NA        L        Roughly traced    <NA> 67781,17475,17008
#> 56  75501_L         NA        L        Roughly traced    <NA> 68838,15616,17572
#> 57 119700_L         NA        L        Roughly traced    <NA> 67798,16223,16712
#> 58  42774_L         NA        L        Roughly traced    <NA> 65398,18762,15292
#> 59  75501_L         NA        L Prelim Roughly traced    <NA> 67334,16918,16608
#> 60  18646_L         NA        L Prelim Roughly traced    <NA> 67080,17446,16816
#> 61  82845_L         NA     <NA>        Roughly traced    <NA>                  
#> 62  84895_L         NA        L Prelim Roughly traced    <NA> 65720,19526,16770
#> 63  53587_R         NA        R        Roughly traced    <NA> 35192,20163,15216
#> 64  75501_L         NA        L        Roughly traced    <NA> 67068,16671,16135
#> 65  82845_L         NA        L        Roughly traced    <NA> 65034,19624,14500
#> 66  75501_R         NA        R Prelim Roughly traced    <NA> 36476,21418,13650
#> 67  82845_R         NA        R        Roughly traced    <NA> 35050,20952,13000
#> 68 133153_R         NA        R Prelim Roughly traced    <NA> 32604,21350,15692
#> 69  15899_R         NA        R        Roughly traced   CL258 38201,23462,37894
#> 70  42774_R         NA        L        Roughly traced    <NA> 64676,17919,16649
#> 71  42774_L         NA        L        Roughly traced    <NA> 65542,17088,16234
#> 72  53587_R         NA        R        Roughly traced    <NA> 35028,19879,14750
#> 73  11234_R         NA        R Prelim Roughly traced    <NA> 35228,38257,33280
#> 74  16251_L         NA        L        Roughly traced    <NA> 72396,20028,18998
#> 75  16251_L         NA        L        Roughly traced    <NA> 73100,21418,18478
#> 76  16251_R         NA        R        Roughly traced    <NA> 22596,26628,20862
#> 77  13798_L         NA        L Prelim Roughly traced    <NA> 63818,20952,35610
#> 78 513691_R         NA        R        Roughly traced    <NA> 36360,15082,14908
#> 79 513691_L         NA        L        Roughly traced    <NA> 56828,15422,13200
#> 80  14036_L         NA        L        Roughly traced AVLP488 59551,21867,37206
#> 81  14036_R         NA        R        Roughly traced AVLP488 35630,22862,36560
#> 82  20488_R         NA        R        Roughly traced PVLP118 37516,22541,37905
#> 83  20488_L         NA        L        Roughly traced PVLP118 58524,22274,37456
#> 84  34357_L         NA        L        Roughly traced    <NA> 65438,19072,14424
#> 85  19554_L         NA        L        Roughly traced    <NA> 56283,16344,11915
#> 86  19728_R         NA        R Prelim Roughly traced    <NA> 41468,13240,30486
#> 87  19728_L         NA        L        Primary Anchor    <NA> 57336,11336,30984
#> 88  22315_R         NA        R        Roughly traced    <NA> 30548,18745,17855
#> 89  23278_L         NA        L Prelim Roughly traced    <NA> 54428,10488,31394
#> 90  23278_R         NA        R Prelim Roughly traced    <NA> 42169,11387,29955
#> 91  30347_L         NA        L Prelim Roughly traced    <NA> 47856,10668,15537
#> 92  30347_L         NA        L Prelim Roughly traced    <NA> 49496,10720,15502
#> 93  41527_R         NA        R Prelim Roughly traced    <NA>  41804,6601,24896
#> 94  10474_L         NA        L        Roughly traced     DGI 70794,15623,23717
#> 95  10474_R         NA        R        Roughly traced     DGI 23959,19995,24642
#>    status     voxels vnc-shell anteriorCervicalRootPosition
#> 1  Traced  520388878        NA                             
#> 2  Traced  484205534        NA                             
#> 3  Anchor  482766551        NA                             
#> 4  Traced  254030926        NA                             
#> 5  Traced  233042790        NA                             
#> 6  Traced  484518525        NA                             
#> 7  Traced  340270663        NA                             
#> 8  Traced  269692484        NA                             
#> 9  Traced  528604601        NA                             
#> 10 Traced  303094479        NA                             
#> 11 Traced  696748548        NA                             
#> 12 Traced  451849198        NA                             
#> 13 Traced  387684658        NA                             
#> 14 Traced  623820656        NA                             
#> 15 Traced  318955745        NA                             
#> 16 Traced  307535504        NA                             
#> 17 Traced  532622066        NA                             
#> 18 Traced  784168739        NA                             
#> 19 Traced  340389879        NA                             
#> 20 Traced  291404631        NA                             
#> 21 Traced  365375332        NA                             
#> 22 Traced  443272833        NA                             
#> 23 Traced  344471682        NA                             
#> 24 Traced  333205591        NA                             
#> 25 Traced  625802541        NA                             
#> 26 Traced  378268089        NA                             
#> 27 Traced  382244209        NA                             
#> 28 Traced  359602114        NA                             
#> 29 Traced  294184912        NA                             
#> 30 Traced  242105588        NA                             
#> 31 Traced  255937724        NA                             
#> 32 Traced  324847066        NA                             
#> 33 Traced  412677939        NA                             
#> 34 Traced  296539148        NA                             
#> 35 Traced  269023070        NA                             
#> 36 Traced  223926680        NA                             
#> 37 Traced  284974428        NA                             
#> 38 Traced  310290697        NA                             
#> 39 Traced  789103315        NA                             
#> 40 Traced  766845723        NA                             
#> 41 Traced  311347909        NA                             
#> 42 Traced  350299160        NA                             
#> 43 Traced  333963184        NA                             
#> 44 Traced  262586266        NA                             
#> 45 Traced  297833796        NA                             
#> 46 Traced  423887865        NA                             
#> 47 Traced  297764633        NA                             
#> 48 Traced  389282590        NA                             
#> 49 Traced  306775359        NA                             
#> 50 Traced  707344022        NA                             
#> 51 Traced  297081442        NA                             
#> 52 Traced  592870565        NA                             
#> 53 Traced  214646575        NA                             
#> 54 Traced  284008513        NA                             
#> 55 Traced  358462671        NA                             
#> 56 Traced  307505648        NA                             
#> 57 Traced  297777197        NA                             
#> 58 Traced  263048275        NA                             
#> 59 Traced  224493537        NA                             
#> 60 Traced  293830414        NA                             
#> 61 Traced  170480096        NA                             
#> 62 Traced  403076877        NA                             
#> 63 Traced  274850914        NA                             
#> 64 Traced  260123339        NA                             
#> 65 Traced  209232861        NA                             
#> 66 Traced  307035507        NA                             
#> 67 Traced  231769947        NA                             
#> 68 Traced  436444830        NA                             
#> 69 Traced  689772868        NA                             
#> 70 Traced  323199723        NA                             
#> 71 Traced  368506657        NA                             
#> 72 Traced  262078805        NA                             
#> 73 Traced 1070304879        NA                             
#> 74 Traced 1018937226        NA                             
#> 75 Traced 1067876448        NA                             
#> 76 Traced  916808515        NA                             
#> 77 Traced 1638296270        NA                             
#> 78 Traced 1275543596        NA                             
#> 79 Traced 1450377641        NA                             
#> 80 Traced 1317276665        NA                             
#> 81 Traced 1451113747        NA                             
#> 82 Traced  637648131        NA                             
#> 83 Traced  654066588        NA                             
#> 84 Traced  393127013        NA                             
#> 85 Traced  791223622        NA                             
#> 86 Traced  574257172        NA                             
#> 87 Anchor  559482213        NA                             
#> 88 Traced  607848947        NA                             
#> 89 Traced  518813212        NA                             
#> 90 Traced  566876726        NA                             
#> 91 Traced  639336494        NA                             
#> 92 Traced  615824072        NA                             
#> 93 Traced  351282129        NA                             
#> 94 Traced 3522321719        NA                             
#> 95 Traced 3475651903        NA                             
#>    posteriorCervicalRootPosition entryNerve mancGroup mancType synonyms
#> 1                                      <NA>        NA     <NA>     <NA>
#> 2                                      <NA>        NA     <NA>     <NA>
#> 3                                      <NA>        NA     <NA>     <NA>
#> 4                                      <NA>        NA     <NA>     <NA>
#> 5                                      <NA>        NA     <NA>     <NA>
#> 6                                      <NA>        NA     <NA>     <NA>
#> 7                                      <NA>        NA     <NA>     <NA>
#> 8                                      <NA>        NA     <NA>     <NA>
#> 9                                      <NA>        NA     <NA>     <NA>
#> 10                                     <NA>        NA     <NA>     <NA>
#> 11                                     <NA>        NA     <NA>     <NA>
#> 12                                     <NA>        NA     <NA>     <NA>
#> 13                                     <NA>        NA     <NA>     <NA>
#> 14                                     <NA>        NA     <NA>     <NA>
#> 15                                     <NA>        NA     <NA>     <NA>
#> 16                                     <NA>        NA     <NA>     <NA>
#> 17                                     <NA>        NA     <NA>     <NA>
#> 18                                     <NA>        NA     <NA>     <NA>
#> 19                                     <NA>        NA     <NA>     <NA>
#> 20                                     <NA>        NA     <NA>     <NA>
#> 21                                     <NA>        NA     <NA>     <NA>
#> 22                                     <NA>        NA     <NA>     <NA>
#> 23                                     <NA>        NA     <NA>     <NA>
#> 24                                     <NA>        NA     <NA>     <NA>
#> 25                                     <NA>        NA     <NA>     <NA>
#> 26                                     <NA>        NA     <NA>     <NA>
#> 27                                     <NA>        NA     <NA>     <NA>
#> 28                                     <NA>        NA     <NA>     <NA>
#> 29                                     <NA>        NA     <NA>     <NA>
#> 30                                     <NA>        NA     <NA>     <NA>
#> 31                                     <NA>        NA     <NA>     <NA>
#> 32                                     <NA>        NA     <NA>     <NA>
#> 33                                     <NA>        NA     <NA>     <NA>
#> 34                                     <NA>        NA     <NA>     <NA>
#> 35                                     <NA>        NA     <NA>     <NA>
#> 36                                     <NA>        NA     <NA>     <NA>
#> 37                                     <NA>        NA     <NA>     <NA>
#> 38                                     <NA>        NA     <NA>     <NA>
#> 39                                     <NA>        NA     <NA>     <NA>
#> 40                                     <NA>        NA     <NA>     <NA>
#> 41                                     <NA>        NA     <NA>     <NA>
#> 42                                     <NA>        NA     <NA>     <NA>
#> 43                                     <NA>        NA     <NA>     <NA>
#> 44                                     <NA>        NA     <NA>     <NA>
#> 45                                     <NA>        NA     <NA>     <NA>
#> 46                                     <NA>        NA     <NA>     <NA>
#> 47                                     <NA>        NA     <NA>     <NA>
#> 48                                     <NA>        NA     <NA>     <NA>
#> 49                                     <NA>        NA     <NA>     <NA>
#> 50                                     <NA>        NA     <NA>     <NA>
#> 51                                     <NA>        NA     <NA>     <NA>
#> 52                                     <NA>        NA     <NA>     <NA>
#> 53                                     <NA>        NA     <NA>     <NA>
#> 54                                     <NA>        NA     <NA>     <NA>
#> 55                                     <NA>        NA     <NA>     <NA>
#> 56                                     <NA>        NA     <NA>     <NA>
#> 57                                     <NA>        NA     <NA>     <NA>
#> 58                                     <NA>        NA     <NA>     <NA>
#> 59                                     <NA>        NA     <NA>     <NA>
#> 60                                     <NA>        NA     <NA>     <NA>
#> 61                                     <NA>        NA     <NA>     <NA>
#> 62                                     <NA>        NA     <NA>     <NA>
#> 63                                     <NA>        NA     <NA>     <NA>
#> 64                                     <NA>        NA     <NA>     <NA>
#> 65                                     <NA>        NA     <NA>     <NA>
#> 66                                     <NA>        NA     <NA>     <NA>
#> 67                                     <NA>        NA     <NA>     <NA>
#> 68                                     <NA>        NA     <NA>     <NA>
#> 69                                     <NA>        NA     <NA>     <NA>
#> 70                                     <NA>        NA     <NA>     <NA>
#> 71                                     <NA>        NA     <NA>     <NA>
#> 72                                     <NA>        NA     <NA>     <NA>
#> 73                                     <NA>        NA     <NA>     <NA>
#> 74                                     <NA>        NA     <NA>     <NA>
#> 75                                     <NA>        NA     <NA>     <NA>
#> 76                                     <NA>        NA     <NA>     <NA>
#> 77                                     <NA>        NA     <NA>     <NA>
#> 78                                     <NA>        NA     <NA>     <NA>
#> 79                                     <NA>        NA     <NA>     <NA>
#> 80                                     <NA>        NA     <NA>     <NA>
#> 81                                     <NA>        NA     <NA>     <NA>
#> 82                                     <NA>        NA     <NA>     <NA>
#> 83                                     <NA>        NA     <NA>     <NA>
#> 84                                     <NA>        NA     <NA>     <NA>
#> 85                                     <NA>        NA     <NA>     <NA>
#> 86                                     <NA>        NA     <NA>     <NA>
#> 87                                     <NA>        NA     <NA>     <NA>
#> 88                                     <NA>        NA     <NA>     <NA>
#> 89                                     <NA>        NA     <NA>     <NA>
#> 90                                     <NA>        NA     <NA>     <NA>
#> 91                                     <NA>        NA     <NA>     <NA>
#> 92                                     <NA>        NA     <NA>     <NA>
#> 93                                     <NA>        NA     <NA>     <NA>
#> 94                                     <NA>        NA     <NA>     <NA>
#> 95                                     <NA>        NA     <NA>     <NA>
#>    superClass cellClass cellSubClass rootSide flywireType assignedOlHex1
#> 1        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 2        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 3        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 4        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 5        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 6        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 7        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 8        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 9        <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 10       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 11       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 12       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 13       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 14       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 15       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 16       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 17       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 18       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 19       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 20       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 21       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 22       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 23       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 24       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 25       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 26       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 27       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 28       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 29       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 30       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 31       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 32       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 33       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 34       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 35       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 36       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 37       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 38       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 39       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 40       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 41       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 42       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 43       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 44       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 45       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 46       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 47       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 48       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 49       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 50       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 51       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 52       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 53       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 54       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 55       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 56       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 57       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 58       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 59       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 60       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 61       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 62       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 63       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 64       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 65       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 66       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 67       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 68       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 69       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 70       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 71       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 72       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 73       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 74       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 75       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 76       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 77       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 78       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 79       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 80    central      <NA>         <NA>     <NA>        <NA>             NA
#> 81    central      <NA>         <NA>     <NA>        <NA>             NA
#> 82       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 83       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 84       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 85       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 86       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 87       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 88       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 89       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 90       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 91       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 92       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 93       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 94       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#> 95       <NA>      <NA>         <NA>     <NA>        <NA>             NA
#>    assignedOlHex2 hemibrainBodyId cellBodyFiber  soma instance_group type_group
#> 1              NA            <NA>          <NA>  TRUE          16872         NA
#> 2              NA            <NA>          <NA>  TRUE          16872         NA
#> 3              NA            <NA>          <NA>  TRUE         111782         NA
#> 4              NA            <NA>          <NA>  TRUE          82845         NA
#> 5              NA            <NA>          <NA>  TRUE          82845         NA
#> 6              NA            <NA>          <NA>  TRUE          16872         NA
#> 7              NA            <NA>          <NA>  TRUE          82845         NA
#> 8              NA            <NA>          <NA>  TRUE          82845         NA
#> 9              NA            <NA>          <NA>  TRUE          16872         NA
#> 10             NA            <NA>          <NA>  TRUE          42774         NA
#> 11             NA            <NA>          <NA>  TRUE          18267         NA
#> 12             NA            <NA>          <NA>  TRUE          29577         NA
#> 13             NA            <NA>          <NA>  TRUE          30984         NA
#> 14             NA            <NA>          <NA>  TRUE          10021         NA
#> 15             NA            <NA>          <NA>  TRUE          49620         NA
#> 16             NA            <NA>          <NA>  TRUE          42774         NA
#> 17             NA            <NA>          <NA>  TRUE          16872         NA
#> 18             NA            <NA>          <NA>  TRUE          15899      15899
#> 19             NA            <NA>          <NA>  TRUE          42774         NA
#> 20             NA            <NA>          <NA>  TRUE          18646         NA
#> 21             NA            <NA>          <NA>  TRUE          42774         NA
#> 22             NA            <NA>          <NA>  TRUE          54757         NA
#> 23             NA            <NA>          <NA>  TRUE          49620         NA
#> 24             NA            <NA>          <NA>  TRUE          62916         NA
#> 25             NA            <NA>          <NA>  TRUE          18267         NA
#> 26             NA            <NA>          <NA>  TRUE          84895         NA
#> 27             NA            <NA>          <NA>  TRUE          54757         NA
#> 28             NA            <NA>          <NA>  TRUE          18646         NA
#> 29             NA            <NA>          <NA>  TRUE         119700         NA
#> 30             NA            <NA>          <NA>  TRUE          82845         NA
#> 31             NA            <NA>          <NA>  TRUE          53587         NA
#> 32             NA            <NA>          <NA>  TRUE          18646         NA
#> 33             NA            <NA>          <NA>  TRUE          30900         NA
#> 34             NA            <NA>          <NA>  TRUE          18646         NA
#> 35             NA            <NA>          <NA>  TRUE          53587         NA
#> 36             NA            <NA>          <NA>  TRUE          82845         NA
#> 37             NA            <NA>          <NA>  TRUE          49620         NA
#> 38             NA            <NA>          <NA>  TRUE          18646         NA
#> 39             NA            <NA>          <NA>  TRUE          18267         NA
#> 40             NA            <NA>          <NA>  TRUE          15899      15899
#> 41             NA            <NA>          <NA>  TRUE          75501         NA
#> 42             NA            <NA>          <NA>  TRUE          18646         NA
#> 43             NA            <NA>          <NA>  TRUE          53587         NA
#> 44             NA            <NA>          <NA>  TRUE          53587         NA
#> 45             NA            <NA>          <NA>  TRUE          53587         NA
#> 46             NA            <NA>          <NA>  TRUE          84895         NA
#> 47             NA            <NA>          <NA>  TRUE         119700         NA
#> 48             NA            <NA>          <NA>  TRUE          84895         NA
#> 49             NA            <NA>          <NA>  TRUE         119700         NA
#> 50             NA            <NA>          <NA>  TRUE          15899      15899
#> 51             NA            <NA>          <NA>  TRUE          71977         NA
#> 52             NA            <NA>          <NA>  TRUE          34350         NA
#> 53             NA            <NA>          <NA>  TRUE          82845         NA
#> 54             NA            <NA>          <NA>  TRUE          18646         NA
#> 55             NA            <NA>          <NA>  TRUE          42774         NA
#> 56             NA            <NA>          <NA>  TRUE          75501         NA
#> 57             NA            <NA>          <NA>  TRUE         119700         NA
#> 58             NA            <NA>          <NA>  TRUE          42774         NA
#> 59             NA            <NA>          <NA>  TRUE          75501         NA
#> 60             NA            <NA>          <NA>  TRUE          18646         NA
#> 61             NA            <NA>          <NA> FALSE          82845         NA
#> 62             NA            <NA>          <NA>  TRUE          84895         NA
#> 63             NA            <NA>          <NA>  TRUE          53587         NA
#> 64             NA            <NA>          <NA>  TRUE          75501         NA
#> 65             NA            <NA>          <NA>  TRUE          82845         NA
#> 66             NA            <NA>          <NA>  TRUE          75501         NA
#> 67             NA            <NA>          <NA>  TRUE          82845         NA
#> 68             NA            <NA>          <NA>  TRUE         133153         NA
#> 69             NA            <NA>          <NA>  TRUE          15899      15899
#> 70             NA            <NA>          <NA>  TRUE          42774         NA
#> 71             NA            <NA>          <NA>  TRUE          42774         NA
#> 72             NA            <NA>          <NA>  TRUE          53587         NA
#> 73             NA            <NA>          <NA>  TRUE          11234         NA
#> 74             NA            <NA>          <NA>  TRUE          16251         NA
#> 75             NA            <NA>          <NA>  TRUE          16251         NA
#> 76             NA            <NA>          <NA>  TRUE          16251         NA
#> 77             NA            <NA>          <NA>  TRUE          13798         NA
#> 78             NA            <NA>          <NA>  TRUE         513691         NA
#> 79             NA            <NA>          <NA>  TRUE         513691         NA
#> 80             NA            <NA>          <NA>  TRUE          14036      13264
#> 81             NA            <NA>          <NA>  TRUE          14036      13264
#> 82             NA            <NA>          <NA>  TRUE          20488      13983
#> 83             NA            <NA>          <NA>  TRUE          20488      13983
#> 84             NA            <NA>          <NA>  TRUE          34357         NA
#> 85             NA            <NA>          <NA>  TRUE          19554         NA
#> 86             NA            <NA>          <NA>  TRUE          19728         NA
#> 87             NA            <NA>          <NA>  TRUE          19728         NA
#> 88             NA            <NA>          <NA>  TRUE          22315         NA
#> 89             NA            <NA>          <NA>  TRUE          23278         NA
#> 90             NA            <NA>          <NA>  TRUE          23278         NA
#> 91             NA            <NA>          <NA>  TRUE          30347         NA
#> 92             NA            <NA>          <NA>  TRUE          30347         NA
#> 93             NA            <NA>          <NA>  TRUE          41527         NA
#> 94             NA            <NA>          <NA>  TRUE          10474      10474
#> 95             NA            <NA>          <NA>  TRUE          10474      10474
# }
if (FALSE) {
# neurons where the recorded group and type group disagree
type_group_mismatch <- allg %>% filter(!is.na(group) & !is.na(type_group) & group!=type_group)
allg %>%
  filter(group %in% type_group_mismatch$group | type_group %in% type_group_mismatch$type_group) %>%
 select(bodyid, type, name, group, type_group, instance_group) %>%
 arrange(type, group) %>% View
}