Predict the cell type of male cns neurons from type and name/instance fields

Predict the cell type of male cns neurons from type and name/instance fields

Usage

mcns_predict_type(ids, method = c("auto", "instance", "type", "all"))

Arguments

ids

Body ids in any form understandable by mcns_ids

method

The prediction method to use (type, instance or auto, which uses type when available, instance otherwise). The special value of all adds 3 columns to the metadata data.frame type_t, type_i, type_a containing the results of each of the other methods.

Value

A data.frame when type='all', a character vector otherwise.

Details

For the time being when the instance is of the form 12345_L then 12345 will be returned as the type. This obviously isn't very useful for most purposes (use mcns_predict_group if you want the group).

See also

Examples

# \donttest{
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')
# Descending neurons with the wrong class
mnm.ti %>%
  mutate(ptype=mcns_predict_type(.)) %>%
  filter(grepl("DN[abdgpx]", ptype)) %>%
  filter(class!='Descending' | is.na(class))
#>       bodyid  pre post upstream downstream synweight status    statusLabel
#> 1      11693 1823 4106     4106      11835     15941 Anchor Primary Anchor
#> 2      40623  138  578      578       1205      1783 Anchor Primary Anchor
#> 3  451851712   40  262      262        150       412 Anchor Primary Anchor
#> 4     383627   54   42       42        267       309 Anchor Primary Anchor
#> 5      11442  142  789      789        672      1461 Anchor Primary Anchor
#> 6      11756  135  631      631        628      1259 Anchor Primary Anchor
#> 7      17373  750 2310     2310       6582      8892 Anchor Primary Anchor
#> 8      13432  163 1932     1932        818      2750 Anchor Primary Anchor
#> 9     517881  482 1847     1847       3161      5008 Anchor Primary Anchor
#> 10     13277  987 4413     4413       4830      9243 Anchor Primary Anchor
#> 11     16731  198 1559     1559       1091      2650 Anchor Primary Anchor
#> 12     15569  611  156      156       5877      6033 Anchor Primary Anchor
#> 13     15430  546  187      187       5329      5516 Anchor Primary Anchor
#> 14     15303  549  260      260       5308      5568 Anchor Primary Anchor
#> 15     14707  619  245      245       6011      6256 Anchor Primary Anchor
#> 16     30465  574  814      814       4312      5126 Anchor Primary Anchor
#> 17     10447 2250 5987     5987      13067     19054 Anchor Primary Anchor
#> 18     12151 1863 5368     5368      10738     16106 Anchor Primary Anchor
#> 19     12675  804 4293     4293       4671      8964 Anchor Primary Anchor
#>        voxels cropped         name type somaSide                 class
#> 1  4226774471   FALSE      DNd03_R <NA>        R                  <NA>
#> 2   293449447   FALSE DNg18_EM_2_R <NA>        R                  <NA>
#> 3  1549631143   FALSE      DNg28_L <NA>        L Ascending Interneuron
#> 4   688849981   FALSE      DNg28_L <NA>        L                  <NA>
#> 5  2666944137   FALSE      DNg28_R <NA>        R                  <NA>
#> 6  2543722078   FALSE      DNg28_R <NA>        R                  <NA>
#> 7   886506689   FALSE      DNg58_L <NA>        L                  <NA>
#> 8  1636270805   FALSE      DNg66_M <NA>        M                  <NA>
#> 9   708934341   FALSE      DNg77_R <NA>        R                  <NA>
#> 10 1692960301   FALSE      DNg86_R <NA>        R                  <NA>
#> 11  948589582   FALSE      DNg94_R <NA>        R                  <NA>
#> 12 1513951639   FALSE DNg_EM_276_L <NA>        L                  <NA>
#> 13 1206871763   FALSE DNg_EM_276_L <NA>        L                  <NA>
#> 14 1234442574   FALSE DNg_EM_276_R <NA>        R                  <NA>
#> 15 1280181024   FALSE DNg_EM_276_R <NA>        R                  <NA>
#> 16  889919740   FALSE DNg_EM_287_L <NA>        L Ascending Interneuron
#> 17 3648240832   FALSE DNg_EM_291_M <NA>        M                  <NA>
#> 18 2888765899   FALSE DNg_EM_291_M <NA>        M                  <NA>
#> 19 2254912041   FALSE DNg_EM_296_L <NA>        L                  <NA>
#>         somaLocation nucleusVoxels halfbrainBodyId group        neckLocation
#> 1  33076,34860,19466     202714968              NA 11486 46086, 44690, 43456
#> 2  44506,47110,33696      56527240              NA 40623                  NA
#> 3  58594,37438,18188     469034040              NA 11442 47696, 49760, 43456
#> 4  60600,38112,19198     336024768              NA 11442                  NA
#> 5  36362,39666,17786     278243408              NA 11442                  NA
#> 6                 NA            NA              NA 11442                  NA
#> 7                 NA            NA              NA 15454                  NA
#> 8  48562,46304,33743     149630784              NA 13432 49019, 45627, 43456
#> 9  38287,41225,24025      97670360              NA 18122                  NA
#> 10 47344,46898,34262     135066416              NA 13277                  NA
#> 11 36972,41000,30842     104085912              NA 16731 51587, 44740, 43456
#> 12                NA            NA              NA 14707 50698, 42125, 43456
#> 13                NA            NA              NA 14707 50173, 44612, 43456
#> 14                NA            NA              NA 14707 47661, 44451, 43456
#> 15                NA            NA              NA 14707 47484, 44750, 43456
#> 16 48699,47814,27930      88680384              NA 16856 47675, 42026, 43456
#> 17 47540,38029,41426     168864872              NA 10447 48393, 44141, 43456
#> 18 48922,43966,36072     140183488              NA 10447 48428, 44204, 43456
#> 19 52456,43014,18570     156314488              NA 11173 49043, 49491, 43456
#>    cellBodyFiber  soma      ptype
#> 1           <NA>  TRUE      DNd03
#> 2           <NA>  TRUE DNg18_EM_2
#> 3           <NA>  TRUE      DNg28
#> 4           <NA>  TRUE      DNg28
#> 5           <NA>  TRUE      DNg28
#> 6           <NA> FALSE      DNg28
#> 7           <NA> FALSE      DNg58
#> 8           <NA>  TRUE      DNg66
#> 9           <NA>  TRUE      DNg77
#> 10          <NA>  TRUE      DNg86
#> 11          <NA>  TRUE      DNg94
#> 12          <NA> FALSE DNg_EM_276
#> 13          <NA> FALSE DNg_EM_276
#> 14          <NA> FALSE DNg_EM_276
#> 15          <NA> FALSE DNg_EM_276
#> 16          <NA>  TRUE DNg_EM_287
#> 17          <NA>  TRUE DNg_EM_291
#> 18          <NA>  TRUE DNg_EM_291
#> 19          <NA>  TRUE DNg_EM_296
# }