Set the DVID type, instance or group for some malecns neurons

Set the DVID type, instance or group for some malecns neurons

Usage

mcns_set_dvid_annotations(
  ids,
  type = NULL,
  group = NULL,
  side = NULL,
  instance = T,
  user = getOption("malevnc.dvid_user"),
  ...
)

Arguments

ids

Body ids

type

Character vector specifying cell type e.g. "LHAD1g1"

group

One or more LR groups (ie candidate cell types) to apply. These should normally be the lowest bodyid of the group. Must be the same length as ids unless it has length 1.

side

Character vector specifying the side of each neuron ("L", "R" or "" when it cannot be specified)

instance

Character vector specifying instances (names) for neurons emph or a logical value where TRUE (the default) means to append the side to the type.

user

The DVID user. Defaults to options("malevnc.dvid_user"). neurons.

...

Additional arguments passed to malevnc:::manc_set_dvid_instance and thence to pbapply::pbmapply when there are multiple body ids.

Details

For the male CNS, the evolving standard seems to be to record the following fields (examples given for APL)

type='APL'

instance='APL_R'

group='APL_R'

type_user, instance_user, group_user='jefferisg' to accompany all of those.

Examples

if (FALSE) {
mcns_set_dvid_annotations(10297, type = 'LHAD1g1')
mcns_set_dvid_annotations(10977, type='APL', side='L', group = 10540)
# only set the LR group
mcns_set_dvid_annotations(ids=c(13115, 14424), group = 13115)
}