Read neuronal skeletons via neuprint

Read neuronal skeletons via neuprint

Usage

read_mcns_neurons(
  ids,
  connectors = F,
  units = c("nm", "raw", "microns"),
  heal.threshold = Inf,
  ...
)

Arguments

ids

Bodyids in any form compatible with malevnc::manc_ids

connectors

Whether to fetch synaptic connections for the neuron (default FALSE in contrast to neuprint_read_neurons).

units

Units of the returned neurons (default nm)

heal.threshold

The threshold for

...

Additional arguments passed to neuprint_read_neurons

Value

A neuronlist object containing one or more neuron objects.

See also

Examples

# nb convert
n30102=read_mcns_neurons(30102)
# neuronlist
n30102
#> 'neuronlist' containing 1 'neuprintneuron' object and 'data.frame' with 46 vars [159.6 kB]
boundingbox(n30102)
#>        [,1]   [,2]   [,3]
#> [1,] 168448  92160 197632
#> [2,] 279552 169984 274432
#> attr(,"class")
#> [1] "boundingbox"
# neuron
n30102[[1]]
#> 'neuron' with 2082 vertices in 1 tree and additional classes 'neuprintneuron', 'catmaidneuron'
if (FALSE) { # \dontrun{
nclear3d()
plot3d(malecns.surf, alpha=.1)
plot3d(n30102, lwd=2, soma=2000)
} # }