Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

library(dplyr)
# \donttest{
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')

# neurons without a superclass but quite a few synapses
mnm.nc=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.synweight>2000")
mnm.nc |> arrange(desc(synweight))
#>       bodyid post pre downstream upstream synweight           statusLabel
#> 1     955682 2697 851       5469     2697      8166                  <NA>
#> 2     955684  589 809       5744      589      6333                  <NA>
#> 3     954994 1329 698       4297     1329      5626                  <NA>
#> 4     519161 2336 320       2700     2336      5036                Orphan
#> 5     555297  236 604       3884      236      4120                Orphan
#> 6      88667 1887 222       1702     1887      3589                Anchor
#> 7     956907 1938 306       1531     1938      3469                  <NA>
#> 8     806539  202 447       3169      202      3371       Orphan-artifact
#> 9      73540  332 322       3016      332      3348 Prelim Roughly traced
#> 10    926152 1134 321       2087     1134      3221       Orphan-artifact
#> 11     66613 1672 245       1527     1672      3199 Prelim Roughly traced
#> 12    956675   55 421       3139       55      3194                  <NA>
#> 13     92244 1753 222       1426     1753      3179                Orphan
#> 14    942401  351 465       2705      351      3056       Orphan-artifact
#> 15    946767  198 446       2632      198      2830                  <NA>
#> 16    955284  146 350       2656      146      2802                  <NA>
#> 17    104626   53 318       2732       53      2785                Anchor
#> 18    559672 1395 247       1385     1395      2780                Orphan
#> 19    955927  156 391       2573      156      2729                  <NA>
#> 20    110922   62 267       2640       62      2702                Anchor
#> 21    114076  150 300       2529      150      2679                Orphan
#> 22    955135  228 287       2444      228      2672                  <NA>
#> 23    955615  450 382       2111      450      2561                  <NA>
#> 24    955155  597 297       1952      597      2549                  <NA>
#> 25    118423  103 257       2424      103      2527                Anchor
#> 26    555399 1033 424       1491     1033      2524           Unimportant
#> 27 516174168  137 305       2355      137      2492            PRT Orphan
#> 28    954991  387 301       2062      387      2449                  <NA>
#> 29    956789  648 254       1791      648      2439                  <NA>
#> 30    136752  315 245       2093      315      2408                  <NA>
#> 31    147605 1180 183       1208     1180      2388                Orphan
#> 32    954995  399 305       1971      399      2370                  <NA>
#> 33    555267   64 289       2300       64      2364                Anchor
#> 34    210839   70 268       2294       70      2364            PRT Orphan
#> 35    956674  866 278       1489      866      2355                  <NA>
#> 36    908376   26 235       2318       26      2344       Orphan-artifact
#> 37    809279  107 279       2117      107      2224       Orphan-artifact
#> 38    909276   57 275       2131       57      2188       Orphan-artifact
#> 39    210579 1399 178        782     1399      2181                Anchor
#> 40    810259   74 256       2040       74      2114       Orphan-artifact
#> 41    954670  284 251       1812      284      2096                  <NA>
#> 42    953574 1195 227        888     1195      2083                  <NA>
#> 43    544325  238 254       1820      238      2058      RT Hard to trace
#> 44    147742  257 204       1790      257      2047                Orphan
#> 45    903906   52 327       1985       52      2037       Orphan-artifact
#> 46    946550  949 128       1067      949      2016                Orphan
#>    subclass superclass          class entryNerve rootSide           status
#> 1      <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 2      <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 3      <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 4      <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 5      <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 6      <NA>       <NA>           <NA>       <NA>     <NA>           Anchor
#> 7      <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 8      <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 9      <NA>       <NA>           <NA>       <NA>     <NA>           Traced
#> 10     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 11     <NA>       <NA>           <NA>       <NA>     <NA>           Traced
#> 12     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 13     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 14     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 15     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 16     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 17     <NA>       <NA>           <NA>       <NA>     <NA>           Anchor
#> 18     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 19     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 20     <NA>       <NA>           <NA>       <NA>     <NA>           Anchor
#> 21     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 22     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 23     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 24     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 25     <NA>       <NA>           <NA>       <NA>     <NA>           Anchor
#> 26     <NA>       <NA>           <NA>       <NA>     <NA>      Unimportant
#> 27     <NA>       <NA>           <NA>       <NA>     <NA>           Traced
#> 28     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 29     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 30     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 31     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 32     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 33     <NA>       <NA>           <NA>       <NA>     <NA>           Anchor
#> 34     <NA>       <NA>           <NA>       <NA>     <NA>           Traced
#> 35     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 36     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 37     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 38     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 39     <NA>       <NA>           <NA>       <NA>     <NA>           Anchor
#> 40     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 41     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 42     <NA>       <NA>           <NA>       <NA>     <NA>             <NA>
#> 43     <NA>       <NA> mechanosensory       aPhN        R RT Hard to trace
#> 44     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 45     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#> 46     <NA>       <NA>           <NA>       <NA>     <NA>           Orphan
#>    totalNtPredictions predictedNtConfidence   predictedNt
#> 1                 851             0.6905599 acetylcholine
#> 2                 809             0.4990132       unclear
#> 3                 698             0.6480936 acetylcholine
#> 4                 320             0.5911017          gaba
#> 5                 604             0.6630964 acetylcholine
#> 6                 222             0.7629604          gaba
#> 7                 306             0.5808011 acetylcholine
#> 8                 447             0.6745443 acetylcholine
#> 9                 322             0.3064155       unclear
#> 10                321             0.8521824          gaba
#> 11                245             0.8042635          gaba
#> 12                421             0.6915368 acetylcholine
#> 13                222             0.8121917          gaba
#> 14                465             0.3836070       unclear
#> 15                446             0.6567487 acetylcholine
#> 16                350             0.7753956 acetylcholine
#> 17                318             0.6225174          gaba
#> 18                247             0.5886770          gaba
#> 19                391             0.7377123 acetylcholine
#> 20                267             0.7381830 acetylcholine
#> 21                300             0.7077942 acetylcholine
#> 22                287             0.5754398 acetylcholine
#> 23                382             0.3301920       unclear
#> 24                297             0.6054308 acetylcholine
#> 25                257             0.2942655       unclear
#> 26                424             0.3488267       unclear
#> 27                305             0.5985467 acetylcholine
#> 28                301             0.7325168 acetylcholine
#> 29                254             0.7406185 acetylcholine
#> 30                245             0.4928851       unclear
#> 31                183             0.8316643          gaba
#> 32                305             0.7259765 acetylcholine
#> 33                289             0.7676080          gaba
#> 34                268             0.6868132 acetylcholine
#> 35                278             0.7137923 acetylcholine
#> 36                235             0.7470717          gaba
#> 37                279             0.7437456 acetylcholine
#> 38                275             0.3345208       unclear
#> 39                178             0.6651275      dopamine
#> 40                256             0.8303618          gaba
#> 41                251             0.7457834 acetylcholine
#> 42                227             0.8509192          gaba
#> 43                254             0.3842044       unclear
#> 44                204             0.4670316       unclear
#> 45                327             0.6268798 acetylcholine
#> 46                128             0.4909406       unclear
#>    celltypeTotalNtPredictions celltypePredictedNt   consensusNt      voxels
#> 1                         851             unclear acetylcholine   692326498
#> 2                         809             unclear       unclear   476087509
#> 3                         698             unclear acetylcholine   768054781
#> 4                         320             unclear          gaba   275039336
#> 5                         604             unclear acetylcholine   183037288
#> 6                         222             unclear          gaba   197698192
#> 7                         306             unclear acetylcholine  1179077539
#> 8                         447             unclear acetylcholine   204307113
#> 9                         322             unclear       unclear   164205396
#> 10                        321             unclear          gaba   833342245
#> 11                        245             unclear          gaba   184607589
#> 12                        421             unclear acetylcholine   217610342
#> 13                        222             unclear          gaba   172460933
#> 14                        465             unclear       unclear   244236620
#> 15                        446             unclear acetylcholine   417909777
#> 16                        350             unclear acetylcholine   346648352
#> 17                        318             unclear          gaba   118186516
#> 18                        247             unclear          gaba   139035186
#> 19                        391             unclear acetylcholine   286965855
#> 20                        267             unclear acetylcholine   130934995
#> 21                        300             unclear acetylcholine   127874261
#> 22                        287             unclear acetylcholine   106365452
#> 23                        382             unclear       unclear   142895359
#> 24                        297             unclear acetylcholine   389079304
#> 25                        257             unclear       unclear   115227387
#> 26                        424             unclear       unclear 15774680322
#> 27                        305             unclear acetylcholine   147509581
#> 28                        301             unclear acetylcholine   287877411
#> 29                        254             unclear acetylcholine   415371655
#> 30                        245             unclear       unclear   313431542
#> 31                        183             unclear          gaba   118445955
#> 32                        305             unclear acetylcholine   309045905
#> 33                        289             unclear          gaba   135641183
#> 34                        268             unclear acetylcholine    66258699
#> 35                        278             unclear acetylcholine   387945679
#> 36                        235             unclear          gaba   115459022
#> 37                        279             unclear acetylcholine   229596508
#> 38                        275             unclear       unclear   169169631
#> 39                        178             unclear      dopamine   101864445
#> 40                        256             unclear          gaba   230638948
#> 41                        251             unclear acetylcholine   139554521
#> 42                        227             unclear          gaba   564805869
#> 43                        254             unclear       unclear    93097711
#> 44                        204             unclear       unclear    88932916
#> 45                        327             unclear acetylcholine   176044266
#> 46                        128             unclear       unclear   129131037
#>      type flywireType itoleeHl somaSide supertype somaNeuromere trumanHl
#> 1    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 2    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 3    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 4    <NA>        <NA>     <NA>        L      <NA>          <NA>     <NA>
#> 5    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 6    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 7    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 8    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 9    <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 10   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 11   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 12   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 13   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 14   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 15   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 16   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 17   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 18   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 19   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 20   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 21   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 22   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 23   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 24   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 25   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 26   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 27   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 28   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 29   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 30   <NA>        <NA>     <NA>        R      <NA>          <NA>     <NA>
#> 31   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 32   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 33   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 34   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 35   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 36   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 37   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 38   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 39   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 40   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 41   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 42   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 43 aPhM2b        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 44   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 45   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#> 46   <NA>        <NA>     <NA>     <NA>      <NA>          <NA>     <NA>
#>           somaLocation celltypePredictedNtConfidence
#> 1                                                 NA
#> 2                                                 NA
#> 3                                                 NA
#> 4                                                 NA
#> 5                                                 NA
#> 6                                                 NA
#> 7                                                 NA
#> 8                                                 NA
#> 9                                                 NA
#> 10                                                NA
#> 11                                                NA
#> 12                                                NA
#> 13                                                NA
#> 14                                                NA
#> 15                                                NA
#> 16                                                NA
#> 17                                                NA
#> 18                                                NA
#> 19                                                NA
#> 20                                                NA
#> 21                                                NA
#> 22                                                NA
#> 23                                                NA
#> 24                                                NA
#> 25                                                NA
#> 26                                                NA
#> 27                                                NA
#> 28                                                NA
#> 29                                                NA
#> 30 31887, 20287, 16543                            NA
#> 31                                                NA
#> 32                                                NA
#> 33                                                NA
#> 34                                                NA
#> 35                                                NA
#> 36                                                NA
#> 37                                                NA
#> 38                                                NA
#> 39                                                NA
#> 40                                                NA
#> 41                                                NA
#> 42                                                NA
#> 43                                                NA
#> 44                                                NA
#> 45                                                NA
#> 46                                                NA
#>    anteriorCervicalRootPosition synonyms fruDsx matchingNotes receptorType
#> 1                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 2                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 3                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 4                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 5                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 6                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 7                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 8                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 9                          <NA>     <NA>   <NA>          <NA>         <NA>
#> 10                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 11                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 12                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 13                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 14                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 15                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 16                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 17                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 18                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 19                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 20                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 21                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 22                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 23                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 24                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 25                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 26                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 27                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 28                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 29                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 30                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 31                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 32                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 33                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 34                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 35                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 36                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 37                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 38                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 39                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 40                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 41                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 42                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 43                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 44                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 45                         <NA>     <NA>   <NA>          <NA>         <NA>
#> 46                         <NA>     <NA>   <NA>          <NA>         <NA>
#>    mancGroup group mancType birthtime mancBodyid mcnsSerial mancSerial
#> 1         NA    NA     <NA>      <NA>         NA         NA         NA
#> 2         NA    NA     <NA>      <NA>         NA         NA         NA
#> 3         NA    NA     <NA>      <NA>         NA         NA         NA
#> 4         NA    NA     <NA>      <NA>         NA         NA         NA
#> 5         NA    NA     <NA>      <NA>         NA         NA         NA
#> 6         NA    NA     <NA>      <NA>         NA         NA         NA
#> 7         NA    NA     <NA>      <NA>         NA         NA         NA
#> 8         NA    NA     <NA>      <NA>     153319         NA         NA
#> 9         NA    NA     <NA>      <NA>         NA         NA         NA
#> 10        NA    NA     <NA>      <NA>      11325         NA         NA
#> 11        NA    NA     <NA>      <NA>         NA         NA         NA
#> 12        NA    NA     <NA>      <NA>         NA         NA         NA
#> 13        NA    NA     <NA>      <NA>         NA         NA         NA
#> 14        NA    NA     <NA>      <NA>         NA         NA         NA
#> 15        NA    NA     <NA>      <NA>         NA         NA         NA
#> 16        NA    NA     <NA>      <NA>         NA         NA         NA
#> 17        NA    NA     <NA>      <NA>         NA         NA         NA
#> 18        NA    NA     <NA>      <NA>         NA         NA         NA
#> 19        NA    NA     <NA>      <NA>         NA         NA         NA
#> 20        NA    NA     <NA>      <NA>         NA         NA         NA
#> 21        NA    NA     <NA>      <NA>         NA         NA         NA
#> 22        NA    NA     <NA>      <NA>         NA         NA         NA
#> 23        NA    NA     <NA>      <NA>         NA         NA         NA
#> 24        NA    NA     <NA>      <NA>         NA         NA         NA
#> 25        NA    NA     <NA>      <NA>         NA         NA         NA
#> 26        NA    NA     <NA>      <NA>         NA         NA         NA
#> 27        NA    NA     <NA>      <NA>         NA         NA         NA
#> 28        NA    NA     <NA>      <NA>         NA         NA         NA
#> 29        NA    NA     <NA>      <NA>         NA         NA         NA
#> 30        NA    NA     <NA>      <NA>         NA         NA         NA
#> 31        NA    NA     <NA>      <NA>         NA         NA         NA
#> 32        NA    NA     <NA>      <NA>         NA         NA         NA
#> 33        NA    NA     <NA>      <NA>         NA         NA         NA
#> 34        NA    NA     <NA>      <NA>         NA         NA         NA
#> 35        NA    NA     <NA>      <NA>         NA         NA         NA
#> 36        NA    NA     <NA>      <NA>         NA         NA         NA
#> 37        NA    NA     <NA>      <NA>      16226         NA         NA
#> 38        NA    NA     <NA>      <NA>      28494         NA         NA
#> 39        NA    NA     <NA>      <NA>         NA         NA         NA
#> 40        NA    NA     <NA>      <NA>      29891         NA         NA
#> 41        NA    NA     <NA>      <NA>         NA         NA         NA
#> 42        NA    NA     <NA>      <NA>         NA         NA         NA
#> 43        NA    NA     <NA>      <NA>         NA         NA         NA
#> 44        NA    NA     <NA>      <NA>         NA         NA         NA
#> 45        NA    NA     <NA>      <NA>         NA         NA         NA
#> 46        NA    NA     <NA>      <NA>         NA         NA         NA
#>    exitNerve tosomaLocation          name hemibrainType flywireId serialMotif
#> 1       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 2       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 3       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 4       <NA>           <NA> KC_L_fragment          <NA>      <NA>        <NA>
#> 5       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 6       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 7       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 8       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 9       <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 10      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 11      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 12      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 13      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 14      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 15      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 16      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 17      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 18      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 19      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 20      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 21      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 22      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 23      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 24      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 25      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 26      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 27      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 28      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 29      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 30      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 31      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 32      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 33      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 34      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 35      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 36      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 37      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 38      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 39      <NA>           <NA> MBON_fragment          <NA>      <NA>        <NA>
#> 40      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 41      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 42      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 43      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 44      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 45      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#> 46      <NA>           <NA>          <NA>          <NA>      <NA>        <NA>
#>    posteriorCervicalRootPosition dimorphism assignedOlHex1 assignedOlHex2
#> 1                           <NA>       <NA>             NA             NA
#> 2                           <NA>       <NA>             NA             NA
#> 3                           <NA>       <NA>             NA             NA
#> 4                           <NA>       <NA>             NA             NA
#> 5                           <NA>       <NA>             NA             NA
#> 6                           <NA>       <NA>             NA             NA
#> 7                           <NA>       <NA>             NA             NA
#> 8                           <NA>       <NA>             NA             NA
#> 9                           <NA>       <NA>             NA             NA
#> 10                          <NA>       <NA>             NA             NA
#> 11                          <NA>       <NA>             NA             NA
#> 12                          <NA>       <NA>             NA             NA
#> 13                          <NA>       <NA>             NA             NA
#> 14                          <NA>       <NA>             NA             NA
#> 15                          <NA>       <NA>             NA             NA
#> 16                          <NA>       <NA>             NA             NA
#> 17                          <NA>       <NA>             NA             NA
#> 18                          <NA>       <NA>             NA             NA
#> 19                          <NA>       <NA>             NA             NA
#> 20                          <NA>       <NA>             NA             NA
#> 21                          <NA>       <NA>             NA             NA
#> 22                          <NA>       <NA>             NA             NA
#> 23                          <NA>       <NA>             NA             NA
#> 24                          <NA>       <NA>             NA             NA
#> 25                          <NA>       <NA>             NA             NA
#> 26                          <NA>       <NA>             NA             NA
#> 27                          <NA>       <NA>             NA             NA
#> 28                          <NA>       <NA>             NA             NA
#> 29                          <NA>       <NA>             NA             NA
#> 30                          <NA>       <NA>             NA             NA
#> 31                          <NA>       <NA>             NA             NA
#> 32                          <NA>       <NA>             NA             NA
#> 33                          <NA>       <NA>             NA             NA
#> 34                          <NA>       <NA>             NA             NA
#> 35                          <NA>       <NA>             NA             NA
#> 36                          <NA>       <NA>             NA             NA
#> 37                          <NA>       <NA>             NA             NA
#> 38                          <NA>       <NA>             NA             NA
#> 39                          <NA>       <NA>             NA             NA
#> 40                          <NA>       <NA>             NA             NA
#> 41                          <NA>       <NA>             NA             NA
#> 42                          <NA>       <NA>             NA             NA
#> 43                          <NA>       <NA>             NA             NA
#> 44                          <NA>       <NA>             NA             NA
#> 45                          <NA>       <NA>             NA             NA
#> 46                          <NA>       <NA>             NA             NA
#>    hemibrainBodyId location locationType cellBodyFiber  soma
#> 1             <NA>     <NA>         <NA>          <NA> FALSE
#> 2             <NA>     <NA>         <NA>          <NA> FALSE
#> 3             <NA>     <NA>         <NA>          <NA> FALSE
#> 4             <NA>     <NA>         <NA>          <NA> FALSE
#> 5             <NA>     <NA>         <NA>          <NA> FALSE
#> 6             <NA>     <NA>         <NA>          <NA> FALSE
#> 7             <NA>     <NA>         <NA>          <NA> FALSE
#> 8             <NA>     <NA>         <NA>          <NA> FALSE
#> 9             <NA>     <NA>         <NA>          <NA> FALSE
#> 10            <NA>     <NA>         <NA>          <NA> FALSE
#> 11            <NA>     <NA>         <NA>          <NA> FALSE
#> 12            <NA>     <NA>         <NA>          <NA> FALSE
#> 13            <NA>     <NA>         <NA>          <NA> FALSE
#> 14            <NA>     <NA>         <NA>          <NA> FALSE
#> 15            <NA>     <NA>         <NA>          <NA> FALSE
#> 16            <NA>     <NA>         <NA>          <NA> FALSE
#> 17            <NA>     <NA>         <NA>          <NA> FALSE
#> 18            <NA>     <NA>         <NA>          <NA> FALSE
#> 19            <NA>     <NA>         <NA>          <NA> FALSE
#> 20            <NA>     <NA>         <NA>          <NA> FALSE
#> 21            <NA>     <NA>         <NA>          <NA> FALSE
#> 22            <NA>     <NA>         <NA>          <NA> FALSE
#> 23            <NA>     <NA>         <NA>          <NA> FALSE
#> 24            <NA>     <NA>         <NA>          <NA> FALSE
#> 25            <NA>     <NA>         <NA>          <NA> FALSE
#> 26            <NA>     <NA>         <NA>          <NA> FALSE
#> 27            <NA>     <NA>         <NA>          <NA> FALSE
#> 28            <NA>     <NA>         <NA>          <NA> FALSE
#> 29            <NA>     <NA>         <NA>          <NA> FALSE
#> 30            <NA>     <NA>         <NA>          <NA>  TRUE
#> 31            <NA>     <NA>         <NA>          <NA> FALSE
#> 32            <NA>     <NA>         <NA>          <NA> FALSE
#> 33            <NA>     <NA>         <NA>          <NA> FALSE
#> 34            <NA>     <NA>         <NA>          <NA> FALSE
#> 35            <NA>     <NA>         <NA>          <NA> FALSE
#> 36            <NA>     <NA>         <NA>          <NA> FALSE
#> 37            <NA>     <NA>         <NA>          <NA> FALSE
#> 38            <NA>     <NA>         <NA>          <NA> FALSE
#> 39            <NA>     <NA>         <NA>          <NA> FALSE
#> 40            <NA>     <NA>         <NA>          <NA> FALSE
#> 41            <NA>     <NA>         <NA>          <NA> FALSE
#> 42            <NA>     <NA>         <NA>          <NA> FALSE
#> 43            <NA>     <NA>         <NA>          <NA> FALSE
#> 44            <NA>     <NA>         <NA>          <NA> FALSE
#> 45            <NA>     <NA>         <NA>          <NA> FALSE
#> 46            <NA>     <NA>         <NA>          <NA> FALSE
# }

# Which neurons don't have a superclass, but possibly should
mnm.nsc=mcns_neuprint_meta("where:NOT exists(n.superclass)")
mnm.nsc |> count(statusLabel)
#>              statusLabel    n
#> 1              0.5assign   45
#> 2                 Anchor  281
#> 3                   Glia  331
#> 4                 Leaves  150
#> 5                 Orphan 3397
#> 6        Orphan hotknife  481
#> 7        Orphan-artifact 1452
#> 8           Out of scope  981
#> 9             PRT Orphan   22
#> 10 Prelim Roughly traced   19
#> 11        Primary Anchor  473
#> 12      RT Hard to trace    2
#> 13        Roughly traced    1
#> 14        Sensory Anchor  139
#> 15           Soma Anchor   25
#> 16           Unimportant  123
#> 17                  <NA>  160

# neurons that are RT or PRT should probably have a superclass
mnm.nscprt=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.statusLabel CONTAINS 'Roughly'")
mnm.nscprt |> count()
#>    n
#> 1 20