Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

# \donttest{
library(dplyr)
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')

# neurons without a superclass but quite a few synapses
mnm.nc=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.synweight>2000")
mnm.nc |> arrange(desc(synweight))
#>       bodyid post pre downstream upstream synweight           statusLabel
#> 1     955682 2697 851       5469     2697      8166                  <NA>
#> 2     955684  589 809       5744      589      6333                  <NA>
#> 3     954994 1344 700       4307     1344      5651                  <NA>
#> 4     519161 2403 322       2718     2403      5121                Orphan
#> 5     555297  236 604       3884      236      4120                Orphan
#> 6      88667 1908 225       1724     1908      3632                Anchor
#> 7     956907 1965 306       1531     1965      3496                  <NA>
#> 8     806539  202 447       3169      202      3371       Orphan-artifact
#> 9      73540  332 322       3016      332      3348 Prelim Roughly traced
#> 10    926152 1134 321       2087     1134      3221       Orphan-artifact
#> 11     66613 1686 245       1527     1686      3213 Prelim Roughly traced
#> 12    956675   55 421       3139       55      3194                  <NA>
#> 13     92244 1753 222       1426     1753      3179                Orphan
#> 14    942401  351 465       2705      351      3056       Orphan-artifact
#> 15    946767  198 446       2632      198      2830                  <NA>
#> 16    955927  163 401       2644      163      2807                  <NA>
#> 17    955284  146 350       2656      146      2802                  <NA>
#> 18    104626   53 318       2732       53      2785                Anchor
#> 19    559672 1395 247       1385     1395      2780                Orphan
#> 20    955135  237 291       2471      237      2708                  <NA>
#> 21    110922   62 267       2640       62      2702                Anchor
#> 22    114076  150 300       2529      150      2679                Orphan
#> 23 516174168  150 319       2494      150      2644            PRT Orphan
#> 24    955155  609 302       1982      609      2591                  <NA>
#> 25    955615  450 382       2111      450      2561                  <NA>
#> 26    118423  104 259       2439      104      2543                Anchor
#> 27    555399 1033 424       1491     1033      2524           Unimportant
#> 28    954991  392 308       2124      392      2516                  <NA>
#> 29    136752  376 245       2093      376      2469                  <NA>
#> 30    956789  648 254       1791      648      2439                  <NA>
#> 31    147605 1180 183       1208     1180      2388                Orphan
#> 32    956674  868 281       1506      868      2374                  <NA>
#> 33    954995  399 305       1971      399      2370                  <NA>
#> 34    555267   64 289       2300       64      2364                Anchor
#> 35    210839   70 268       2294       70      2364            PRT Orphan
#> 36    908376   26 235       2318       26      2344       Orphan-artifact
#> 37    954670  309 268       1947      309      2256                  <NA>
#> 38    809279  107 279       2117      107      2224       Orphan-artifact
#> 39    909276   57 275       2131       57      2188       Orphan-artifact
#> 40    810259   74 256       2040       74      2114       Orphan-artifact
#> 41    953574 1202 231        906     1202      2108                  <NA>
#> 42 290057755   71 224       2020       71      2091             0.5assign
#> 43    544325  238 254       1820      238      2058      RT Hard to trace
#> 44    147742  257 204       1790      257      2047                Orphan
#> 45    903906   52 327       1985       52      2037       Orphan-artifact
#> 46    946550  949 128       1067      949      2016                Orphan
#>    superclass type subclass class rootSide entryNerve           status
#> 1        <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 2        <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 3        <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 4        <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 5        <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 6        <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Anchor
#> 7        <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 8        <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 9        <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Traced
#> 10       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 11       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Traced
#> 12       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 13       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 14       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 15       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 16       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 17       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 18       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Anchor
#> 19       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 20       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 21       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Anchor
#> 22       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 23       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Traced
#> 24       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 25       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 26       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Anchor
#> 27       <NA> <NA>     <NA>  <NA>     <NA>       <NA>      Unimportant
#> 28       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 29       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 30       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 31       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 32       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 33       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 34       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Anchor
#> 35       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Traced
#> 36       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 37       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 38       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 39       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 40       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 41       <NA> <NA>     <NA>  <NA>     <NA>       <NA>             <NA>
#> 42       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Assign
#> 43       <NA> <NA>     <NA>  <NA>        R       aPhN RT Hard to trace
#> 44       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 45       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#> 46       <NA> <NA>     <NA>  <NA>     <NA>       <NA>           Orphan
#>    totalNtPredictions predictedNtConfidence   predictedNt
#> 1                 851             0.6882697 acetylcholine
#> 2                 809             0.4961571       unclear
#> 3                 700             0.6453297 acetylcholine
#> 4                 322             0.5897908          gaba
#> 5                 604             0.6609749 acetylcholine
#> 6                 225             0.7598869          gaba
#> 7                 306             0.5789229 acetylcholine
#> 8                 447             0.6723387 acetylcholine
#> 9                 322             0.3080053       unclear
#> 10                321             0.8467842          gaba
#> 11                245             0.7993161          gaba
#> 12                421             0.6892003 acetylcholine
#> 13                222             0.8071632          gaba
#> 14                465             0.3824963       unclear
#> 15                446             0.6546214 acetylcholine
#> 16                401             0.7368743 acetylcholine
#> 17                350             0.7727528 acetylcholine
#> 18                318             0.6187874          gaba
#> 19                247             0.5856910          gaba
#> 20                291             0.5686773 acetylcholine
#> 21                267             0.7357013 acetylcholine
#> 22                300             0.7054712 acetylcholine
#> 23                319             0.5936913 acetylcholine
#> 24                302             0.6042004 acetylcholine
#> 25                382             0.3295135       unclear
#> 26                259             0.2966501       unclear
#> 27                424             0.3480597       unclear
#> 28                308             0.7316513 acetylcholine
#> 29                245             0.4916106       unclear
#> 30                254             0.7381036 acetylcholine
#> 31                183             0.8264490          gaba
#> 32                281             0.7123947 acetylcholine
#> 33                305             0.7235335 acetylcholine
#> 34                289             0.7628830          gaba
#> 35                268             0.6844937 acetylcholine
#> 36                235             0.7424799          gaba
#> 37                268             0.7327254 acetylcholine
#> 38                279             0.7412243 acetylcholine
#> 39                275             0.3365083       unclear
#> 40                256             0.8250818          gaba
#> 41                231             0.8422147          gaba
#> 42                224             0.7097775 acetylcholine
#> 43                254             0.3866957       unclear
#> 44                204             0.4705864       unclear
#> 45                327             0.6248420 acetylcholine
#> 46                128             0.4883891       unclear
#>    celltypeTotalNtPredictions celltypePredictedNt celltypePredictedNtConfidence
#> 1                         851             unclear                            NA
#> 2                         809             unclear                            NA
#> 3                         700             unclear                            NA
#> 4                         322             unclear                            NA
#> 5                         604             unclear                            NA
#> 6                         225             unclear                            NA
#> 7                         306             unclear                            NA
#> 8                         447             unclear                            NA
#> 9                         322             unclear                            NA
#> 10                        321             unclear                            NA
#> 11                        245             unclear                            NA
#> 12                        421             unclear                            NA
#> 13                        222             unclear                            NA
#> 14                        465             unclear                            NA
#> 15                        446             unclear                            NA
#> 16                        401             unclear                            NA
#> 17                        350             unclear                            NA
#> 18                        318             unclear                            NA
#> 19                        247             unclear                            NA
#> 20                        291             unclear                            NA
#> 21                        267             unclear                            NA
#> 22                        300             unclear                            NA
#> 23                        319             unclear                            NA
#> 24                        302             unclear                            NA
#> 25                        382             unclear                            NA
#> 26                        259             unclear                            NA
#> 27                        424             unclear                            NA
#> 28                        308             unclear                            NA
#> 29                        245             unclear                            NA
#> 30                        254             unclear                            NA
#> 31                        183             unclear                            NA
#> 32                        281             unclear                            NA
#> 33                        305             unclear                            NA
#> 34                        289             unclear                            NA
#> 35                        268             unclear                            NA
#> 36                        235             unclear                            NA
#> 37                        268             unclear                            NA
#> 38                        279             unclear                            NA
#> 39                        275             unclear                            NA
#> 40                        256             unclear                            NA
#> 41                        231             unclear                            NA
#> 42                        224             unclear                            NA
#> 43                       1472       acetylcholine                     0.6163095
#> 44                        204             unclear                            NA
#> 45                        327             unclear                            NA
#> 46                        128             unclear                            NA
#>      consensusNt      voxels flywireType somaSide itoleeHl supertype
#> 1  acetylcholine   692326498        <NA>     <NA>     <NA>      <NA>
#> 2        unclear   476087509        <NA>     <NA>     <NA>      <NA>
#> 3  acetylcholine   769950427        <NA>     <NA>     <NA>      <NA>
#> 4           gaba   278587841        <NA>        L     <NA>      <NA>
#> 5  acetylcholine   183037288        <NA>     <NA>     <NA>      <NA>
#> 6           gaba   199021086        <NA>     <NA>     <NA>      <NA>
#> 7  acetylcholine  1181128023        <NA>     <NA>     <NA>      <NA>
#> 8  acetylcholine   204307113        <NA>     <NA>     <NA>      <NA>
#> 9        unclear   164205396        <NA>     <NA>     <NA>      <NA>
#> 10          gaba   833342245        <NA>     <NA>     <NA>      <NA>
#> 11          gaba   185592243        <NA>     <NA>     <NA>      <NA>
#> 12 acetylcholine   217610342        <NA>     <NA>     <NA>      <NA>
#> 13          gaba   172460933        <NA>     <NA>     <NA>      <NA>
#> 14       unclear   244236620        <NA>     <NA>     <NA>      <NA>
#> 15 acetylcholine   417909777        <NA>     <NA>     <NA>      <NA>
#> 16 acetylcholine   291166682        <NA>     <NA>     <NA>      <NA>
#> 17 acetylcholine   346648352        <NA>     <NA>     <NA>      <NA>
#> 18          gaba   118186516        <NA>     <NA>     <NA>      <NA>
#> 19          gaba   139035186        <NA>     <NA>     <NA>      <NA>
#> 20 acetylcholine   107081588        <NA>     <NA>     <NA>      <NA>
#> 21 acetylcholine   130934995        <NA>     <NA>     <NA>      <NA>
#> 22 acetylcholine   127874261        <NA>     <NA>     <NA>      <NA>
#> 23 acetylcholine   148541348        <NA>     <NA>     <NA>      <NA>
#> 24 acetylcholine   393057814        <NA>     <NA>     <NA>      <NA>
#> 25       unclear   142895359        <NA>     <NA>     <NA>      <NA>
#> 26       unclear   115258843        <NA>     <NA>     <NA>      <NA>
#> 27       unclear 15774680322        <NA>     <NA>     <NA>      <NA>
#> 28 acetylcholine   291580020        <NA>     <NA>     <NA>      <NA>
#> 29       unclear   323241784        <NA>        R     <NA>      <NA>
#> 30 acetylcholine   415371655        <NA>     <NA>     <NA>      <NA>
#> 31          gaba   118445955        <NA>     <NA>     <NA>      <NA>
#> 32 acetylcholine   389656375        <NA>     <NA>     <NA>      <NA>
#> 33 acetylcholine   309045905        <NA>     <NA>     <NA>      <NA>
#> 34          gaba   135641183        <NA>     <NA>     <NA>      <NA>
#> 35 acetylcholine    66258699        <NA>     <NA>     <NA>      <NA>
#> 36          gaba   115459022        <NA>     <NA>     <NA>      <NA>
#> 37 acetylcholine   141835905        <NA>     <NA>     <NA>      <NA>
#> 38 acetylcholine   229596508        <NA>     <NA>     <NA>      <NA>
#> 39       unclear   169169631        <NA>     <NA>     <NA>      <NA>
#> 40          gaba   230638948        <NA>     <NA>     <NA>      <NA>
#> 41          gaba   567492443        <NA>     <NA>     <NA>      <NA>
#> 42 acetylcholine    29944601        <NA>     <NA>     <NA>      <NA>
#> 43 acetylcholine    93097711        <NA>     <NA>     <NA>      <NA>
#> 44       unclear    88932916        <NA>     <NA>     <NA>      <NA>
#> 45 acetylcholine   176044266        <NA>     <NA>     <NA>      <NA>
#> 46       unclear   129131037        <NA>     <NA>     <NA>      <NA>
#>    somaNeuromere trumanHl        somaLocation group mancGroup mancType
#> 1           <NA>     <NA>                        NA        NA     <NA>
#> 2           <NA>     <NA>                        NA        NA     <NA>
#> 3           <NA>     <NA>                        NA        NA     <NA>
#> 4           <NA>     <NA>                        NA        NA     <NA>
#> 5           <NA>     <NA>                        NA        NA     <NA>
#> 6           <NA>     <NA>                        NA        NA     <NA>
#> 7           <NA>     <NA>                        NA        NA     <NA>
#> 8           <NA>     <NA>                        NA        NA     <NA>
#> 9           <NA>     <NA>                        NA        NA     <NA>
#> 10          <NA>     <NA>                        NA        NA     <NA>
#> 11          <NA>     <NA>                        NA        NA     <NA>
#> 12          <NA>     <NA>                        NA        NA     <NA>
#> 13          <NA>     <NA>                        NA        NA     <NA>
#> 14          <NA>     <NA>                        NA        NA     <NA>
#> 15          <NA>     <NA>                        NA        NA     <NA>
#> 16          <NA>     <NA>                        NA        NA     <NA>
#> 17          <NA>     <NA>                        NA        NA     <NA>
#> 18          <NA>     <NA>                        NA        NA     <NA>
#> 19          <NA>     <NA>                        NA        NA     <NA>
#> 20          <NA>     <NA>                        NA        NA     <NA>
#> 21          <NA>     <NA>                        NA        NA     <NA>
#> 22          <NA>     <NA>                        NA        NA     <NA>
#> 23          <NA>     <NA>                        NA        NA     <NA>
#> 24          <NA>     <NA>                        NA        NA     <NA>
#> 25          <NA>     <NA>                        NA        NA     <NA>
#> 26          <NA>     <NA>                        NA        NA     <NA>
#> 27          <NA>     <NA>                        NA        NA     <NA>
#> 28          <NA>     <NA>                        NA        NA     <NA>
#> 29          <NA>     <NA> 31887, 20287, 16543    NA        NA     <NA>
#> 30          <NA>     <NA>                        NA        NA     <NA>
#> 31          <NA>     <NA>                        NA        NA     <NA>
#> 32          <NA>     <NA>                        NA        NA     <NA>
#> 33          <NA>     <NA>                        NA        NA     <NA>
#> 34          <NA>     <NA>                        NA        NA     <NA>
#> 35          <NA>     <NA>                        NA        NA     <NA>
#> 36          <NA>     <NA>                        NA        NA     <NA>
#> 37          <NA>     <NA>                        NA        NA     <NA>
#> 38          <NA>     <NA>                        NA        NA     <NA>
#> 39          <NA>     <NA>                        NA        NA     <NA>
#> 40          <NA>     <NA>                        NA        NA     <NA>
#> 41          <NA>     <NA>                        NA        NA     <NA>
#> 42          <NA>     <NA>                        NA        NA     <NA>
#> 43          <NA>     <NA>                        NA        NA     <NA>
#> 44          <NA>     <NA>                        NA        NA     <NA>
#> 45          <NA>     <NA>                        NA        NA     <NA>
#> 46          <NA>     <NA>                        NA        NA     <NA>
#>    birthtime mancBodyid mancSerial mcnsSerial synonyms fruDsx exitNerve
#> 1       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 2       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 3       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 4       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 5       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 6       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 7       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 8       <NA>     153319         NA         NA     <NA>   <NA>      <NA>
#> 9       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 10      <NA>      11325         NA         NA     <NA>   <NA>      <NA>
#> 11      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 12      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 13      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 14      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 15      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 16      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 17      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 18      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 19      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 20      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 21      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 22      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 23      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 24      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 25      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 26      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 27      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 28      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 29      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 30      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 31      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 32      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 33      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 34      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 35      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 36      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 37      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 38      <NA>      16226         NA         NA     <NA>   <NA>      <NA>
#> 39      <NA>      28494         NA         NA     <NA>   <NA>      <NA>
#> 40      <NA>      29891         NA         NA     <NA>   <NA>      <NA>
#> 41      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 42      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 43      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 44      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 45      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 46      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#>    tosomaLocation receptorType matchingNotes hemibrainType          name
#> 1            <NA>         <NA>          <NA>          <NA>          <NA>
#> 2            <NA>         <NA>          <NA>          <NA>          <NA>
#> 3            <NA>         <NA>          <NA>          <NA>          <NA>
#> 4            <NA>         <NA>          <NA>          <NA> KC_L_fragment
#> 5            <NA>         <NA>          <NA>          <NA>          <NA>
#> 6            <NA>         <NA>          <NA>          <NA>          <NA>
#> 7            <NA>         <NA>          <NA>          <NA>          <NA>
#> 8            <NA>         <NA>          <NA>          <NA>          <NA>
#> 9            <NA>         <NA>          <NA>          <NA>          <NA>
#> 10           <NA>         <NA>          <NA>          <NA>          <NA>
#> 11           <NA>         <NA>          <NA>          <NA>          <NA>
#> 12           <NA>         <NA>          <NA>          <NA>          <NA>
#> 13           <NA>         <NA>          <NA>          <NA>          <NA>
#> 14           <NA>         <NA>          <NA>          <NA>          <NA>
#> 15           <NA>         <NA>          <NA>          <NA>          <NA>
#> 16           <NA>         <NA>          <NA>          <NA>          <NA>
#> 17           <NA>         <NA>          <NA>          <NA>          <NA>
#> 18           <NA>         <NA>          <NA>          <NA>          <NA>
#> 19           <NA>         <NA>          <NA>          <NA>          <NA>
#> 20           <NA>         <NA>          <NA>          <NA>          <NA>
#> 21           <NA>         <NA>          <NA>          <NA>          <NA>
#> 22           <NA>         <NA>          <NA>          <NA>          <NA>
#> 23           <NA>         <NA>          <NA>          <NA>          <NA>
#> 24           <NA>         <NA>          <NA>          <NA>          <NA>
#> 25           <NA>         <NA>          <NA>          <NA>          <NA>
#> 26           <NA>         <NA>          <NA>          <NA>          <NA>
#> 27           <NA>         <NA>          <NA>          <NA>          <NA>
#> 28           <NA>         <NA>          <NA>          <NA>          <NA>
#> 29           <NA>         <NA>          <NA>          <NA>          <NA>
#> 30           <NA>         <NA>          <NA>          <NA>          <NA>
#> 31           <NA>         <NA>          <NA>          <NA>          <NA>
#> 32           <NA>         <NA>          <NA>          <NA>          <NA>
#> 33           <NA>         <NA>          <NA>          <NA>          <NA>
#> 34           <NA>         <NA>          <NA>          <NA>          <NA>
#> 35           <NA>         <NA>          <NA>          <NA>          <NA>
#> 36           <NA>         <NA>          <NA>          <NA>          <NA>
#> 37           <NA>         <NA>          <NA>          <NA>          <NA>
#> 38           <NA>         <NA>          <NA>          <NA>          <NA>
#> 39           <NA>         <NA>          <NA>          <NA>          <NA>
#> 40           <NA>         <NA>          <NA>          <NA>          <NA>
#> 41           <NA>         <NA>          <NA>          <NA>          <NA>
#> 42           <NA>         <NA>          <NA>          <NA>          <NA>
#> 43           <NA>         <NA>          <NA>          <NA>          <NA>
#> 44           <NA>         <NA>          <NA>          <NA>          <NA>
#> 45           <NA>         <NA>          <NA>          <NA>          <NA>
#> 46           <NA>         <NA>          <NA>          <NA>          <NA>
#>    anteriorCervicalRootPosition flywireId posteriorCervicalRootPosition
#> 1                          <NA>      <NA>                          <NA>
#> 2                          <NA>      <NA>                          <NA>
#> 3                          <NA>      <NA>                          <NA>
#> 4                          <NA>      <NA>                          <NA>
#> 5                          <NA>      <NA>                          <NA>
#> 6                          <NA>      <NA>                          <NA>
#> 7                          <NA>      <NA>                          <NA>
#> 8                          <NA>      <NA>                          <NA>
#> 9                          <NA>      <NA>                          <NA>
#> 10                         <NA>      <NA>                          <NA>
#> 11                         <NA>      <NA>                          <NA>
#> 12                         <NA>      <NA>                          <NA>
#> 13                         <NA>      <NA>                          <NA>
#> 14                         <NA>      <NA>                          <NA>
#> 15                         <NA>      <NA>                          <NA>
#> 16                         <NA>      <NA>                          <NA>
#> 17                         <NA>      <NA>                          <NA>
#> 18                         <NA>      <NA>                          <NA>
#> 19                         <NA>      <NA>                          <NA>
#> 20                         <NA>      <NA>                          <NA>
#> 21                         <NA>      <NA>                          <NA>
#> 22                         <NA>      <NA>                          <NA>
#> 23                         <NA>      <NA>                          <NA>
#> 24                         <NA>      <NA>                          <NA>
#> 25                         <NA>      <NA>                          <NA>
#> 26                         <NA>      <NA>                          <NA>
#> 27                         <NA>      <NA>                          <NA>
#> 28                         <NA>      <NA>                          <NA>
#> 29                         <NA>      <NA>                          <NA>
#> 30                         <NA>      <NA>                          <NA>
#> 31                         <NA>      <NA>                          <NA>
#> 32                         <NA>      <NA>                          <NA>
#> 33                         <NA>      <NA>                          <NA>
#> 34                         <NA>      <NA>                          <NA>
#> 35                         <NA>      <NA>                          <NA>
#> 36                         <NA>      <NA>                          <NA>
#> 37                         <NA>      <NA>                          <NA>
#> 38                         <NA>      <NA>                          <NA>
#> 39                         <NA>      <NA>                          <NA>
#> 40                         <NA>      <NA>                          <NA>
#> 41                         <NA>      <NA>                          <NA>
#> 42                         <NA>      <NA>                          <NA>
#> 43                         <NA>      <NA>                          <NA>
#> 44                         <NA>      <NA>                          <NA>
#> 45                         <NA>      <NA>                          <NA>
#> 46                         <NA>      <NA>                          <NA>
#>    serialMotif dimorphism assignedOlHex1 assignedOlHex2 hemibrainBodyId
#> 1         <NA>       <NA>             NA             NA            <NA>
#> 2         <NA>       <NA>             NA             NA            <NA>
#> 3         <NA>       <NA>             NA             NA            <NA>
#> 4         <NA>       <NA>             NA             NA            <NA>
#> 5         <NA>       <NA>             NA             NA            <NA>
#> 6         <NA>       <NA>             NA             NA            <NA>
#> 7         <NA>       <NA>             NA             NA            <NA>
#> 8         <NA>       <NA>             NA             NA            <NA>
#> 9         <NA>       <NA>             NA             NA            <NA>
#> 10        <NA>       <NA>             NA             NA            <NA>
#> 11        <NA>       <NA>             NA             NA            <NA>
#> 12        <NA>       <NA>             NA             NA            <NA>
#> 13        <NA>       <NA>             NA             NA            <NA>
#> 14        <NA>       <NA>             NA             NA            <NA>
#> 15        <NA>       <NA>             NA             NA            <NA>
#> 16        <NA>       <NA>             NA             NA            <NA>
#> 17        <NA>       <NA>             NA             NA            <NA>
#> 18        <NA>       <NA>             NA             NA            <NA>
#> 19        <NA>       <NA>             NA             NA            <NA>
#> 20        <NA>       <NA>             NA             NA            <NA>
#> 21        <NA>       <NA>             NA             NA            <NA>
#> 22        <NA>       <NA>             NA             NA            <NA>
#> 23        <NA>       <NA>             NA             NA            <NA>
#> 24        <NA>       <NA>             NA             NA            <NA>
#> 25        <NA>       <NA>             NA             NA            <NA>
#> 26        <NA>       <NA>             NA             NA            <NA>
#> 27        <NA>       <NA>             NA             NA            <NA>
#> 28        <NA>       <NA>             NA             NA            <NA>
#> 29        <NA>       <NA>             NA             NA            <NA>
#> 30        <NA>       <NA>             NA             NA            <NA>
#> 31        <NA>       <NA>             NA             NA            <NA>
#> 32        <NA>       <NA>             NA             NA            <NA>
#> 33        <NA>       <NA>             NA             NA            <NA>
#> 34        <NA>       <NA>             NA             NA            <NA>
#> 35        <NA>       <NA>             NA             NA            <NA>
#> 36        <NA>       <NA>             NA             NA            <NA>
#> 37        <NA>       <NA>             NA             NA            <NA>
#> 38        <NA>       <NA>             NA             NA            <NA>
#> 39        <NA>       <NA>             NA             NA            <NA>
#> 40        <NA>       <NA>             NA             NA            <NA>
#> 41        <NA>       <NA>             NA             NA            <NA>
#> 42        <NA>       <NA>             NA             NA            <NA>
#> 43        <NA>       <NA>             NA             NA            <NA>
#> 44        <NA>       <NA>             NA             NA            <NA>
#> 45        <NA>       <NA>             NA             NA            <NA>
#> 46        <NA>       <NA>             NA             NA            <NA>
#>    location locationType cellBodyFiber  soma
#> 1      <NA>         <NA>          <NA> FALSE
#> 2      <NA>         <NA>          <NA> FALSE
#> 3      <NA>         <NA>          <NA> FALSE
#> 4      <NA>         <NA>          <NA> FALSE
#> 5      <NA>         <NA>          <NA> FALSE
#> 6      <NA>         <NA>          <NA> FALSE
#> 7      <NA>         <NA>          <NA> FALSE
#> 8      <NA>         <NA>          <NA> FALSE
#> 9      <NA>         <NA>          <NA> FALSE
#> 10     <NA>         <NA>          <NA> FALSE
#> 11     <NA>         <NA>          <NA> FALSE
#> 12     <NA>         <NA>          <NA> FALSE
#> 13     <NA>         <NA>          <NA> FALSE
#> 14     <NA>         <NA>          <NA> FALSE
#> 15     <NA>         <NA>          <NA> FALSE
#> 16     <NA>         <NA>          <NA> FALSE
#> 17     <NA>         <NA>          <NA> FALSE
#> 18     <NA>         <NA>          <NA> FALSE
#> 19     <NA>         <NA>          <NA> FALSE
#> 20     <NA>         <NA>          <NA> FALSE
#> 21     <NA>         <NA>          <NA> FALSE
#> 22     <NA>         <NA>          <NA> FALSE
#> 23     <NA>         <NA>          <NA> FALSE
#> 24     <NA>         <NA>          <NA> FALSE
#> 25     <NA>         <NA>          <NA> FALSE
#> 26     <NA>         <NA>          <NA> FALSE
#> 27     <NA>         <NA>          <NA> FALSE
#> 28     <NA>         <NA>          <NA> FALSE
#> 29     <NA>         <NA>          <NA>  TRUE
#> 30     <NA>         <NA>          <NA> FALSE
#> 31     <NA>         <NA>          <NA> FALSE
#> 32     <NA>         <NA>          <NA> FALSE
#> 33     <NA>         <NA>          <NA> FALSE
#> 34     <NA>         <NA>          <NA> FALSE
#> 35     <NA>         <NA>          <NA> FALSE
#> 36     <NA>         <NA>          <NA> FALSE
#> 37     <NA>         <NA>          <NA> FALSE
#> 38     <NA>         <NA>          <NA> FALSE
#> 39     <NA>         <NA>          <NA> FALSE
#> 40     <NA>         <NA>          <NA> FALSE
#> 41     <NA>         <NA>          <NA> FALSE
#> 42     <NA>         <NA>          <NA> FALSE
#> 43     <NA>         <NA>          <NA> FALSE
#> 44     <NA>         <NA>          <NA> FALSE
#> 45     <NA>         <NA>          <NA> FALSE
#> 46     <NA>         <NA>          <NA> FALSE
# }
library(dplyr)
# Which neurons don't have a superclass, but possibly should
mnm.nsc=mcns_neuprint_meta("where:NOT exists(n.superclass)")
mnm.nsc |> count(statusLabel)
#>              statusLabel    n
#> 1              0.5assign   46
#> 2                 Anchor  282
#> 3                   Glia  329
#> 4          Hard to trace    2
#> 5                 Leaves  150
#> 6                 Orphan 3373
#> 7        Orphan hotknife  481
#> 8        Orphan-artifact 1450
#> 9           Out of scope  981
#> 10            PRT Orphan   22
#> 11 Prelim Roughly traced   19
#> 12        Primary Anchor  472
#> 13      RT Hard to trace    2
#> 14        Roughly traced    1
#> 15        Sensory Anchor  139
#> 16           Soma Anchor   25
#> 17           Unimportant  123
#> 18                  <NA>  173

# neurons that are RT or PRT should probably have a superclass
mnm.nscprt=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.statusLabel CONTAINS 'Roughly'")
mnm.nscprt |> count()
#>    n
#> 1 20