Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

library(dplyr)
# \donttest{
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')

# neurons without a superclass but quite a few synapses
mnm.nc=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.synweight>2000")
mnm.nc |> arrange(desc(synweight))
#>       bodyid post  pre downstream upstream synweight           statusLabel
#> 1     331473  615 1176       7695      615      8310            PRT Orphan
#> 2     533707  177  884       7343      177      7520       Orphan-artifact
#> 3     555399 2504 1425       4175     2504      6679           Unimportant
#> 4     519161 2490  326       2724     2490      5214                Orphan
#> 5     944585  653  562       4534      653      5187                Orphan
#> 6     555297  274  607       3919      274      4193                Orphan
#> 7      88667 1925  222       1702     1925      3627                Anchor
#> 8  236599764   98  531       3509       98      3607                Orphan
#> 9     806539  224  447       3313      224      3537       Orphan-artifact
#> 10    232828 1312  351       2162     1312      3474 Prelim Roughly traced
#> 11    953607  129  429       3278      129      3407                  <NA>
#> 12    926152 1182  321       2184     1182      3366       Orphan-artifact
#> 13     73540  332  322       3016      332      3348 Prelim Roughly traced
#> 14    942401  403  465       2865      403      3268       Orphan-artifact
#> 15     66613 1678  247       1536     1678      3214 Prelim Roughly traced
#> 16     92244 1763  222       1426     1763      3189                Orphan
#> 17    946767  230  446       2750      230      2980                  <NA>
#> 18    946890  283  324       2546      283      2829                  <NA>
#> 19    559672 1407  248       1389     1407      2796                Orphan
#> 20    104626   53  319       2733       53      2786                Anchor
#> 21    114076  157  301       2535      157      2692                Orphan
#> 22    530376  446  244       2181      446      2627                Orphan
#> 23    934797 1566  264       1048     1566      2614            PRT Orphan
#> 24    906982   72  341       2503       72      2575       Orphan-artifact
#> 25    806609  106  321       2467      106      2573       Orphan-artifact
#> 26 516174168  159  308       2369      159      2528            PRT Orphan
#> 27    928870  399  241       2080      399      2479        Primary Anchor
#> 28     10621  679  557       1726      679      2405           Unimportant
#> 29    210839   79  269       2311       79      2390            PRT Orphan
#> 30    147605 1180  183       1208     1180      2388                Orphan
#> 31    555267   65  292       2314       65      2379                Anchor
#> 32    908376   26  235       2318       26      2344       Orphan-artifact
#> 33    909276   74  275       2262       74      2336       Orphan-artifact
#> 34    809279  117  280       2216      117      2333       Orphan-artifact
#> 35    941745  187  201       2128      187      2315       Orphan-artifact
#> 36    953574 1282  227        953     1282      2235                  <NA>
#> 37    210579 1432  179        788     1432      2220                Anchor
#> 38    810259   78  256       2134       78      2212       Orphan-artifact
#> 39    580488  416  202       1755      416      2171        Primary Anchor
#> 40    903906   73  327       2085       73      2158       Orphan-artifact
#> 41    953169  101  247       2055      101      2156                  <NA>
#> 42    819159   63  348       2078       63      2141       Orphan-artifact
#> 43    147742  257  204       1790      257      2047                Orphan
#> 44    946550  975  128       1068      975      2043                Orphan
#> 45    953507  672  189       1361      672      2033       Orphan-artifact
#> 46    141338 1403  103        612     1403      2015       Orphan-artifact
#>    superclass class entryNerve rootSide subclass      status      voxels
#> 1        <NA>  <NA>       <NA>     <NA>     <NA>      Traced   365414996
#> 2        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   404897796
#> 3        <NA>  <NA>       <NA>     <NA>     <NA> Unimportant 15774680322
#> 4        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   275039336
#> 5        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   240560432
#> 6        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   183037288
#> 7        <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   197698192
#> 8        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   161511491
#> 9        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   204307113
#> 10       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   418846190
#> 11       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   195663702
#> 12       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   833342245
#> 13       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   164205396
#> 14       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   244236620
#> 15       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   184607589
#> 16       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   172460933
#> 17       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   417909777
#> 18       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   419414444
#> 19       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   139035186
#> 20       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   118186516
#> 21       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   127874261
#> 22       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   152621707
#> 23       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   136391713
#> 24       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   205761867
#> 25       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   183589861
#> 26       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   147509581
#> 27       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   143101585
#> 28       <NA>  <NA>       <NA>     <NA>     <NA> Unimportant  6475367429
#> 29       <NA>  <NA>       <NA>     <NA>     <NA>      Traced    66258699
#> 30       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   118445955
#> 31       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   135641183
#> 32       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   115459022
#> 33       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   169169631
#> 34       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   229596508
#> 35       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   116788659
#> 36       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   564805869
#> 37       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   100881025
#> 38       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   230638948
#> 39       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   260194496
#> 40       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   176044266
#> 41       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>    71780552
#> 42       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   316825823
#> 43       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan    88932916
#> 44       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   129131037
#> 45       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   212778138
#> 46       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan    79808810
#>    somaSide        somaLocation        itoleeHl anteriorCervicalRootPosition
#> 1      <NA>                                <NA>                         <NA>
#> 2      <NA>                                <NA>                         <NA>
#> 3      <NA>                                <NA>                         <NA>
#> 4         L                                <NA>                         <NA>
#> 5      <NA>                                <NA>                         <NA>
#> 6      <NA>                                <NA>                         <NA>
#> 7      <NA>                                <NA>                         <NA>
#> 8      <NA>                                <NA>                         <NA>
#> 9      <NA>                                <NA>                         <NA>
#> 10     <NA> 73607, 21955, 19078 VLPl&d1_lateral                         <NA>
#> 11     <NA>                                <NA>                         <NA>
#> 12     <NA>                                <NA>                         <NA>
#> 13     <NA>                                <NA>                         <NA>
#> 14     <NA>                                <NA>                         <NA>
#> 15     <NA>                                <NA>                         <NA>
#> 16     <NA>                                <NA>                         <NA>
#> 17     <NA>                                <NA>                         <NA>
#> 18     <NA>                                <NA>                         <NA>
#> 19     <NA>                                <NA>                         <NA>
#> 20     <NA>                                <NA>                         <NA>
#> 21     <NA>                                <NA>                         <NA>
#> 22     <NA>                                <NA>                         <NA>
#> 23     <NA>                                <NA>                         <NA>
#> 24     <NA>                                <NA>                         <NA>
#> 25     <NA>                                <NA>                         <NA>
#> 26     <NA>                                <NA>                         <NA>
#> 27     <NA>                                <NA>                         <NA>
#> 28     <NA>                                <NA>                         <NA>
#> 29     <NA>                                <NA>                         <NA>
#> 30     <NA>                                <NA>                         <NA>
#> 31     <NA>                                <NA>                         <NA>
#> 32     <NA>                                <NA>                         <NA>
#> 33     <NA>                                <NA>                         <NA>
#> 34     <NA>                                <NA>                         <NA>
#> 35     <NA>                                <NA>                         <NA>
#> 36     <NA>                                <NA>                         <NA>
#> 37     <NA>                                <NA>                         <NA>
#> 38     <NA>                                <NA>                         <NA>
#> 39     <NA> 74352, 28712, 20890            <NA>                         <NA>
#> 40     <NA>                                <NA>                         <NA>
#> 41     <NA>                                <NA>                         <NA>
#> 42     <NA>                                <NA>                         <NA>
#> 43     <NA>                                <NA>                         <NA>
#> 44     <NA>                                <NA>                         <NA>
#> 45     <NA>                                <NA>                         <NA>
#> 46     <NA>                                <NA>                         <NA>
#>    synonyms fruDsx mancType somaNeuromere trumanHl birthtime group mancBodyid
#> 1      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 2      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 3      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 4      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 5      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 6      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 7      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 8      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 9      <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA     153319
#> 10     <NA>   <NA>     <NA>          <NA>     <NA>      <NA> 28005         NA
#> 11     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 12     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA      11325
#> 13     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 14     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 15     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 16     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 17     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 18     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 19     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 20     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 21     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 22     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 23     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 24     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 25     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA     155040
#> 26     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 27     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 28     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 29     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 30     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 31     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 32     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 33     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA      28494
#> 34     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA      16226
#> 35     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 36     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 37     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 38     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA      29891
#> 39     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 40     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 41     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 42     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 43     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 44     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 45     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#> 46     <NA>   <NA>     <NA>          <NA>     <NA>      <NA>    NA         NA
#>    mancGroup type mcnsSerial mancSerial exitNerve tosomaLocation matchingNotes
#> 1         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 2         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 3         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 4         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 5         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 6         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 7         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 8         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 9         NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 10        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 11        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 12        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 13        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 14        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 15        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 16        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 17        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 18        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 19        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 20        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 21        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 22        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 23        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 24        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 25        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 26        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 27        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 28        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 29        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 30        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 31        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 32        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 33        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 34        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 35        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 36        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 37        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 38        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 39        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 40        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 41        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 42        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 43        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 44        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 45        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#> 46        NA <NA>         NA       <NA>      <NA>           <NA>          <NA>
#>    flywireType hemibrainType          name flywireId
#> 1         <NA>          <NA>          <NA>      <NA>
#> 2         <NA>          <NA>          <NA>      <NA>
#> 3         <NA>          <NA>          <NA>      <NA>
#> 4         <NA>          <NA> KC_R_fragment      <NA>
#> 5         <NA>          <NA>          <NA>      <NA>
#> 6         <NA>          <NA>          <NA>      <NA>
#> 7         <NA>          <NA>          <NA>      <NA>
#> 8         <NA>          <NA>          <NA>      <NA>
#> 9         <NA>          <NA>          <NA>      <NA>
#> 10        <NA>          <NA>          <NA>      <NA>
#> 11        <NA>          <NA>          <NA>      <NA>
#> 12        <NA>          <NA>          <NA>      <NA>
#> 13        <NA>          <NA>          <NA>      <NA>
#> 14        <NA>          <NA>          <NA>      <NA>
#> 15        <NA>          <NA>          <NA>      <NA>
#> 16        <NA>          <NA>          <NA>      <NA>
#> 17        <NA>          <NA>          <NA>      <NA>
#> 18        <NA>          <NA>          <NA>      <NA>
#> 19        <NA>          <NA>          <NA>      <NA>
#> 20        <NA>          <NA>          <NA>      <NA>
#> 21        <NA>          <NA>          <NA>      <NA>
#> 22        <NA>          <NA>          <NA>      <NA>
#> 23        <NA>          <NA>          <NA>      <NA>
#> 24        <NA>          <NA>          <NA>      <NA>
#> 25        <NA>          <NA>          <NA>      <NA>
#> 26        <NA>          <NA>          <NA>      <NA>
#> 27        <NA>          <NA>          <NA>      <NA>
#> 28        <NA>          <NA>          <NA>      <NA>
#> 29        <NA>          <NA>          <NA>      <NA>
#> 30        <NA>          <NA>          <NA>      <NA>
#> 31        <NA>          <NA>          <NA>      <NA>
#> 32        <NA>          <NA>          <NA>      <NA>
#> 33        <NA>          <NA>          <NA>      <NA>
#> 34        <NA>          <NA>          <NA>      <NA>
#> 35        <NA>          <NA>          <NA>      <NA>
#> 36        <NA>          <NA>          <NA>      <NA>
#> 37        <NA>          <NA> MBON_fragment      <NA>
#> 38        <NA>          <NA>          <NA>      <NA>
#> 39        <NA>          <NA>          <NA>      <NA>
#> 40        <NA>          <NA>          <NA>      <NA>
#> 41        <NA>          <NA>          <NA>      <NA>
#> 42        <NA>          <NA>          <NA>      <NA>
#> 43        <NA>          <NA>          <NA>      <NA>
#> 44        <NA>          <NA>          <NA>      <NA>
#> 45        <NA>          <NA>          <NA>      <NA>
#> 46        <NA>          <NA>          <NA>      <NA>
#>    posteriorCervicalRootPosition dimorphism serialMotif hemibrainBodyId
#> 1                           <NA>       <NA>        <NA>            <NA>
#> 2                           <NA>       <NA>        <NA>            <NA>
#> 3                           <NA>       <NA>        <NA>            <NA>
#> 4                           <NA>       <NA>        <NA>            <NA>
#> 5                           <NA>       <NA>        <NA>            <NA>
#> 6                           <NA>       <NA>        <NA>            <NA>
#> 7                           <NA>       <NA>        <NA>            <NA>
#> 8                           <NA>       <NA>        <NA>            <NA>
#> 9                           <NA>       <NA>        <NA>            <NA>
#> 10                          <NA>       <NA>        <NA>            <NA>
#> 11                          <NA>       <NA>        <NA>            <NA>
#> 12                          <NA>       <NA>        <NA>            <NA>
#> 13                          <NA>       <NA>        <NA>            <NA>
#> 14                          <NA>       <NA>        <NA>            <NA>
#> 15                          <NA>       <NA>        <NA>            <NA>
#> 16                          <NA>       <NA>        <NA>            <NA>
#> 17                          <NA>       <NA>        <NA>            <NA>
#> 18                          <NA>       <NA>        <NA>            <NA>
#> 19                          <NA>       <NA>        <NA>            <NA>
#> 20                          <NA>       <NA>        <NA>            <NA>
#> 21                          <NA>       <NA>        <NA>            <NA>
#> 22                          <NA>       <NA>        <NA>            <NA>
#> 23                          <NA>       <NA>        <NA>            <NA>
#> 24                          <NA>       <NA>        <NA>            <NA>
#> 25                          <NA>       <NA>        <NA>            <NA>
#> 26                          <NA>       <NA>        <NA>            <NA>
#> 27                          <NA>       <NA>        <NA>            <NA>
#> 28                          <NA>       <NA>        <NA>            <NA>
#> 29                          <NA>       <NA>        <NA>            <NA>
#> 30                          <NA>       <NA>        <NA>            <NA>
#> 31                          <NA>       <NA>        <NA>            <NA>
#> 32                          <NA>       <NA>        <NA>            <NA>
#> 33                          <NA>       <NA>        <NA>            <NA>
#> 34                          <NA>       <NA>        <NA>            <NA>
#> 35                          <NA>       <NA>        <NA>            <NA>
#> 36                          <NA>       <NA>        <NA>            <NA>
#> 37                          <NA>       <NA>        <NA>            <NA>
#> 38                          <NA>       <NA>        <NA>            <NA>
#> 39                          <NA>       <NA>        <NA>            <NA>
#> 40                          <NA>       <NA>        <NA>            <NA>
#> 41                          <NA>       <NA>        <NA>            <NA>
#> 42                          <NA>       <NA>        <NA>            <NA>
#> 43                          <NA>       <NA>        <NA>            <NA>
#> 44                          <NA>       <NA>        <NA>            <NA>
#> 45                          <NA>       <NA>        <NA>            <NA>
#> 46                          <NA>       <NA>        <NA>            <NA>
#>    assignedOlHex1 assignedOlHex2 locationType cellBodyFiber  soma
#> 1              NA             NA         <NA>          <NA> FALSE
#> 2              NA             NA         <NA>          <NA> FALSE
#> 3              NA             NA         <NA>          <NA> FALSE
#> 4              NA             NA         <NA>          <NA> FALSE
#> 5              NA             NA         <NA>          <NA> FALSE
#> 6              NA             NA         <NA>          <NA> FALSE
#> 7              NA             NA         <NA>          <NA> FALSE
#> 8              NA             NA         <NA>          <NA> FALSE
#> 9              NA             NA         <NA>          <NA> FALSE
#> 10             NA             NA         <NA>          <NA>  TRUE
#> 11             NA             NA         <NA>          <NA> FALSE
#> 12             NA             NA         <NA>          <NA> FALSE
#> 13             NA             NA         <NA>          <NA> FALSE
#> 14             NA             NA         <NA>          <NA> FALSE
#> 15             NA             NA         <NA>          <NA> FALSE
#> 16             NA             NA         <NA>          <NA> FALSE
#> 17             NA             NA         <NA>          <NA> FALSE
#> 18             NA             NA         <NA>          <NA> FALSE
#> 19             NA             NA         <NA>          <NA> FALSE
#> 20             NA             NA         <NA>          <NA> FALSE
#> 21             NA             NA         <NA>          <NA> FALSE
#> 22             NA             NA         <NA>          <NA> FALSE
#> 23             NA             NA         <NA>          <NA> FALSE
#> 24             NA             NA         <NA>          <NA> FALSE
#> 25             NA             NA         <NA>          <NA> FALSE
#> 26             NA             NA         <NA>          <NA> FALSE
#> 27             NA             NA         <NA>          <NA> FALSE
#> 28             NA             NA         <NA>          <NA> FALSE
#> 29             NA             NA         <NA>          <NA> FALSE
#> 30             NA             NA         <NA>          <NA> FALSE
#> 31             NA             NA         <NA>          <NA> FALSE
#> 32             NA             NA         <NA>          <NA> FALSE
#> 33             NA             NA         <NA>          <NA> FALSE
#> 34             NA             NA         <NA>          <NA> FALSE
#> 35             NA             NA         <NA>          <NA> FALSE
#> 36             NA             NA         <NA>          <NA> FALSE
#> 37             NA             NA         <NA>          <NA> FALSE
#> 38             NA             NA         <NA>          <NA> FALSE
#> 39             NA             NA         <NA>          <NA>  TRUE
#> 40             NA             NA         <NA>          <NA> FALSE
#> 41             NA             NA         <NA>          <NA> FALSE
#> 42             NA             NA         <NA>          <NA> FALSE
#> 43             NA             NA         <NA>          <NA> FALSE
#> 44             NA             NA         <NA>          <NA> FALSE
#> 45             NA             NA         <NA>          <NA> FALSE
#> 46             NA             NA         <NA>          <NA> FALSE
# }

# Which neurons don't have a superclass, but possibly should
mnm.nsc=mcns_neuprint_meta("where:NOT exists(n.superclass)")
mnm.nsc |> count(statusLabel)
#>              statusLabel    n
#> 1              0.5assign   60
#> 2                 Anchor  300
#> 3                   Glia  516
#> 4          Hard to trace    1
#> 5                 Leaves  222
#> 6                 Orphan 3697
#> 7        Orphan hotknife  657
#> 8        Orphan-artifact 1729
#> 9           Out of scope 1300
#> 10            PRT Orphan   42
#> 11 Prelim Roughly traced   69
#> 12        Primary Anchor  155
#> 13        Roughly traced    3
#> 14        Sensory Anchor  139
#> 15           Soma Anchor    9
#> 16           Unimportant  615
#> 17                  <NA>  335

# neurons that are RT or PRT should probably have a superclass
mnm.nscprt=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.statusLabel CONTAINS 'Roughly'")
mnm.nscprt |> count()
#>    n
#> 1 72