Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

# \donttest{
library(dplyr)
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')

# neurons without a superclass but quite a few synapses
mnm.nc=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.synweight>2000")
mnm.nc |> arrange(desc(synweight))
#>       bodyid post pre downstream upstream synweight           statusLabel
#> 1     955682 2696 851       5467     2696      8163                  <NA>
#> 2     955684  589 809       5744      589      6333                  <NA>
#> 3     954994 1344 700       4305     1344      5649                  <NA>
#> 4     957716 1258 466       4213     1258      5471 Prelim Roughly traced
#> 5     519161 2401 322       2717     2401      5118                Orphan
#> 6     957664  792 400       3361      792      4153 Prelim Roughly traced
#> 7     555297  236 604       3882      236      4118                Orphan
#> 8      88667 1908 225       1724     1908      3632                Anchor
#> 9     956907 1965 306       1531     1965      3496                  <NA>
#> 10    806539  202 447       3169      202      3371       Orphan-artifact
#> 11     73540  332 322       3016      332      3348 Prelim Roughly traced
#> 12    926152 1134 321       2087     1134      3221       Orphan-artifact
#> 13     66613 1686 245       1527     1686      3213 Prelim Roughly traced
#> 14    956675   55 421       3139       55      3194                  <NA>
#> 15     92244 1753 222       1426     1753      3179                Orphan
#> 16    942401  351 465       2705      351      3056       Orphan-artifact
#> 17    946767  198 446       2631      198      2829                  <NA>
#> 18    955927  163 401       2644      163      2807                  <NA>
#> 19    955284  146 350       2655      146      2801                  <NA>
#> 20    104626   53 318       2732       53      2785                Anchor
#> 21    559672 1395 247       1385     1395      2780                Orphan
#> 22    955135  235 291       2469      235      2704                  <NA>
#> 23    110922   62 267       2640       62      2702                Anchor
#> 24    114076  150 300       2529      150      2679                Orphan
#> 25 516174168  128 319       2472      128      2600            PRT Orphan
#> 26    955155  609 302       1982      609      2591                  <NA>
#> 27    955615  450 382       2109      450      2559                  <NA>
#> 28    118423  101 259       2436      101      2537                Anchor
#> 29    555399 1033 424       1491     1033      2524           Unimportant
#> 30    954991  392 308       2124      392      2516                  <NA>
#> 31    136752  375 245       2093      375      2468                  <NA>
#> 32    956789  648 254       1790      648      2438                  <NA>
#> 33    147605 1180 183       1207     1180      2387                Orphan
#> 34    956674  868 281       1506      868      2374                  <NA>
#> 35    954995  399 305       1971      399      2370                  <NA>
#> 36    555267   64 289       2300       64      2364                Anchor
#> 37    210839   70 268       2294       70      2364            PRT Orphan
#> 38    908376   26 235       2318       26      2344       Orphan-artifact
#> 39    954670  305 268       1943      305      2248                  <NA>
#> 40    809279  107 279       2117      107      2224       Orphan-artifact
#> 41    957751  616 218       1592      616      2208                  <NA>
#> 42    909276   57 275       2131       57      2188       Orphan-artifact
#> 43    810259   74 256       2039       74      2113       Orphan-artifact
#> 44    953574 1202 231        904     1202      2106                  <NA>
#> 45 290057755   58 224       2008       58      2066             0.5assign
#> 46    544325  238 254       1820      238      2058      RT Hard to trace
#> 47    147742  257 204       1790      257      2047                Orphan
#> 48    903906   52 327       1983       52      2035       Orphan-artifact
#> 49    946550  949 128       1067      949      2016                Orphan
#>    subclass superclass type class entryNerve rootSide      status
#> 1      <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 2      <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 3      <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 4      <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Traced
#> 5      <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 6      <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Traced
#> 7      <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 8      <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Anchor
#> 9      <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 10     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 11     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Traced
#> 12     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 13     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Traced
#> 14     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 15     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 16     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 17     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 18     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 19     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 20     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Anchor
#> 21     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 22     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 23     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Anchor
#> 24     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 25     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Traced
#> 26     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 27     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 28     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Anchor
#> 29     <NA>       <NA> <NA>  <NA>       <NA>     <NA> Unimportant
#> 30     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 31     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 32     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 33     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 34     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 35     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 36     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Anchor
#> 37     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Traced
#> 38     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 39     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 40     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 41     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 42     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 43     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 44     <NA>       <NA> <NA>  <NA>       <NA>     <NA>        <NA>
#> 45     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Assign
#> 46     <NA>       <NA> <NA>  <NA>       aPhN        R      Traced
#> 47     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 48     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#> 49     <NA>       <NA> <NA>  <NA>       <NA>     <NA>      Orphan
#>    totalNtPredictions predictedNtConfidence   predictedNt
#> 1                 851             0.6882885 acetylcholine
#> 2                 809             0.4953307       unclear
#> 3                 700             0.6453711 acetylcholine
#> 4                 466             0.6273553 acetylcholine
#> 5                 322             0.5889296          gaba
#> 6                 400             0.5098772     serotonin
#> 7                 604             0.6609908 acetylcholine
#> 8                 225             0.7586356          gaba
#> 9                 306             0.5789716 acetylcholine
#> 10                447             0.6723263 acetylcholine
#> 11                322             0.3063286       unclear
#> 12                321             0.8453475          gaba
#> 13                245             0.7979734          gaba
#> 14                421             0.6892378 acetylcholine
#> 15                222             0.8058083          gaba
#> 16                465             0.3824132       unclear
#> 17                446             0.6546597 acetylcholine
#> 18                401             0.7369105 acetylcholine
#> 19                350             0.7727820 acetylcholine
#> 20                318             0.6177278          gaba
#> 21                247             0.5847769          gaba
#> 22                291             0.5687093 acetylcholine
#> 23                267             0.7356939 acetylcholine
#> 24                300             0.7054706 acetylcholine
#> 25                319             0.5937066 acetylcholine
#> 26                302             0.6042480 acetylcholine
#> 27                382             0.3294308       unclear
#> 28                259             0.2974450       unclear
#> 29                424             0.3481269       unclear
#> 30                308             0.7316861 acetylcholine
#> 31                245             0.4915912       unclear
#> 32                254             0.7381366 acetylcholine
#> 33                183             0.8250495          gaba
#> 34                281             0.7124321 acetylcholine
#> 35                305             0.7235691 acetylcholine
#> 36                289             0.7616221          gaba
#> 37                268             0.6845148 acetylcholine
#> 38                235             0.7412565          gaba
#> 39                268             0.7327555 acetylcholine
#> 40                279             0.7412511 acetylcholine
#> 41                218             0.3450897       unclear
#> 42                275             0.3343678       unclear
#> 43                256             0.8236932          gaba
#> 44                231             0.8407942          gaba
#> 45                224             0.7097978 acetylcholine
#> 46                254             0.3840813       unclear
#> 47                204             0.4672385       unclear
#> 48                327             0.6248300 acetylcholine
#> 49                128             0.4877605       unclear
#>    celltypeTotalNtPredictions celltypePredictedNt celltypePredictedNtConfidence
#> 1                         851             unclear                            NA
#> 2                         809             unclear                            NA
#> 3                         700             unclear                            NA
#> 4                         466             unclear                            NA
#> 5                         322             unclear                            NA
#> 6                         400             unclear                            NA
#> 7                         604             unclear                            NA
#> 8                         225             unclear                            NA
#> 9                         306             unclear                            NA
#> 10                        447             unclear                            NA
#> 11                        322             unclear                            NA
#> 12                        321             unclear                            NA
#> 13                        245             unclear                            NA
#> 14                        421             unclear                            NA
#> 15                        222             unclear                            NA
#> 16                        465             unclear                            NA
#> 17                        446             unclear                            NA
#> 18                        401             unclear                            NA
#> 19                        350             unclear                            NA
#> 20                        318             unclear                            NA
#> 21                        247             unclear                            NA
#> 22                        291             unclear                            NA
#> 23                        267             unclear                            NA
#> 24                        300             unclear                            NA
#> 25                        319             unclear                            NA
#> 26                        302             unclear                            NA
#> 27                        382             unclear                            NA
#> 28                        259             unclear                            NA
#> 29                        424             unclear                            NA
#> 30                        308             unclear                            NA
#> 31                        245             unclear                            NA
#> 32                        254             unclear                            NA
#> 33                        183             unclear                            NA
#> 34                        281             unclear                            NA
#> 35                        305             unclear                            NA
#> 36                        289             unclear                            NA
#> 37                        268             unclear                            NA
#> 38                        235             unclear                            NA
#> 39                        268             unclear                            NA
#> 40                        279             unclear                            NA
#> 41                        218             unclear                            NA
#> 42                        275             unclear                            NA
#> 43                        256             unclear                            NA
#> 44                        231             unclear                            NA
#> 45                        224             unclear                            NA
#> 46                        254             unclear                            NA
#> 47                        204             unclear                            NA
#> 48                        327             unclear                            NA
#> 49                        128             unclear                            NA
#>      consensusNt      voxels flywireType itoleeHl somaSide supertype
#> 1  acetylcholine   692326498        <NA>     <NA>     <NA>      <NA>
#> 2        unclear   476087509        <NA>     <NA>     <NA>      <NA>
#> 3  acetylcholine   769950427        <NA>     <NA>     <NA>      <NA>
#> 4  acetylcholine   481029228        <NA>     <NA>        R      <NA>
#> 5           gaba   278587841        <NA>     <NA>        L      <NA>
#> 6      serotonin   489236167        <NA>     <NA>     <NA>      <NA>
#> 7  acetylcholine   183037288        <NA>     <NA>     <NA>      <NA>
#> 8           gaba   199021086        <NA>     <NA>     <NA>      <NA>
#> 9  acetylcholine  1181128023        <NA>     <NA>     <NA>      <NA>
#> 10 acetylcholine   204307113        <NA>     <NA>     <NA>      <NA>
#> 11       unclear   164205396        <NA>     <NA>     <NA>      <NA>
#> 12          gaba   833342245        <NA>     <NA>     <NA>      <NA>
#> 13          gaba   185592243        <NA>     <NA>     <NA>      <NA>
#> 14 acetylcholine   217610342        <NA>     <NA>     <NA>      <NA>
#> 15          gaba   172460933        <NA>     <NA>     <NA>      <NA>
#> 16       unclear   244236620        <NA>     <NA>     <NA>      <NA>
#> 17 acetylcholine   417909777        <NA>     <NA>     <NA>      <NA>
#> 18 acetylcholine   291166682        <NA>     <NA>     <NA>      <NA>
#> 19 acetylcholine   346648352        <NA>     <NA>     <NA>      <NA>
#> 20          gaba   118186516        <NA>     <NA>     <NA>      <NA>
#> 21          gaba   139035186        <NA>     <NA>     <NA>      <NA>
#> 22 acetylcholine   107081588        <NA>     <NA>     <NA>      <NA>
#> 23 acetylcholine   130934995        <NA>     <NA>     <NA>      <NA>
#> 24 acetylcholine   127874261        <NA>     <NA>     <NA>      <NA>
#> 25 acetylcholine   148541348        <NA>     <NA>     <NA>      <NA>
#> 26 acetylcholine   393057814        <NA>     <NA>     <NA>      <NA>
#> 27       unclear   142895359        <NA>     <NA>     <NA>      <NA>
#> 28       unclear   115258843        <NA>     <NA>     <NA>      <NA>
#> 29       unclear 15774680322        <NA>     <NA>     <NA>      <NA>
#> 30 acetylcholine   291580020        <NA>     <NA>     <NA>      <NA>
#> 31       unclear   323241784        <NA>     <NA>        R      <NA>
#> 32 acetylcholine   415371655        <NA>     <NA>     <NA>      <NA>
#> 33          gaba   118445955        <NA>     <NA>     <NA>      <NA>
#> 34 acetylcholine   389656375        <NA>     <NA>     <NA>      <NA>
#> 35 acetylcholine   309045905        <NA>     <NA>     <NA>      <NA>
#> 36          gaba   135641183        <NA>     <NA>     <NA>      <NA>
#> 37 acetylcholine    66258699        <NA>     <NA>     <NA>      <NA>
#> 38          gaba   115459022        <NA>     <NA>     <NA>      <NA>
#> 39 acetylcholine   141835905        <NA>     <NA>     <NA>      <NA>
#> 40 acetylcholine   229596508        <NA>     <NA>     <NA>      <NA>
#> 41       unclear   254548622        <NA>     <NA>     <NA>      <NA>
#> 42       unclear   169169631        <NA>     <NA>     <NA>      <NA>
#> 43          gaba   230638948        <NA>     <NA>     <NA>      <NA>
#> 44          gaba   567492443        <NA>     <NA>     <NA>      <NA>
#> 45 acetylcholine    29944601        <NA>     <NA>     <NA>      <NA>
#> 46       unclear    93097711        <NA>     <NA>     <NA>      <NA>
#> 47       unclear    88932916        <NA>     <NA>     <NA>      <NA>
#> 48 acetylcholine   176044266        <NA>     <NA>     <NA>      <NA>
#> 49       unclear   129131037        <NA>     <NA>     <NA>      <NA>
#>    somaNeuromere trumanHl        somaLocation mancType group mancGroup
#> 1           <NA>     <NA>                         <NA>    NA        NA
#> 2           <NA>     <NA>                         <NA>    NA        NA
#> 3           <NA>     <NA>                         <NA>    NA        NA
#> 4           <NA>     <NA> 44983, 14505, 34017     <NA>    NA        NA
#> 5           <NA>     <NA>                         <NA>    NA        NA
#> 6           <NA>     <NA> 69682, 13204, 19654     <NA> 36888        NA
#> 7           <NA>     <NA>                         <NA>    NA        NA
#> 8           <NA>     <NA>                         <NA>    NA        NA
#> 9           <NA>     <NA>                         <NA>    NA        NA
#> 10          <NA>     <NA>                         <NA>    NA        NA
#> 11          <NA>     <NA>                         <NA>    NA        NA
#> 12          <NA>     <NA>                         <NA>    NA        NA
#> 13          <NA>     <NA>                         <NA>    NA        NA
#> 14          <NA>     <NA>                         <NA>    NA        NA
#> 15          <NA>     <NA>                         <NA>    NA        NA
#> 16          <NA>     <NA>                         <NA>    NA        NA
#> 17          <NA>     <NA>                         <NA>    NA        NA
#> 18          <NA>     <NA>                         <NA>    NA        NA
#> 19          <NA>     <NA>                         <NA>    NA        NA
#> 20          <NA>     <NA>                         <NA>    NA        NA
#> 21          <NA>     <NA>                         <NA>    NA        NA
#> 22          <NA>     <NA>                         <NA>    NA        NA
#> 23          <NA>     <NA>                         <NA>    NA        NA
#> 24          <NA>     <NA>                         <NA>    NA        NA
#> 25          <NA>     <NA>                         <NA>    NA        NA
#> 26          <NA>     <NA>                         <NA>    NA        NA
#> 27          <NA>     <NA>                         <NA>    NA        NA
#> 28          <NA>     <NA>                         <NA>    NA        NA
#> 29          <NA>     <NA>                         <NA>    NA        NA
#> 30          <NA>     <NA>                         <NA>    NA        NA
#> 31          <NA>     <NA> 31887, 20287, 16543     <NA>    NA        NA
#> 32          <NA>     <NA>                         <NA>    NA        NA
#> 33          <NA>     <NA>                         <NA>    NA        NA
#> 34          <NA>     <NA>                         <NA>    NA        NA
#> 35          <NA>     <NA>                         <NA>    NA        NA
#> 36          <NA>     <NA>                         <NA>    NA        NA
#> 37          <NA>     <NA>                         <NA>    NA        NA
#> 38          <NA>     <NA>                         <NA>    NA        NA
#> 39          <NA>     <NA>                         <NA>    NA        NA
#> 40          <NA>     <NA>                         <NA>    NA        NA
#> 41          <NA>     <NA>                         <NA>    NA        NA
#> 42          <NA>     <NA>                         <NA>    NA        NA
#> 43          <NA>     <NA>                         <NA>    NA        NA
#> 44          <NA>     <NA>                         <NA>    NA        NA
#> 45          <NA>     <NA>                         <NA>    NA        NA
#> 46          <NA>     <NA>                         <NA>    NA        NA
#> 47          <NA>     <NA>                         <NA>    NA        NA
#> 48          <NA>     <NA>                         <NA>    NA        NA
#> 49          <NA>     <NA>                         <NA>    NA        NA
#>    birthtime mancBodyid mcnsSerial mancSerial synonyms fruDsx exitNerve
#> 1       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 2       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 3       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 4       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 5       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 6       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 7       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 8       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 9       <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 10      <NA>     153319         NA         NA     <NA>   <NA>      <NA>
#> 11      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 12      <NA>      11325         NA         NA     <NA>   <NA>      <NA>
#> 13      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 14      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 15      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 16      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 17      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 18      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 19      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 20      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 21      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 22      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 23      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 24      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 25      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 26      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 27      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 28      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 29      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 30      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 31      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 32      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 33      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 34      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 35      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 36      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 37      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 38      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 39      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 40      <NA>      16226         NA         NA     <NA>   <NA>      <NA>
#> 41      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 42      <NA>      28494         NA         NA     <NA>   <NA>      <NA>
#> 43      <NA>      29891         NA         NA     <NA>   <NA>      <NA>
#> 44      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 45      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 46      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 47      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 48      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#> 49      <NA>         NA         NA         NA     <NA>   <NA>      <NA>
#>    tosomaLocation receptorType matchingNotes hemibrainType          name
#> 1            <NA>         <NA>          <NA>          <NA>          <NA>
#> 2            <NA>         <NA>          <NA>          <NA>          <NA>
#> 3            <NA>         <NA>          <NA>          <NA>          <NA>
#> 4            <NA>         <NA>          <NA>          <NA>          <NA>
#> 5            <NA>         <NA>          <NA>          <NA> KC_L_fragment
#> 6            <NA>         <NA>          <NA>          <NA>       36888_L
#> 7            <NA>         <NA>          <NA>          <NA>          <NA>
#> 8            <NA>         <NA>          <NA>          <NA>          <NA>
#> 9            <NA>         <NA>          <NA>          <NA>          <NA>
#> 10           <NA>         <NA>          <NA>          <NA>          <NA>
#> 11           <NA>         <NA>          <NA>          <NA>          <NA>
#> 12           <NA>         <NA>          <NA>          <NA>          <NA>
#> 13           <NA>         <NA>          <NA>          <NA>          <NA>
#> 14           <NA>         <NA>          <NA>          <NA>          <NA>
#> 15           <NA>         <NA>          <NA>          <NA>          <NA>
#> 16           <NA>         <NA>          <NA>          <NA>          <NA>
#> 17           <NA>         <NA>          <NA>          <NA>          <NA>
#> 18           <NA>         <NA>          <NA>          <NA>          <NA>
#> 19           <NA>         <NA>          <NA>          <NA>          <NA>
#> 20           <NA>         <NA>          <NA>          <NA>          <NA>
#> 21           <NA>         <NA>          <NA>          <NA>          <NA>
#> 22           <NA>         <NA>          <NA>          <NA>          <NA>
#> 23           <NA>         <NA>          <NA>          <NA>          <NA>
#> 24           <NA>         <NA>          <NA>          <NA>          <NA>
#> 25           <NA>         <NA>          <NA>          <NA>          <NA>
#> 26           <NA>         <NA>          <NA>          <NA>          <NA>
#> 27           <NA>         <NA>          <NA>          <NA>          <NA>
#> 28           <NA>         <NA>          <NA>          <NA>          <NA>
#> 29           <NA>         <NA>          <NA>          <NA>          <NA>
#> 30           <NA>         <NA>          <NA>          <NA>          <NA>
#> 31           <NA>         <NA>          <NA>          <NA>          <NA>
#> 32           <NA>         <NA>          <NA>          <NA>          <NA>
#> 33           <NA>         <NA>          <NA>          <NA>          <NA>
#> 34           <NA>         <NA>          <NA>          <NA>          <NA>
#> 35           <NA>         <NA>          <NA>          <NA>          <NA>
#> 36           <NA>         <NA>          <NA>          <NA>          <NA>
#> 37           <NA>         <NA>          <NA>          <NA>          <NA>
#> 38           <NA>         <NA>          <NA>          <NA>          <NA>
#> 39           <NA>         <NA>          <NA>          <NA>          <NA>
#> 40           <NA>         <NA>          <NA>          <NA>          <NA>
#> 41           <NA>         <NA>          <NA>          <NA>          <NA>
#> 42           <NA>         <NA>          <NA>          <NA>          <NA>
#> 43           <NA>         <NA>          <NA>          <NA>          <NA>
#> 44           <NA>         <NA>          <NA>          <NA>          <NA>
#> 45           <NA>         <NA>          <NA>          <NA>          <NA>
#> 46           <NA>         <NA>          <NA>          <NA>          <NA>
#> 47           <NA>         <NA>          <NA>          <NA>          <NA>
#> 48           <NA>         <NA>          <NA>          <NA>          <NA>
#> 49           <NA>         <NA>          <NA>          <NA>          <NA>
#>    anteriorCervicalRootPosition flywireId posteriorCervicalRootPosition
#> 1                          <NA>      <NA>                          <NA>
#> 2                          <NA>      <NA>                          <NA>
#> 3                          <NA>      <NA>                          <NA>
#> 4                          <NA>      <NA>                          <NA>
#> 5                          <NA>      <NA>                          <NA>
#> 6                          <NA>      <NA>                          <NA>
#> 7                          <NA>      <NA>                          <NA>
#> 8                          <NA>      <NA>                          <NA>
#> 9                          <NA>      <NA>                          <NA>
#> 10                         <NA>      <NA>                          <NA>
#> 11                         <NA>      <NA>                          <NA>
#> 12                         <NA>      <NA>                          <NA>
#> 13                         <NA>      <NA>                          <NA>
#> 14                         <NA>      <NA>                          <NA>
#> 15                         <NA>      <NA>                          <NA>
#> 16                         <NA>      <NA>                          <NA>
#> 17                         <NA>      <NA>                          <NA>
#> 18                         <NA>      <NA>                          <NA>
#> 19                         <NA>      <NA>                          <NA>
#> 20                         <NA>      <NA>                          <NA>
#> 21                         <NA>      <NA>                          <NA>
#> 22                         <NA>      <NA>                          <NA>
#> 23                         <NA>      <NA>                          <NA>
#> 24                         <NA>      <NA>                          <NA>
#> 25                         <NA>      <NA>                          <NA>
#> 26                         <NA>      <NA>                          <NA>
#> 27                         <NA>      <NA>                          <NA>
#> 28                         <NA>      <NA>                          <NA>
#> 29                         <NA>      <NA>                          <NA>
#> 30                         <NA>      <NA>                          <NA>
#> 31                         <NA>      <NA>                          <NA>
#> 32                         <NA>      <NA>                          <NA>
#> 33                         <NA>      <NA>                          <NA>
#> 34                         <NA>      <NA>                          <NA>
#> 35                         <NA>      <NA>                          <NA>
#> 36                         <NA>      <NA>                          <NA>
#> 37                         <NA>      <NA>                          <NA>
#> 38                         <NA>      <NA>                          <NA>
#> 39                         <NA>      <NA>                          <NA>
#> 40                         <NA>      <NA>                          <NA>
#> 41                         <NA>      <NA>                          <NA>
#> 42                         <NA>      <NA>                          <NA>
#> 43                         <NA>      <NA>                          <NA>
#> 44                         <NA>      <NA>                          <NA>
#> 45                         <NA>      <NA>                          <NA>
#> 46                         <NA>      <NA>                          <NA>
#> 47                         <NA>      <NA>                          <NA>
#> 48                         <NA>      <NA>                          <NA>
#> 49                         <NA>      <NA>                          <NA>
#>    serialMotif dimorphism hemibrainBodyId assignedOlHex1 assignedOlHex2
#> 1         <NA>       <NA>            <NA>             NA             NA
#> 2         <NA>       <NA>            <NA>             NA             NA
#> 3         <NA>       <NA>            <NA>             NA             NA
#> 4         <NA>       <NA>            <NA>             NA             NA
#> 5         <NA>       <NA>            <NA>             NA             NA
#> 6         <NA>       <NA>            <NA>             NA             NA
#> 7         <NA>       <NA>            <NA>             NA             NA
#> 8         <NA>       <NA>            <NA>             NA             NA
#> 9         <NA>       <NA>            <NA>             NA             NA
#> 10        <NA>       <NA>            <NA>             NA             NA
#> 11        <NA>       <NA>            <NA>             NA             NA
#> 12        <NA>       <NA>            <NA>             NA             NA
#> 13        <NA>       <NA>            <NA>             NA             NA
#> 14        <NA>       <NA>            <NA>             NA             NA
#> 15        <NA>       <NA>            <NA>             NA             NA
#> 16        <NA>       <NA>            <NA>             NA             NA
#> 17        <NA>       <NA>            <NA>             NA             NA
#> 18        <NA>       <NA>            <NA>             NA             NA
#> 19        <NA>       <NA>            <NA>             NA             NA
#> 20        <NA>       <NA>            <NA>             NA             NA
#> 21        <NA>       <NA>            <NA>             NA             NA
#> 22        <NA>       <NA>            <NA>             NA             NA
#> 23        <NA>       <NA>            <NA>             NA             NA
#> 24        <NA>       <NA>            <NA>             NA             NA
#> 25        <NA>       <NA>            <NA>             NA             NA
#> 26        <NA>       <NA>            <NA>             NA             NA
#> 27        <NA>       <NA>            <NA>             NA             NA
#> 28        <NA>       <NA>            <NA>             NA             NA
#> 29        <NA>       <NA>            <NA>             NA             NA
#> 30        <NA>       <NA>            <NA>             NA             NA
#> 31        <NA>       <NA>            <NA>             NA             NA
#> 32        <NA>       <NA>            <NA>             NA             NA
#> 33        <NA>       <NA>            <NA>             NA             NA
#> 34        <NA>       <NA>            <NA>             NA             NA
#> 35        <NA>       <NA>            <NA>             NA             NA
#> 36        <NA>       <NA>            <NA>             NA             NA
#> 37        <NA>       <NA>            <NA>             NA             NA
#> 38        <NA>       <NA>            <NA>             NA             NA
#> 39        <NA>       <NA>            <NA>             NA             NA
#> 40        <NA>       <NA>            <NA>             NA             NA
#> 41        <NA>       <NA>            <NA>             NA             NA
#> 42        <NA>       <NA>            <NA>             NA             NA
#> 43        <NA>       <NA>            <NA>             NA             NA
#> 44        <NA>       <NA>            <NA>             NA             NA
#> 45        <NA>       <NA>            <NA>             NA             NA
#> 46        <NA>       <NA>            <NA>             NA             NA
#> 47        <NA>       <NA>            <NA>             NA             NA
#> 48        <NA>       <NA>            <NA>             NA             NA
#> 49        <NA>       <NA>            <NA>             NA             NA
#>    location locationType cellBodyFiber  soma
#> 1      <NA>         <NA>          <NA> FALSE
#> 2      <NA>         <NA>          <NA> FALSE
#> 3      <NA>         <NA>          <NA> FALSE
#> 4      <NA>         <NA>          <NA>  TRUE
#> 5      <NA>         <NA>          <NA> FALSE
#> 6      <NA>         <NA>          <NA>  TRUE
#> 7      <NA>         <NA>          <NA> FALSE
#> 8      <NA>         <NA>          <NA> FALSE
#> 9      <NA>         <NA>          <NA> FALSE
#> 10     <NA>         <NA>          <NA> FALSE
#> 11     <NA>         <NA>          <NA> FALSE
#> 12     <NA>         <NA>          <NA> FALSE
#> 13     <NA>         <NA>          <NA> FALSE
#> 14     <NA>         <NA>          <NA> FALSE
#> 15     <NA>         <NA>          <NA> FALSE
#> 16     <NA>         <NA>          <NA> FALSE
#> 17     <NA>         <NA>          <NA> FALSE
#> 18     <NA>         <NA>          <NA> FALSE
#> 19     <NA>         <NA>          <NA> FALSE
#> 20     <NA>         <NA>          <NA> FALSE
#> 21     <NA>         <NA>          <NA> FALSE
#> 22     <NA>         <NA>          <NA> FALSE
#> 23     <NA>         <NA>          <NA> FALSE
#> 24     <NA>         <NA>          <NA> FALSE
#> 25     <NA>         <NA>          <NA> FALSE
#> 26     <NA>         <NA>          <NA> FALSE
#> 27     <NA>         <NA>          <NA> FALSE
#> 28     <NA>         <NA>          <NA> FALSE
#> 29     <NA>         <NA>          <NA> FALSE
#> 30     <NA>         <NA>          <NA> FALSE
#> 31     <NA>         <NA>          <NA>  TRUE
#> 32     <NA>         <NA>          <NA> FALSE
#> 33     <NA>         <NA>          <NA> FALSE
#> 34     <NA>         <NA>          <NA> FALSE
#> 35     <NA>         <NA>          <NA> FALSE
#> 36     <NA>         <NA>          <NA> FALSE
#> 37     <NA>         <NA>          <NA> FALSE
#> 38     <NA>         <NA>          <NA> FALSE
#> 39     <NA>         <NA>          <NA> FALSE
#> 40     <NA>         <NA>          <NA> FALSE
#> 41     <NA>         <NA>          <NA> FALSE
#> 42     <NA>         <NA>          <NA> FALSE
#> 43     <NA>         <NA>          <NA> FALSE
#> 44     <NA>         <NA>          <NA> FALSE
#> 45     <NA>         <NA>          <NA> FALSE
#> 46     <NA>         <NA>          <NA> FALSE
#> 47     <NA>         <NA>          <NA> FALSE
#> 48     <NA>         <NA>          <NA> FALSE
#> 49     <NA>         <NA>          <NA> FALSE
# }
library(dplyr)
# Which neurons don't have a superclass, but possibly should
mnm.nsc=mcns_neuprint_meta("where:NOT exists(n.superclass)")
mnm.nsc |> count(statusLabel)
#>              statusLabel    n
#> 1              0.5assign   48
#> 2                 Anchor  528
#> 3                   Glia  389
#> 4          Hard to trace    3
#> 5                 Leaves  411
#> 6                 Orphan 3516
#> 7        Orphan hotknife 1017
#> 8        Orphan-artifact 1932
#> 9           Out of scope 1599
#> 10            PRT Orphan   23
#> 11 Prelim Roughly traced   24
#> 12        Primary Anchor  494
#> 13      RT Hard to trace    2
#> 14        Roughly traced    2
#> 15        Sensory Anchor  195
#> 16           Soma Anchor   32
#> 17           Unimportant  162
#> 18                  <NA>  463

# neurons that are RT or PRT should probably have a superclass
mnm.nscprt=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.statusLabel CONTAINS 'Roughly'")
mnm.nscprt |> count()
#>    n
#> 1 26