Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

# \donttest{
library(dplyr)
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')

# neurons without a superclass but quite a few synapses
mnm.nc=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.synweight>2000")
mnm.nc |> arrange(desc(synweight))
#>       bodyid post  pre downstream upstream synweight flywireType          name
#> 1     955682 2696  851       5467     2696      8163        <NA>          <NA>
#> 2     957899 1776 1343       5431     1776      7207        <NA>          <NA>
#> 3     955684  589  809       5744      589      6333        <NA>          <NA>
#> 4     954994 1344  700       4305     1344      5649        <NA>          <NA>
#> 5     957716 1258  466       4213     1258      5471        <NA>          <NA>
#> 6     519161 2401  322       2717     2401      5118        <NA> KC_L_fragment
#> 7     957664  792  400       3361      792      4153        <NA>       36888_L
#> 8     555297  236  604       3882      236      4118        <NA>          <NA>
#> 9      88667 1908  225       1724     1908      3632        <NA>          <NA>
#> 10    956907 1965  306       1531     1965      3496        <NA>          <NA>
#> 11    806539  202  447       3169      202      3371        <NA>          <NA>
#> 12     73540  332  322       3016      332      3348        <NA>          <NA>
#> 13    926152 1134  321       2087     1134      3221        <NA>          <NA>
#> 14     66613 1686  245       1527     1686      3213        <NA>          <NA>
#> 15    956675   55  421       3139       55      3194        <NA>          <NA>
#> 16     92244 1753  222       1426     1753      3179        <NA>          <NA>
#> 17    942401  351  465       2705      351      3056        <NA>          <NA>
#> 18    946767  198  446       2631      198      2829        <NA>          <NA>
#> 19    955927  163  401       2644      163      2807        <NA>          <NA>
#> 20    955284  146  350       2655      146      2801        <NA>          <NA>
#> 21    104626   53  318       2732       53      2785        <NA>          <NA>
#> 22    559672 1395  247       1385     1395      2780        <NA>          <NA>
#> 23    955135  235  291       2469      235      2704        <NA>          <NA>
#> 24    110922   62  267       2640       62      2702        <NA>          <NA>
#> 25    114076  150  300       2529      150      2679        <NA>          <NA>
#> 26 516174168  128  319       2472      128      2600        <NA>          <NA>
#> 27    955155  609  302       1982      609      2591        <NA>          <NA>
#> 28    955615  450  382       2109      450      2559        <NA>          <NA>
#> 29    118423  101  259       2436      101      2537        <NA>          <NA>
#> 30    555399 1033  424       1491     1033      2524        <NA>          <NA>
#> 31    954991  392  308       2124      392      2516        <NA>          <NA>
#> 32    136752  375  245       2093      375      2468        <NA>          <NA>
#> 33    956789  648  254       1790      648      2438        <NA>          <NA>
#> 34    147605 1180  183       1207     1180      2387        <NA>          <NA>
#> 35    956674  868  281       1506      868      2374        <NA>          <NA>
#> 36    954995  399  305       1971      399      2370        <NA>          <NA>
#> 37    555267   64  289       2300       64      2364        <NA>          <NA>
#> 38    210839   70  268       2294       70      2364        <NA>          <NA>
#> 39    908376   26  235       2318       26      2344        <NA>          <NA>
#> 40    954670  305  268       1943      305      2248        <NA>          <NA>
#> 41    809279  107  279       2117      107      2224        <NA>          <NA>
#> 42     67566  621  219       1595      621      2216        <NA>          <NA>
#> 43    909276   57  275       2131       57      2188        <NA>          <NA>
#> 44    810259   74  256       2039       74      2113        <NA>          <NA>
#> 45    953574 1202  231        904     1202      2106        <NA>          <NA>
#> 46 290057755   58  224       2008       58      2066        <NA>          <NA>
#> 47    544325  238  254       1820      238      2058        <NA>          <NA>
#> 48    147742  257  204       1790      257      2047        <NA>          <NA>
#> 49    903906   52  327       1983       52      2035        <NA>          <NA>
#> 50    946550  949  128       1067      949      2016        <NA>          <NA>
#>              statusLabel subclass superclass type class entryNerve rootSide
#> 1                   <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 2         Primary Anchor     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 3                   <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 4                   <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 5  Prelim Roughly traced     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 6                 Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 7  Prelim Roughly traced     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 8                 Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 9                 Anchor     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 10                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 11       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 12 Prelim Roughly traced     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 13       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 14 Prelim Roughly traced     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 15                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 16                Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 17       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 18                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 19                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 20                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 21                Anchor     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 22                Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 23                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 24                Anchor     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 25                Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 26            PRT Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 27                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 28                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 29                Anchor     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 30           Unimportant     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 31                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 32                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 33                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 34                Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 35                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 36                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 37                Anchor     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 38            PRT Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 39       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 40                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 41       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 42 Prelim Roughly traced     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 43       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 44       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 45                  <NA>     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 46             0.5assign     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 47      RT Hard to trace     <NA>       <NA> <NA>  <NA>       aPhN        R
#> 48                Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 49       Orphan-artifact     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#> 50                Orphan     <NA>       <NA> <NA>  <NA>       <NA>     <NA>
#>         status totalNtPredictions predictedNtConfidence   predictedNt
#> 1         <NA>                851             0.6925855 acetylcholine
#> 2       Anchor               1343             0.3761587       unclear
#> 3         <NA>                809             0.4953595       unclear
#> 4         <NA>                700             0.6495688 acetylcholine
#> 5       Traced                466             0.6308490 acetylcholine
#> 6       Orphan                322             0.5889654          gaba
#> 7       Traced                400             0.5098772     serotonin
#> 8       Orphan                604             0.6644562 acetylcholine
#> 9       Anchor                225             0.7586910          gaba
#> 10        <NA>                306             0.5828933 acetylcholine
#> 11      Orphan                447             0.6763132 acetylcholine
#> 12      Traced                322             0.3063286       unclear
#> 13      Orphan                321             0.8454184          gaba
#> 14      Traced                245             0.7980360          gaba
#> 15        <NA>                421             0.6934516 acetylcholine
#> 16      Orphan                222             0.8058730          gaba
#> 17      Orphan                465             0.3849989       unclear
#> 18        <NA>                446             0.6588567 acetylcholine
#> 19        <NA>                401             0.7412854 acetylcholine
#> 20        <NA>                350             0.7775649 acetylcholine
#> 21      Anchor                318             0.6177690          gaba
#> 22      Orphan                247             0.5848054          gaba
#> 23        <NA>                291             0.5725706 acetylcholine
#> 24      Anchor                267             0.7401769 acetylcholine
#> 25      Orphan                300             0.7093383 acetylcholine
#> 26      Traced                319             0.5962877 acetylcholine
#> 27        <NA>                302             0.6082783 acetylcholine
#> 28        <NA>                382             0.3298464       unclear
#> 29      Anchor                259             0.2973763       unclear
#> 30 Unimportant                424             0.3488079       unclear
#> 31        <NA>                308             0.7362820 acetylcholine
#> 32        <NA>                245             0.4927335       unclear
#> 33        <NA>                254             0.7426522 acetylcholine
#> 34      Orphan                183             0.8251171          gaba
#> 35        <NA>                281             0.7165631 acetylcholine
#> 36        <NA>                305             0.7281367 acetylcholine
#> 37      Anchor                289             0.7616830          gaba
#> 38      Traced                268             0.6886637 acetylcholine
#> 39      Orphan                235             0.7413131          gaba
#> 40        <NA>                268             0.7373407 acetylcholine
#> 41      Orphan                279             0.7457504 acetylcholine
#> 42      Traced                219             0.3435284       unclear
#> 43      Orphan                275             0.3343678       unclear
#> 44      Orphan                256             0.8237638          gaba
#> 45        <NA>                231             0.8408666          gaba
#> 46      Assign                224             0.7136616 acetylcholine
#> 47      Traced                254             0.3840813       unclear
#> 48      Orphan                204             0.4672385       unclear
#> 49      Orphan                327             0.6285104 acetylcholine
#> 50      Orphan                128             0.4877978       unclear
#>    celltypeTotalNtPredictions celltypePredictedNt celltypePredictedNtConfidence
#> 1                         851             unclear                            NA
#> 2                        1343             unclear                            NA
#> 3                         809             unclear                            NA
#> 4                         700             unclear                            NA
#> 5                         466             unclear                            NA
#> 6                         322             unclear                            NA
#> 7                         400             unclear                            NA
#> 8                         604             unclear                            NA
#> 9                         225             unclear                            NA
#> 10                        306             unclear                            NA
#> 11                        447             unclear                            NA
#> 12                        322             unclear                            NA
#> 13                        321             unclear                            NA
#> 14                        245             unclear                            NA
#> 15                        421             unclear                            NA
#> 16                        222             unclear                            NA
#> 17                        465             unclear                            NA
#> 18                        446             unclear                            NA
#> 19                        401             unclear                            NA
#> 20                        350             unclear                            NA
#> 21                        318             unclear                            NA
#> 22                        247             unclear                            NA
#> 23                        291             unclear                            NA
#> 24                        267             unclear                            NA
#> 25                        300             unclear                            NA
#> 26                        319             unclear                            NA
#> 27                        302             unclear                            NA
#> 28                        382             unclear                            NA
#> 29                        259             unclear                            NA
#> 30                        424             unclear                            NA
#> 31                        308             unclear                            NA
#> 32                        245             unclear                            NA
#> 33                        254             unclear                            NA
#> 34                        183             unclear                            NA
#> 35                        281             unclear                            NA
#> 36                        305             unclear                            NA
#> 37                        289             unclear                            NA
#> 38                        268             unclear                            NA
#> 39                        235             unclear                            NA
#> 40                        268             unclear                            NA
#> 41                        279             unclear                            NA
#> 42                        219             unclear                            NA
#> 43                        275             unclear                            NA
#> 44                        256             unclear                            NA
#> 45                        231             unclear                            NA
#> 46                        224             unclear                            NA
#> 47                        254             unclear                            NA
#> 48                        204             unclear                            NA
#> 49                        327             unclear                            NA
#> 50                        128             unclear                            NA
#>      consensusNt      voxels itoleeHl somaSide supertype somaNeuromere trumanHl
#> 1  acetylcholine   692326498     <NA>     <NA>      <NA>          <NA>     <NA>
#> 2        unclear  1275432396     <NA>     <NA>      <NA>          <NA>     <NA>
#> 3        unclear   476087509     <NA>     <NA>      <NA>          <NA>     <NA>
#> 4  acetylcholine   769950427     <NA>     <NA>      <NA>          <NA>     <NA>
#> 5  acetylcholine   481029228     <NA>        R      <NA>          <NA>     <NA>
#> 6           gaba   278587841     <NA>        L      <NA>          <NA>     <NA>
#> 7      serotonin   489236167     <NA>     <NA>      <NA>          <NA>     <NA>
#> 8  acetylcholine   183037288     <NA>     <NA>      <NA>          <NA>     <NA>
#> 9           gaba   199021086     <NA>     <NA>      <NA>          <NA>     <NA>
#> 10 acetylcholine  1181128023     <NA>     <NA>      <NA>          <NA>     <NA>
#> 11 acetylcholine   204307113     <NA>     <NA>      <NA>          <NA>     <NA>
#> 12       unclear   164205396     <NA>     <NA>      <NA>          <NA>     <NA>
#> 13          gaba   833342245     <NA>     <NA>      <NA>          <NA>     <NA>
#> 14          gaba   185592243     <NA>     <NA>      <NA>          <NA>     <NA>
#> 15 acetylcholine   217610342     <NA>     <NA>      <NA>          <NA>     <NA>
#> 16          gaba   172460933     <NA>     <NA>      <NA>          <NA>     <NA>
#> 17       unclear   244236620     <NA>     <NA>      <NA>          <NA>     <NA>
#> 18 acetylcholine   417909777     <NA>     <NA>      <NA>          <NA>     <NA>
#> 19 acetylcholine   291166682     <NA>     <NA>      <NA>          <NA>     <NA>
#> 20 acetylcholine   346648352     <NA>     <NA>      <NA>          <NA>     <NA>
#> 21          gaba   118186516     <NA>     <NA>      <NA>          <NA>     <NA>
#> 22          gaba   139035186     <NA>     <NA>      <NA>          <NA>     <NA>
#> 23 acetylcholine   107081588     <NA>     <NA>      <NA>          <NA>     <NA>
#> 24 acetylcholine   130934995     <NA>     <NA>      <NA>          <NA>     <NA>
#> 25 acetylcholine   127874261     <NA>     <NA>      <NA>          <NA>     <NA>
#> 26 acetylcholine   148541348     <NA>     <NA>      <NA>          <NA>     <NA>
#> 27 acetylcholine   393057814     <NA>     <NA>      <NA>          <NA>     <NA>
#> 28       unclear   142895359     <NA>     <NA>      <NA>          <NA>     <NA>
#> 29       unclear   115258843     <NA>     <NA>      <NA>          <NA>     <NA>
#> 30       unclear 15774680322     <NA>     <NA>      <NA>          <NA>     <NA>
#> 31 acetylcholine   291580020     <NA>     <NA>      <NA>          <NA>     <NA>
#> 32       unclear   323241784     <NA>        R      <NA>          <NA>     <NA>
#> 33 acetylcholine   415371655     <NA>     <NA>      <NA>          <NA>     <NA>
#> 34          gaba   118445955     <NA>     <NA>      <NA>          <NA>     <NA>
#> 35 acetylcholine   389656375     <NA>     <NA>      <NA>          <NA>     <NA>
#> 36 acetylcholine   309045905     <NA>     <NA>      <NA>          <NA>     <NA>
#> 37          gaba   135641183     <NA>     <NA>      <NA>          <NA>     <NA>
#> 38 acetylcholine    66258699     <NA>     <NA>      <NA>          <NA>     <NA>
#> 39          gaba   115459022     <NA>     <NA>      <NA>          <NA>     <NA>
#> 40 acetylcholine   141835905     <NA>     <NA>      <NA>          <NA>     <NA>
#> 41 acetylcholine   229596508     <NA>     <NA>      <NA>          <NA>     <NA>
#> 42       unclear   459673267     <NA>        R      <NA>          <NA>     <NA>
#> 43       unclear   169169631     <NA>     <NA>      <NA>          <NA>     <NA>
#> 44          gaba   230638948     <NA>     <NA>      <NA>          <NA>     <NA>
#> 45          gaba   567492443     <NA>     <NA>      <NA>          <NA>     <NA>
#> 46 acetylcholine    29944601     <NA>     <NA>      <NA>          <NA>     <NA>
#> 47       unclear    93097711     <NA>     <NA>      <NA>          <NA>     <NA>
#> 48       unclear    88932916     <NA>     <NA>      <NA>          <NA>     <NA>
#> 49 acetylcholine   176044266     <NA>     <NA>      <NA>          <NA>     <NA>
#> 50       unclear   129131037     <NA>     <NA>      <NA>          <NA>     <NA>
#>           somaLocation group mancGroup mancType trumanHlVncOnly birthtime
#> 1                         NA        NA     <NA>            <NA>      <NA>
#> 2                         NA        NA     <NA>            <NA>      <NA>
#> 3                         NA        NA     <NA>            <NA>      <NA>
#> 4                         NA        NA     <NA>            <NA>      <NA>
#> 5  44983, 14505, 34017    NA        NA     <NA>            <NA>      <NA>
#> 6                         NA        NA     <NA>            <NA>      <NA>
#> 7  69682, 13204, 19654    NA        NA     <NA>            <NA>      <NA>
#> 8                         NA        NA     <NA>            <NA>      <NA>
#> 9                         NA        NA     <NA>            <NA>      <NA>
#> 10                        NA        NA     <NA>            <NA>      <NA>
#> 11                        NA        NA     <NA>            <NA>      <NA>
#> 12                        NA        NA     <NA>            <NA>      <NA>
#> 13                        NA        NA     <NA>            <NA>      <NA>
#> 14                        NA        NA     <NA>            <NA>      <NA>
#> 15                        NA        NA     <NA>            <NA>      <NA>
#> 16                        NA        NA     <NA>            <NA>      <NA>
#> 17                        NA        NA     <NA>            <NA>      <NA>
#> 18                        NA        NA     <NA>            <NA>      <NA>
#> 19                        NA        NA     <NA>            <NA>      <NA>
#> 20                        NA        NA     <NA>            <NA>      <NA>
#> 21                        NA        NA     <NA>            <NA>      <NA>
#> 22                        NA        NA     <NA>            <NA>      <NA>
#> 23                        NA        NA     <NA>            <NA>      <NA>
#> 24                        NA        NA     <NA>            <NA>      <NA>
#> 25                        NA        NA     <NA>            <NA>      <NA>
#> 26                        NA        NA     <NA>            <NA>      <NA>
#> 27                        NA        NA     <NA>            <NA>      <NA>
#> 28                        NA        NA     <NA>            <NA>      <NA>
#> 29                        NA        NA     <NA>            <NA>      <NA>
#> 30                        NA        NA     <NA>            <NA>      <NA>
#> 31                        NA        NA     <NA>            <NA>      <NA>
#> 32 31887, 20287, 16543    NA        NA     <NA>            <NA>      <NA>
#> 33                        NA        NA     <NA>            <NA>      <NA>
#> 34                        NA        NA     <NA>            <NA>      <NA>
#> 35                        NA        NA     <NA>            <NA>      <NA>
#> 36                        NA        NA     <NA>            <NA>      <NA>
#> 37                        NA        NA     <NA>            <NA>      <NA>
#> 38                        NA        NA     <NA>            <NA>      <NA>
#> 39                        NA        NA     <NA>            <NA>      <NA>
#> 40                        NA        NA     <NA>            <NA>      <NA>
#> 41                        NA        NA     <NA>            <NA>      <NA>
#> 42 45628, 10123, 14545    NA        NA     <NA>            <NA>      <NA>
#> 43                        NA        NA     <NA>            <NA>      <NA>
#> 44                        NA        NA     <NA>            <NA>      <NA>
#> 45                        NA        NA     <NA>            <NA>      <NA>
#> 46                        NA        NA     <NA>            <NA>      <NA>
#> 47                        NA        NA     <NA>            <NA>      <NA>
#> 48                        NA        NA     <NA>            <NA>      <NA>
#> 49                        NA        NA     <NA>            <NA>      <NA>
#> 50                        NA        NA     <NA>            <NA>      <NA>
#>    mancBodyid mcnsSerial mancSerial synonyms fruDsx exitNerve tosomaLocation
#> 1          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 2          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 3          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 4          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 5          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 6          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 7          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 8          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 9          NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 10         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 11     153319         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 12         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 13      11325         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 14         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 15         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 16         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 17         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 18         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 19         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 20         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 21         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 22         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 23         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 24         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 25         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 26         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 27         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 28         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 29         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 30         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 31         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 32         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 33         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 34         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 35         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 36         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 37         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 38         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 39         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 40         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 41      16226         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 42         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 43      28494         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 44      29891         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 45         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 46         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 47         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 48         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 49         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#> 50         NA         NA         NA     <NA>   <NA>      <NA>           <NA>
#>    receptorType matchingNotes hemibrainType anteriorCervicalRootPosition
#> 1          <NA>          <NA>          <NA>                         <NA>
#> 2          <NA>          <NA>          <NA>                         <NA>
#> 3          <NA>          <NA>          <NA>                         <NA>
#> 4          <NA>          <NA>          <NA>                         <NA>
#> 5          <NA>          <NA>          <NA>                         <NA>
#> 6          <NA>          <NA>          <NA>                         <NA>
#> 7          <NA>          <NA>          <NA>                         <NA>
#> 8          <NA>          <NA>          <NA>                         <NA>
#> 9          <NA>          <NA>          <NA>                         <NA>
#> 10         <NA>          <NA>          <NA>                         <NA>
#> 11         <NA>          <NA>          <NA>                         <NA>
#> 12         <NA>          <NA>          <NA>                         <NA>
#> 13         <NA>          <NA>          <NA>                         <NA>
#> 14         <NA>          <NA>          <NA>                         <NA>
#> 15         <NA>          <NA>          <NA>                         <NA>
#> 16         <NA>          <NA>          <NA>                         <NA>
#> 17         <NA>          <NA>          <NA>                         <NA>
#> 18         <NA>          <NA>          <NA>                         <NA>
#> 19         <NA>          <NA>          <NA>                         <NA>
#> 20         <NA>          <NA>          <NA>                         <NA>
#> 21         <NA>          <NA>          <NA>                         <NA>
#> 22         <NA>          <NA>          <NA>                         <NA>
#> 23         <NA>          <NA>          <NA>                         <NA>
#> 24         <NA>          <NA>          <NA>                         <NA>
#> 25         <NA>          <NA>          <NA>                         <NA>
#> 26         <NA>          <NA>          <NA>                         <NA>
#> 27         <NA>          <NA>          <NA>                         <NA>
#> 28         <NA>          <NA>          <NA>                         <NA>
#> 29         <NA>          <NA>          <NA>                         <NA>
#> 30         <NA>          <NA>          <NA>                         <NA>
#> 31         <NA>          <NA>          <NA>                         <NA>
#> 32         <NA>          <NA>          <NA>                         <NA>
#> 33         <NA>          <NA>          <NA>                         <NA>
#> 34         <NA>          <NA>          <NA>                         <NA>
#> 35         <NA>          <NA>          <NA>                         <NA>
#> 36         <NA>          <NA>          <NA>                         <NA>
#> 37         <NA>          <NA>          <NA>                         <NA>
#> 38         <NA>          <NA>          <NA>                         <NA>
#> 39         <NA>          <NA>          <NA>                         <NA>
#> 40         <NA>          <NA>          <NA>                         <NA>
#> 41         <NA>          <NA>          <NA>                         <NA>
#> 42         <NA>          <NA>          <NA>                         <NA>
#> 43         <NA>          <NA>          <NA>                         <NA>
#> 44         <NA>          <NA>          <NA>                         <NA>
#> 45         <NA>          <NA>          <NA>                         <NA>
#> 46         <NA>          <NA>          <NA>                         <NA>
#> 47         <NA>          <NA>          <NA>                         <NA>
#> 48         <NA>          <NA>          <NA>                         <NA>
#> 49         <NA>          <NA>          <NA>                         <NA>
#> 50         <NA>          <NA>          <NA>                         <NA>
#>    dimorphism flywireId posteriorCervicalRootPosition serialMotif
#> 1        <NA>      <NA>                          <NA>        <NA>
#> 2        <NA>      <NA>                          <NA>        <NA>
#> 3        <NA>      <NA>                          <NA>        <NA>
#> 4        <NA>      <NA>                          <NA>        <NA>
#> 5        <NA>      <NA>                          <NA>        <NA>
#> 6        <NA>      <NA>                          <NA>        <NA>
#> 7        <NA>      <NA>                          <NA>        <NA>
#> 8        <NA>      <NA>                          <NA>        <NA>
#> 9        <NA>      <NA>                          <NA>        <NA>
#> 10       <NA>      <NA>                          <NA>        <NA>
#> 11       <NA>      <NA>                          <NA>        <NA>
#> 12       <NA>      <NA>                          <NA>        <NA>
#> 13       <NA>      <NA>                          <NA>        <NA>
#> 14       <NA>      <NA>                          <NA>        <NA>
#> 15       <NA>      <NA>                          <NA>        <NA>
#> 16       <NA>      <NA>                          <NA>        <NA>
#> 17       <NA>      <NA>                          <NA>        <NA>
#> 18       <NA>      <NA>                          <NA>        <NA>
#> 19       <NA>      <NA>                          <NA>        <NA>
#> 20       <NA>      <NA>                          <NA>        <NA>
#> 21       <NA>      <NA>                          <NA>        <NA>
#> 22       <NA>      <NA>                          <NA>        <NA>
#> 23       <NA>      <NA>                          <NA>        <NA>
#> 24       <NA>      <NA>                          <NA>        <NA>
#> 25       <NA>      <NA>                          <NA>        <NA>
#> 26       <NA>      <NA>                          <NA>        <NA>
#> 27       <NA>      <NA>                          <NA>        <NA>
#> 28       <NA>      <NA>                          <NA>        <NA>
#> 29       <NA>      <NA>                          <NA>        <NA>
#> 30       <NA>      <NA>                          <NA>        <NA>
#> 31       <NA>      <NA>                          <NA>        <NA>
#> 32       <NA>      <NA>                          <NA>        <NA>
#> 33       <NA>      <NA>                          <NA>        <NA>
#> 34       <NA>      <NA>                          <NA>        <NA>
#> 35       <NA>      <NA>                          <NA>        <NA>
#> 36       <NA>      <NA>                          <NA>        <NA>
#> 37       <NA>      <NA>                          <NA>        <NA>
#> 38       <NA>      <NA>                          <NA>        <NA>
#> 39       <NA>      <NA>                          <NA>        <NA>
#> 40       <NA>      <NA>                          <NA>        <NA>
#> 41       <NA>      <NA>                          <NA>        <NA>
#> 42       <NA>      <NA>                          <NA>        <NA>
#> 43       <NA>      <NA>                          <NA>        <NA>
#> 44       <NA>      <NA>                          <NA>        <NA>
#> 45       <NA>      <NA>                          <NA>        <NA>
#> 46       <NA>      <NA>                          <NA>        <NA>
#> 47       <NA>      <NA>                          <NA>        <NA>
#> 48       <NA>      <NA>                          <NA>        <NA>
#> 49       <NA>      <NA>                          <NA>        <NA>
#> 50       <NA>      <NA>                          <NA>        <NA>
#>    assignedOlHex1 assignedOlHex2 hemibrainBodyId location locationType
#> 1              NA             NA            <NA>     <NA>         <NA>
#> 2              NA             NA            <NA>     <NA>         <NA>
#> 3              NA             NA            <NA>     <NA>         <NA>
#> 4              NA             NA            <NA>     <NA>         <NA>
#> 5              NA             NA            <NA>     <NA>         <NA>
#> 6              NA             NA            <NA>     <NA>         <NA>
#> 7              NA             NA            <NA>     <NA>         <NA>
#> 8              NA             NA            <NA>     <NA>         <NA>
#> 9              NA             NA            <NA>     <NA>         <NA>
#> 10             NA             NA            <NA>     <NA>         <NA>
#> 11             NA             NA            <NA>     <NA>         <NA>
#> 12             NA             NA            <NA>     <NA>         <NA>
#> 13             NA             NA            <NA>     <NA>         <NA>
#> 14             NA             NA            <NA>     <NA>         <NA>
#> 15             NA             NA            <NA>     <NA>         <NA>
#> 16             NA             NA            <NA>     <NA>         <NA>
#> 17             NA             NA            <NA>     <NA>         <NA>
#> 18             NA             NA            <NA>     <NA>         <NA>
#> 19             NA             NA            <NA>     <NA>         <NA>
#> 20             NA             NA            <NA>     <NA>         <NA>
#> 21             NA             NA            <NA>     <NA>         <NA>
#> 22             NA             NA            <NA>     <NA>         <NA>
#> 23             NA             NA            <NA>     <NA>         <NA>
#> 24             NA             NA            <NA>     <NA>         <NA>
#> 25             NA             NA            <NA>     <NA>         <NA>
#> 26             NA             NA            <NA>     <NA>         <NA>
#> 27             NA             NA            <NA>     <NA>         <NA>
#> 28             NA             NA            <NA>     <NA>         <NA>
#> 29             NA             NA            <NA>     <NA>         <NA>
#> 30             NA             NA            <NA>     <NA>         <NA>
#> 31             NA             NA            <NA>     <NA>         <NA>
#> 32             NA             NA            <NA>     <NA>         <NA>
#> 33             NA             NA            <NA>     <NA>         <NA>
#> 34             NA             NA            <NA>     <NA>         <NA>
#> 35             NA             NA            <NA>     <NA>         <NA>
#> 36             NA             NA            <NA>     <NA>         <NA>
#> 37             NA             NA            <NA>     <NA>         <NA>
#> 38             NA             NA            <NA>     <NA>         <NA>
#> 39             NA             NA            <NA>     <NA>         <NA>
#> 40             NA             NA            <NA>     <NA>         <NA>
#> 41             NA             NA            <NA>     <NA>         <NA>
#> 42             NA             NA            <NA>     <NA>         <NA>
#> 43             NA             NA            <NA>     <NA>         <NA>
#> 44             NA             NA            <NA>     <NA>         <NA>
#> 45             NA             NA            <NA>     <NA>         <NA>
#> 46             NA             NA            <NA>     <NA>         <NA>
#> 47             NA             NA            <NA>     <NA>         <NA>
#> 48             NA             NA            <NA>     <NA>         <NA>
#> 49             NA             NA            <NA>     <NA>         <NA>
#> 50             NA             NA            <NA>     <NA>         <NA>
#>    cellBodyFiber  soma
#> 1           <NA> FALSE
#> 2           <NA> FALSE
#> 3           <NA> FALSE
#> 4           <NA> FALSE
#> 5           <NA>  TRUE
#> 6           <NA> FALSE
#> 7           <NA>  TRUE
#> 8           <NA> FALSE
#> 9           <NA> FALSE
#> 10          <NA> FALSE
#> 11          <NA> FALSE
#> 12          <NA> FALSE
#> 13          <NA> FALSE
#> 14          <NA> FALSE
#> 15          <NA> FALSE
#> 16          <NA> FALSE
#> 17          <NA> FALSE
#> 18          <NA> FALSE
#> 19          <NA> FALSE
#> 20          <NA> FALSE
#> 21          <NA> FALSE
#> 22          <NA> FALSE
#> 23          <NA> FALSE
#> 24          <NA> FALSE
#> 25          <NA> FALSE
#> 26          <NA> FALSE
#> 27          <NA> FALSE
#> 28          <NA> FALSE
#> 29          <NA> FALSE
#> 30          <NA> FALSE
#> 31          <NA> FALSE
#> 32          <NA>  TRUE
#> 33          <NA> FALSE
#> 34          <NA> FALSE
#> 35          <NA> FALSE
#> 36          <NA> FALSE
#> 37          <NA> FALSE
#> 38          <NA> FALSE
#> 39          <NA> FALSE
#> 40          <NA> FALSE
#> 41          <NA> FALSE
#> 42          <NA>  TRUE
#> 43          <NA> FALSE
#> 44          <NA> FALSE
#> 45          <NA> FALSE
#> 46          <NA> FALSE
#> 47          <NA> FALSE
#> 48          <NA> FALSE
#> 49          <NA> FALSE
#> 50          <NA> FALSE
# }
library(dplyr)
# Which neurons don't have a superclass, but possibly should
mnm.nsc=mcns_neuprint_meta("where:NOT exists(n.superclass)")
mnm.nsc |> count(statusLabel)
#>              statusLabel    n
#> 1              0.5assign   48
#> 2                 Anchor  588
#> 3                   Glia  393
#> 4          Hard to trace    1
#> 5                 Leaves  411
#> 6                 Orphan 3499
#> 7        Orphan hotknife 1017
#> 8        Orphan-artifact 1931
#> 9           Out of scope 1591
#> 10            PRT Orphan   23
#> 11 Prelim Roughly traced   25
#> 12        Primary Anchor  432
#> 13      RT Hard to trace    1
#> 14        Roughly traced    2
#> 15        Sensory Anchor  193
#> 16           Soma Anchor   28
#> 17           Unimportant  160
#> 18                  <NA>  469

# neurons that are RT or PRT should probably have a superclass
mnm.nscprt=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.statusLabel CONTAINS 'Roughly'")
mnm.nscprt |> count()
#>    n
#> 1 27