Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

# \donttest{
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')
# }