Fetch neuprint metadata for malecns neurons

Fetch neuprint metadata for malecns neurons

Usage

mcns_neuprint_meta(
  ids = NULL,
  conn = mcns_neuprint(),
  roiInfo = FALSE,
  simplify.xyz = TRUE,
  ...
)

Arguments

ids

body ids. When missing all bodies known to DVID are returned.

conn

Optional, a neuprint_connection object, which also specifies the neuPrint server. Defaults to manc_neuprint() to ensure that query is against the VNC dataset.

roiInfo

whether to include the roiInfo field detailing synapse numbers in different locations. This is omitted by default as it is returned as a character vector of unprocessed JSON.

simplify.xyz

Whether to simplify columns containing XYZ locations to a simple "x,y,z" format rather than a longer form referencing a schema at spatialreference.org. Defaults to TRUE.

...

Additional arguments passed to neuprint_get_meta

Value

A data.frame with one row for each (unique) input id and NAs for all columns except bodyid when neuprint holds no metadata.

Details

in contrast to malevnc::manc_neuprint_meta we leave bodyids as numeric (doubles) since flyem now guarantee them to be less than 2^53 i.e. within the range in which doubles can exactly represent numeric ids.

See also

Examples

# \donttest{
library(dplyr)
# fetch metatada for all bodies in neuprint
mnm=mcns_neuprint_meta()
# fetch metadata for all bodies with a somaLocation
mnm.soma=mcns_neuprint_meta("where:exists(n.somaLocation)")

# type or instance present
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')

# neurons without a superclass but quite a few synapses
mnm.nc=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.synweight>2000")
mnm.nc |> arrange(desc(synweight))
#>       bodyid post  pre downstream upstream synweight           statusLabel
#> 1  765790195  567 1426      12023      567     12590 Prelim Roughly traced
#> 2      75305 5340  643       4793     5340     10133 Prelim Roughly traced
#> 3     331473  615 1176       7695      615      8310            PRT Orphan
#> 4     533707  177  884       7343      177      7520       Orphan-artifact
#> 5     944067 2907  592       3981     2907      6888        Primary Anchor
#> 6     555399 2504 1425       4175     2504      6679           Unimportant
#> 7     581108 2918  389       2804     2918      5722 Prelim Roughly traced
#> 8     941628 2045  442       3260     2045      5305 Prelim Roughly traced
#> 9     519161 2490  326       2724     2490      5214                Orphan
#> 10    944585  653  562       4534      653      5187                Orphan
#> 11    174754 3021  294       2054     3021      5075 Prelim Roughly traced
#> 12    305537  329  543       4733      329      5062 Prelim Roughly traced
#> 13    127314 1414  335       2786     1414      4200 Prelim Roughly traced
#> 14    555297  274  607       3919      274      4193                Orphan
#> 15    918780  182  490       3760      182      3942 Prelim Roughly traced
#> 16     88667 1925  222       1702     1925      3627 Prelim Roughly traced
#> 17 236599764   98  531       3509       98      3607                Orphan
#> 18    232828 1312  351       2162     1312      3474 Prelim Roughly traced
#> 19    953607  129  429       3278      129      3407                  <NA>
#> 20 192342436  228  371       3166      228      3394 Prelim Roughly traced
#> 21    926152 1182  321       2184     1182      3366       Orphan-artifact
#> 22     73540  332  322       3016      332      3348 Prelim Roughly traced
#> 23    423550  389  353       2937      389      3326 Prelim Roughly traced
#> 24    942401  403  465       2865      403      3268       Orphan-artifact
#> 25     66613 1678  247       1536     1678      3214 Prelim Roughly traced
#> 26     92244 1763  222       1426     1763      3189                Orphan
#> 27    943522  796  304       2268      796      3064 Prelim Roughly traced
#> 28    946767  230  446       2750      230      2980                  <NA>
#> 29    946890  283  324       2546      283      2829                  <NA>
#> 30    559672 1407  248       1389     1407      2796                Orphan
#> 31    104626   53  319       2733       53      2786 Prelim Roughly traced
#> 32    110922   67  267       2641       67      2708 Prelim Roughly traced
#> 33     85936 1183  189       1520     1183      2703 Prelim Roughly traced
#> 34    114076  157  301       2535      157      2692                Orphan
#> 35    530376  446  244       2181      446      2627                Orphan
#> 36    107987 1478  133       1142     1478      2620 Prelim Roughly traced
#> 37    934797 1566  264       1048     1566      2614            PRT Orphan
#> 38    943251  426  303       2126      426      2552 Prelim Roughly traced
#> 39    118423  105  257       2426      105      2531 Prelim Roughly traced
#> 40 516174168  159  308       2369      159      2528            PRT Orphan
#> 41    928870  399  241       2080      399      2479        Primary Anchor
#> 42    112430  746  217       1663      746      2409 Prelim Roughly traced
#> 43     10621  679  557       1726      679      2405           Unimportant
#> 44    210839   79  269       2311       79      2390            PRT Orphan
#> 45    147605 1180  183       1208     1180      2388                Orphan
#> 46    555267   65  292       2314       65      2379 Prelim Roughly traced
#> 47    908376   26  235       2318       26      2344       Orphan-artifact
#> 48    941745  187  201       2128      187      2315       Orphan-artifact
#> 49    945946  347  291       1948      347      2295                Anchor
#> 50    953574 1282  227        953     1282      2235                  <NA>
#> 51    210579 1432  179        788     1432      2220 Prelim Roughly traced
#> 52    580488  416  202       1755      416      2171        Primary Anchor
#> 53    953169  101  247       2055      101      2156                  <NA>
#> 54    819159   63  348       2078       63      2141       Orphan-artifact
#> 55    484601  144  281       1904      144      2048 Prelim Roughly traced
#> 56    147742  257  204       1790      257      2047                Orphan
#> 57    946550  975  128       1068      975      2043                Orphan
#> 58    953507  672  189       1361      672      2033       Orphan-artifact
#> 59    141338 1403  103        612     1403      2015       Orphan-artifact
#> 60    953023  344  217       1665      344      2009        Primary Anchor
#>    superclass class entryNerve rootSide subclass      status      voxels
#> 1        <NA>  <NA>       <NA>     <NA>     <NA>      Traced   627946176
#> 2        <NA>  <NA>       <NA>     <NA>     <NA>      Traced   566721866
#> 3        <NA>  <NA>       <NA>     <NA>     <NA>      Traced   365414996
#> 4        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   404897796
#> 5        <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   478617308
#> 6        <NA>  <NA>       <NA>     <NA>     <NA> Unimportant 15774680322
#> 7        <NA>  <NA>       <NA>     <NA>     <NA>      Traced   546792498
#> 8        <NA>  <NA>       <NA>     <NA>     <NA>      Traced   616595182
#> 9        <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   275039336
#> 10       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   240560432
#> 11       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   259659259
#> 12       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   207003096
#> 13       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   418839704
#> 14       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   183037288
#> 15       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   221109490
#> 16       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   197698192
#> 17       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   161511491
#> 18       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   418846190
#> 19       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   195663702
#> 20       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   121644514
#> 21       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   833342245
#> 22       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   164205396
#> 23       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   223780462
#> 24       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   244236620
#> 25       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   184607589
#> 26       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   172460933
#> 27       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   350774957
#> 28       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   417909777
#> 29       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   419414444
#> 30       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   139035186
#> 31       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   118186516
#> 32       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   130934995
#> 33       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   159910308
#> 34       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   127874261
#> 35       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   152621707
#> 36       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   127911160
#> 37       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   136391713
#> 38       <NA>  <NA>       <NA>     <NA>     <NA>      Traced    83487356
#> 39       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   115227387
#> 40       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   147509581
#> 41       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   143101585
#> 42       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   320117047
#> 43       <NA>  <NA>       <NA>     <NA>     <NA> Unimportant  6475367429
#> 44       <NA>  <NA>       <NA>     <NA>     <NA>      Traced    66258699
#> 45       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   118445955
#> 46       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   135641183
#> 47       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   115459022
#> 48       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   116788659
#> 49       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   379830308
#> 50       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>   564805869
#> 51       <NA>  <NA>       <NA>     <NA>     <NA>      Traced   100881025
#> 52       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   260194496
#> 53       <NA>  <NA>       <NA>     <NA>     <NA>        <NA>    71780552
#> 54       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   316825823
#> 55       <NA>  <NA>       <NA>     <NA>     <NA>      Traced    99210529
#> 56       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan    88932916
#> 57       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   129131037
#> 58       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan   212778138
#> 59       <NA>  <NA>       <NA>     <NA>     <NA>      Orphan    79808810
#> 60       <NA>  <NA>       <NA>     <NA>     <NA>      Anchor   192701860
#>    somaSide      somaLocation        itoleeHl anteriorCervicalRootPosition
#> 1      <NA>                              <NA>                         <NA>
#> 2      <NA>                              <NA>                         <NA>
#> 3      <NA>                              <NA>                         <NA>
#> 4      <NA>                              <NA>                         <NA>
#> 5      <NA>                              <NA>                         <NA>
#> 6      <NA>                              <NA>                         <NA>
#> 7      <NA> 25284,24392,20282    VLPl2_medial                         <NA>
#> 8      <NA> 24286,25890,21014    VLPl2_medial                         <NA>
#> 9         L                              <NA>                         <NA>
#> 10     <NA>                              <NA>                         <NA>
#> 11        R                              <NA>                         <NA>
#> 12     <NA>                              <NA>                         <NA>
#> 13        L 59624,26318,37428            <NA>                         <NA>
#> 14     <NA>                              <NA>                         <NA>
#> 15     <NA>                              <NA>                         <NA>
#> 16     <NA>                              <NA>                         <NA>
#> 17     <NA>                              <NA>                         <NA>
#> 18     <NA> 73607,21955,19078 VLPl&d1_lateral                         <NA>
#> 19     <NA>                              <NA>                         <NA>
#> 20     <NA>                              <NA>                         <NA>
#> 21     <NA>                              <NA>                         <NA>
#> 22     <NA>                              <NA>                         <NA>
#> 23     <NA>                              <NA>                         <NA>
#> 24     <NA>                              <NA>                         <NA>
#> 25     <NA>                              <NA>                         <NA>
#> 26     <NA>                              <NA>                         <NA>
#> 27     <NA> 25122,11941,24828            <NA>                         <NA>
#> 28     <NA>                              <NA>                         <NA>
#> 29     <NA>                              <NA>                         <NA>
#> 30     <NA>                              <NA>                         <NA>
#> 31     <NA>                              <NA>                         <NA>
#> 32     <NA>                              <NA>                         <NA>
#> 33     <NA>                              <NA>                         <NA>
#> 34     <NA>                              <NA>                         <NA>
#> 35     <NA>                              <NA>                         <NA>
#> 36        R                              <NA>                         <NA>
#> 37     <NA>                              <NA>                         <NA>
#> 38     <NA>                              <NA>                         <NA>
#> 39     <NA>                              <NA>                         <NA>
#> 40     <NA>                              <NA>                         <NA>
#> 41     <NA>                              <NA>                         <NA>
#> 42        L 71840,27186,18458            <NA>                         <NA>
#> 43     <NA>                              <NA>                         <NA>
#> 44     <NA>                              <NA>                         <NA>
#> 45     <NA>                              <NA>                         <NA>
#> 46     <NA>                              <NA>                         <NA>
#> 47     <NA>                              <NA>                         <NA>
#> 48     <NA>                              <NA>                         <NA>
#> 49     <NA>                              <NA>                         <NA>
#> 50     <NA>                              <NA>                         <NA>
#> 51     <NA>                              <NA>                         <NA>
#> 52     <NA> 74352,28712,20890            <NA>                         <NA>
#> 53     <NA>                              <NA>                         <NA>
#> 54     <NA>                              <NA>                         <NA>
#> 55     <NA>                              <NA>                         <NA>
#> 56     <NA>                              <NA>                         <NA>
#> 57     <NA>                              <NA>                         <NA>
#> 58     <NA>                              <NA>                         <NA>
#> 59     <NA>                              <NA>                         <NA>
#> 60     <NA> 77063,50223,32121            <NA>                         <NA>
#>    synonyms fruDsx somaNeuromere trumanHl birthtime group mancBodyid mancGroup
#> 1      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 2      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 3      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 4      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 5      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 6      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 7      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 8      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 9      <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 10     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 11     <NA>   <NA>          <NA>     <NA>      <NA> 18483         NA        NA
#> 12     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 13     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 14     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 15     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 16     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 17     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 18     <NA>   <NA>          <NA>     <NA>      <NA> 28005         NA        NA
#> 19     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 20     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 21     <NA>   <NA>          <NA>     <NA>      <NA>    NA      11325        NA
#> 22     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 23     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 24     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 25     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 26     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 27     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 28     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 29     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 30     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 31     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 32     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 33     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 34     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 35     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 36     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 37     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 38     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 39     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 40     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 41     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 42     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 43     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 44     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 45     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 46     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 47     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 48     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 49     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 50     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 51     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 52     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 53     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 54     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 55     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 56     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 57     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 58     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 59     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#> 60     <NA>   <NA>          <NA>     <NA>      <NA>    NA         NA        NA
#>    mancType type mancSerial exitNerve matchingNotes mcnsSerial hemibrainType
#> 1      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 2      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 3      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 4      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 5      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 6      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 7      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 8      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 9      <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 10     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 11     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 12     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 13     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 14     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 15     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 16     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 17     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 18     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 19     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 20     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 21     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 22     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 23     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 24     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 25     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 26     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 27     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 28     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 29     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 30     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 31     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 32     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 33     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 34     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 35     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 36     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 37     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 38     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 39     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 40     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 41     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 42     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 43     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 44     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 45     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 46     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 47     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 48     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 49     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 50     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 51     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 52     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 53     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 54     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 55     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 56     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 57     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 58     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 59     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#> 60     <NA> <NA>         NA      <NA>          <NA>         NA          <NA>
#>             name flywireType flywireId posteriorCervicalRootPosition
#> 1           <NA>        <NA>      <NA>                          <NA>
#> 2           <NA>        <NA>      <NA>                          <NA>
#> 3           <NA>        <NA>      <NA>                          <NA>
#> 4           <NA>        <NA>      <NA>                          <NA>
#> 5           <NA>        <NA>      <NA>                          <NA>
#> 6           <NA>        <NA>      <NA>                          <NA>
#> 7           <NA>        <NA>      <NA>                          <NA>
#> 8           <NA>        <NA>      <NA>                          <NA>
#> 9  KC_R_fragment        <NA>      <NA>                          <NA>
#> 10          <NA>        <NA>      <NA>                          <NA>
#> 11 KC_R_fragment        <NA>      <NA>                          <NA>
#> 12          <NA>        <NA>      <NA>                          <NA>
#> 13          <NA>        <NA>      <NA>                          <NA>
#> 14          <NA>        <NA>      <NA>                          <NA>
#> 15          <NA>        <NA>      <NA>                          <NA>
#> 16          <NA>        <NA>      <NA>                          <NA>
#> 17          <NA>        <NA>      <NA>                          <NA>
#> 18          <NA>        <NA>      <NA>                          <NA>
#> 19          <NA>        <NA>      <NA>                          <NA>
#> 20          <NA>        <NA>      <NA>                          <NA>
#> 21          <NA>        <NA>      <NA>                          <NA>
#> 22          <NA>        <NA>      <NA>                          <NA>
#> 23          <NA>        <NA>      <NA>                          <NA>
#> 24          <NA>        <NA>      <NA>                          <NA>
#> 25          <NA>        <NA>      <NA>                          <NA>
#> 26          <NA>        <NA>      <NA>                          <NA>
#> 27          <NA>        <NA>      <NA>                          <NA>
#> 28          <NA>        <NA>      <NA>                          <NA>
#> 29          <NA>        <NA>      <NA>                          <NA>
#> 30          <NA>        <NA>      <NA>                          <NA>
#> 31          <NA>        <NA>      <NA>                          <NA>
#> 32          <NA>        <NA>      <NA>                          <NA>
#> 33          <NA>        <NA>      <NA>                          <NA>
#> 34          <NA>        <NA>      <NA>                          <NA>
#> 35          <NA>        <NA>      <NA>                          <NA>
#> 36 KC_R_fragment        <NA>      <NA>                          <NA>
#> 37          <NA>        <NA>      <NA>                          <NA>
#> 38          <NA>        <NA>      <NA>                          <NA>
#> 39          <NA>        <NA>      <NA>                          <NA>
#> 40          <NA>        <NA>      <NA>                          <NA>
#> 41          <NA>        <NA>      <NA>                          <NA>
#> 42          <NA>        <NA>      <NA>                          <NA>
#> 43          <NA>        <NA>      <NA>                          <NA>
#> 44          <NA>        <NA>      <NA>                          <NA>
#> 45          <NA>        <NA>      <NA>                          <NA>
#> 46          <NA>        <NA>      <NA>                          <NA>
#> 47          <NA>        <NA>      <NA>                          <NA>
#> 48          <NA>        <NA>      <NA>                          <NA>
#> 49          <NA>        <NA>      <NA>                          <NA>
#> 50          <NA>        <NA>      <NA>                          <NA>
#> 51 MBON_fragment        <NA>      <NA>                          <NA>
#> 52          <NA>        <NA>      <NA>                          <NA>
#> 53          <NA>        <NA>      <NA>                          <NA>
#> 54          <NA>        <NA>      <NA>                          <NA>
#> 55          <NA>        <NA>      <NA>                          <NA>
#> 56          <NA>        <NA>      <NA>                          <NA>
#> 57          <NA>        <NA>      <NA>                          <NA>
#> 58          <NA>        <NA>      <NA>                          <NA>
#> 59          <NA>        <NA>      <NA>                          <NA>
#> 60          <NA>        <NA>      <NA>                          <NA>
#>    serialMotif dimorphism hemibrainBodyId assignedOlHex1 assignedOlHex2
#> 1         <NA>       <NA>            <NA>             NA             NA
#> 2         <NA>       <NA>            <NA>             NA             NA
#> 3         <NA>       <NA>            <NA>             NA             NA
#> 4         <NA>       <NA>            <NA>             NA             NA
#> 5         <NA>       <NA>            <NA>             NA             NA
#> 6         <NA>       <NA>            <NA>             NA             NA
#> 7         <NA>       <NA>            <NA>             NA             NA
#> 8         <NA>       <NA>            <NA>             NA             NA
#> 9         <NA>       <NA>            <NA>             NA             NA
#> 10        <NA>       <NA>            <NA>             NA             NA
#> 11        <NA>       <NA>            <NA>             NA             NA
#> 12        <NA>       <NA>            <NA>             NA             NA
#> 13        <NA>       <NA>            <NA>             NA             NA
#> 14        <NA>       <NA>            <NA>             NA             NA
#> 15        <NA>       <NA>            <NA>             NA             NA
#> 16        <NA>       <NA>            <NA>             NA             NA
#> 17        <NA>       <NA>            <NA>             NA             NA
#> 18        <NA>       <NA>            <NA>             NA             NA
#> 19        <NA>       <NA>            <NA>             NA             NA
#> 20        <NA>       <NA>            <NA>             NA             NA
#> 21        <NA>       <NA>            <NA>             NA             NA
#> 22        <NA>       <NA>            <NA>             NA             NA
#> 23        <NA>       <NA>            <NA>             NA             NA
#> 24        <NA>       <NA>            <NA>             NA             NA
#> 25        <NA>       <NA>            <NA>             NA             NA
#> 26        <NA>       <NA>            <NA>             NA             NA
#> 27        <NA>       <NA>            <NA>             NA             NA
#> 28        <NA>       <NA>            <NA>             NA             NA
#> 29        <NA>       <NA>            <NA>             NA             NA
#> 30        <NA>       <NA>            <NA>             NA             NA
#> 31        <NA>       <NA>            <NA>             NA             NA
#> 32        <NA>       <NA>            <NA>             NA             NA
#> 33        <NA>       <NA>            <NA>             NA             NA
#> 34        <NA>       <NA>            <NA>             NA             NA
#> 35        <NA>       <NA>            <NA>             NA             NA
#> 36        <NA>       <NA>            <NA>             NA             NA
#> 37        <NA>       <NA>            <NA>             NA             NA
#> 38        <NA>       <NA>            <NA>             NA             NA
#> 39        <NA>       <NA>            <NA>             NA             NA
#> 40        <NA>       <NA>            <NA>             NA             NA
#> 41        <NA>       <NA>            <NA>             NA             NA
#> 42        <NA>       <NA>            <NA>             NA             NA
#> 43        <NA>       <NA>            <NA>             NA             NA
#> 44        <NA>       <NA>            <NA>             NA             NA
#> 45        <NA>       <NA>            <NA>             NA             NA
#> 46        <NA>       <NA>            <NA>             NA             NA
#> 47        <NA>       <NA>            <NA>             NA             NA
#> 48        <NA>       <NA>            <NA>             NA             NA
#> 49        <NA>       <NA>            <NA>             NA             NA
#> 50        <NA>       <NA>            <NA>             NA             NA
#> 51        <NA>       <NA>            <NA>             NA             NA
#> 52        <NA>       <NA>            <NA>             NA             NA
#> 53        <NA>       <NA>            <NA>             NA             NA
#> 54        <NA>       <NA>            <NA>             NA             NA
#> 55        <NA>       <NA>            <NA>             NA             NA
#> 56        <NA>       <NA>            <NA>             NA             NA
#> 57        <NA>       <NA>            <NA>             NA             NA
#> 58        <NA>       <NA>            <NA>             NA             NA
#> 59        <NA>       <NA>            <NA>             NA             NA
#> 60        <NA>       <NA>            <NA>             NA             NA
#>    locationType cellBodyFiber  soma
#> 1          <NA>          <NA> FALSE
#> 2          <NA>          <NA> FALSE
#> 3          <NA>          <NA> FALSE
#> 4          <NA>          <NA> FALSE
#> 5          <NA>          <NA> FALSE
#> 6          <NA>          <NA> FALSE
#> 7          <NA>          <NA>  TRUE
#> 8          <NA>          <NA>  TRUE
#> 9          <NA>          <NA> FALSE
#> 10         <NA>          <NA> FALSE
#> 11         <NA>          <NA> FALSE
#> 12         <NA>          <NA> FALSE
#> 13         <NA>          <NA>  TRUE
#> 14         <NA>          <NA> FALSE
#> 15         <NA>          <NA> FALSE
#> 16         <NA>          <NA> FALSE
#> 17         <NA>          <NA> FALSE
#> 18         <NA>          <NA>  TRUE
#> 19         <NA>          <NA> FALSE
#> 20         <NA>          <NA> FALSE
#> 21         <NA>          <NA> FALSE
#> 22         <NA>          <NA> FALSE
#> 23         <NA>          <NA> FALSE
#> 24         <NA>          <NA> FALSE
#> 25         <NA>          <NA> FALSE
#> 26         <NA>          <NA> FALSE
#> 27         <NA>          <NA>  TRUE
#> 28         <NA>          <NA> FALSE
#> 29         <NA>          <NA> FALSE
#> 30         <NA>          <NA> FALSE
#> 31         <NA>          <NA> FALSE
#> 32         <NA>          <NA> FALSE
#> 33         <NA>          <NA> FALSE
#> 34         <NA>          <NA> FALSE
#> 35         <NA>          <NA> FALSE
#> 36         <NA>          <NA> FALSE
#> 37         <NA>          <NA> FALSE
#> 38         <NA>          <NA> FALSE
#> 39         <NA>          <NA> FALSE
#> 40         <NA>          <NA> FALSE
#> 41         <NA>          <NA> FALSE
#> 42         <NA>          <NA>  TRUE
#> 43         <NA>          <NA> FALSE
#> 44         <NA>          <NA> FALSE
#> 45         <NA>          <NA> FALSE
#> 46         <NA>          <NA> FALSE
#> 47         <NA>          <NA> FALSE
#> 48         <NA>          <NA> FALSE
#> 49         <NA>          <NA> FALSE
#> 50         <NA>          <NA> FALSE
#> 51         <NA>          <NA> FALSE
#> 52         <NA>          <NA>  TRUE
#> 53         <NA>          <NA> FALSE
#> 54         <NA>          <NA> FALSE
#> 55         <NA>          <NA> FALSE
#> 56         <NA>          <NA> FALSE
#> 57         <NA>          <NA> FALSE
#> 58         <NA>          <NA> FALSE
#> 59         <NA>          <NA> FALSE
#> 60         <NA>          <NA>  TRUE

# }

# Which neurons don't have a superclass, but possibly should
mnm.nsc=mcns_neuprint_meta("where:NOT exists(n.superclass)")
mnm.nsc |> count(statusLabel)
#>              statusLabel    n
#> 1              0.5assign   51
#> 2                 Anchor  215
#> 3                   Glia  508
#> 4          Hard to trace    1
#> 5                 Leaves  223
#> 6                 Orphan 3715
#> 7        Orphan hotknife  662
#> 8        Orphan-artifact 1699
#> 9           Out of scope 1300
#> 10            PRT Orphan   42
#> 11 Prelim Roughly traced  373
#> 12        Primary Anchor  126
#> 13        Roughly traced    3
#> 14        Sensory Anchor  170
#> 15           Unimportant  614
#> 16                  <NA>  323

# neurons that are RT or PRT should probably have a superclass
mnm.nscprt=mcns_neuprint_meta("where:NOT exists(n.superclass) AND n.statusLabel CONTAINS 'Roughly'")
mnm.nscprt |> count()
#>     n
#> 1 376