Predict the group of neurons using instance or type information

Predict the group of neurons using instance or type information

Usage

mcns_predict_group(
  ids,
  method = c("auto", "group", "instance", "type", "all"),
  badtypes = c(NA, "", "Lamina_R1-R6", "Descending", "KC", "ER", "LC", "PB",
    "Ascending Interneuron", "Delta", "P1_L candidate", "LT", "MeMe", "PFGs", "Mi", "VT",
    "ML", "EL", "FB", "Dm", "DNp", "FC", "OL", "T", "Y")
)

Arguments

ids

Body ids in any form understood by mcns_ids. If you have a metadata dataframe as returned by mcns_neuprint_meta then this is ideal as that function is called under the hood.

method

A string specifying which of 3 methods to use to identify the group. "all" means to return all 3, while "auto" (the default) means to look at each method in turn successively filling in missing group values.

badtypes

Values of the type column which should be ignored for the purposes of defining cell type groups. This will be because they contain bad values or because the types are too broad to be very useful.

Value

For method="all" a dataframe as returned by

mcns_neuprint_meta with additional columns

instance_group and type_group. Otherwise a numeric vector.

Details

Grouping information for neurons in the male cns is presently scattered in several locations. These include the numeric group field, the type field or the instance field. If the type field has the same value, the neurons should form a group. However there are some values

Examples

# \donttest{
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# return all body ids with a group type or instance
tig_ids=mcns_ids('where:exists(n.group) OR exists(n.type) OR exists (n.instance)')
allg=mcns_predict_group(tig_ids, method = 'all')
# neurons where the recorded group and instance group disagree
allg %>% filter(!is.na(group) & !is.na(instance_group) & group!=instance_group)
#>   bodyid  pre post upstream downstream synweight status           statusLabel
#> 1  11211 1051 5936     5936       6355     12291 Anchor        Primary Anchor
#> 2  11803  992 4404     4404       6047     10451 Anchor        Primary Anchor
#> 3  27083 1212 2047     2047      12146     14193 Anchor        Primary Anchor
#> 4 524150  944 4665     4665       8028     12693 Anchor        Primary Anchor
#> 5  13837  925 3952     3952       7840     11792 Anchor        Primary Anchor
#> 6  57240 1587 3064     3064      15160     18224 Traced Prelim Roughly traced
#> 7 513129  550 1626     1626       4704      6330 Traced Prelim Roughly traced
#>   cropped    name           type      somaLocation nucleusVoxels     voxels
#> 1   FALSE 11211_R           <NA> 45404,48928,28412     103840368 2286457771
#> 2   FALSE 11211_L           <NA> 52290,47120,31620     112753856 1915409289
#> 3   FALSE 27083_L           <NA> 55314,20366,38448     106673824  991224730
#> 4   FALSE 13837_R           <NA> 47814,18265,35894     104159464 1706719281
#> 5   FALSE 13837_L           <NA> 52000,18403,35249      90686240 1390399523
#> 6   FALSE 57240_R           <NA>                NA            NA 1257798925
#> 7   FALSE 30391_L P1_L candidate 55388,14978,35062      82900288  629491986
#>   halfbrainBodyId class group neckLocation cellBodyFiber  soma instance_group
#> 1              NA  <NA> 11803           NA          <NA>  TRUE          11211
#> 2              NA  <NA> 11803           NA          <NA>  TRUE          11211
#> 3              NA  <NA> 12724           NA          <NA>  TRUE          27083
#> 4              NA  <NA> 13165           NA          <NA>  TRUE          13837
#> 5              NA  <NA> 13165           NA          <NA>  TRUE          13837
#> 6              NA  <NA> 24136           NA          <NA> FALSE          57240
#> 7              NA  <NA> 25542           NA          <NA>  TRUE          30391
#>   type_group
#> 1         NA
#> 2         NA
#> 3         NA
#> 4         NA
#> 5         NA
#> 6         NA
#> 7         NA
# }
if (FALSE) {
# neurons where the recorded group and type group disagree
type_group_mismatch <- allg %>% filter(!is.na(group) & !is.na(type_group) & group!=type_group)
allg %>%
  filter(group %in% type_group_mismatch$group | type_group %in% type_group_mismatch$type_group) %>%
 select(bodyid, type, name, group, type_group, instance_group) %>%
 arrange(type, group) %>% View
}