Predict the group of neurons using instance or type information

Predict the group of neurons using instance or type information

Usage

mcns_predict_group(
  ids,
  method = c("auto", "group", "instance", "type", "all"),
  badtypes = c(NA, "", "Lamina_R1-R6", "Descending", "KC", "ER", "LC", "PB",
    "Ascending Interneuron", "Delta", "P1_L candidate", "LT", "MeMe", "PFGs", "Mi", "VT",
    "ML", "EL", "FB", "Dm", "DNp", "FC", "OL", "T", "Y")
)

Arguments

ids

Body ids in any form understood by mcns_ids. If you have a metadata dataframe as returned by mcns_neuprint_meta then this is ideal as that function is called under the hood.

method

A string specifying which of 3 methods to use to identify the group. "all" means to return all 3, while "auto" (the default) means to look at each method in turn successively filling in missing group values.

badtypes

Values of the type column which should be ignored for the purposes of defining cell type groups. This will be because they contain bad values or because the types are too broad to be very useful.

Value

For method="all" a dataframe as returned by

mcns_neuprint_meta with additional columns

instance_group and type_group. Otherwise a numeric vector.

Details

Grouping information for neurons in the male cns is presently scattered in several locations. These include the numeric group field, the type field or the instance field. If the type field has the same value, the neurons should form a group. However there are some values

Examples

# \donttest{
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# return all body ids with a group type or instance
tig_ids=mcns_ids('where:exists(n.group) OR exists(n.type) OR exists (n.instance)')
allg=mcns_predict_group(tig_ids, method = 'all')
# neurons where the recorded group and instance group disagree
allg %>% filter(!is.na(group) & !is.na(instance_group) & group!=instance_group)
#>    bodyid  pre  post upstream downstream synweight status           statusLabel
#> 1   10333 2004 11514    11514      16549     28063 Anchor        Primary Anchor
#> 2  581131 1908 11794    11794      15654     27448 Assign             0.5assign
#> 3   13975  949  2684     2684       9466     12150 Traced Prelim Roughly traced
#> 4   13161 1123  3127     3127      11509     14636 Traced Prelim Roughly traced
#> 5   11211 1051  5936     5936       6355     12291 Anchor        Primary Anchor
#> 6   11803  992  4404     4404       6047     10451 Anchor        Primary Anchor
#> 7   27083 1212  2047     2047      12146     14193 Anchor        Primary Anchor
#> 8   17153  565  5048     5048       5131     10179 Traced Prelim Roughly traced
#> 9  513506  642  4767     4767       5722     10489 Traced Prelim Roughly traced
#> 10 513117  673  4482     4482       5836     10318 Traced Prelim Roughly traced
#> 11 524150  944  4665     4665       8028     12693 Anchor        Primary Anchor
#> 12  13837  957  3965     3965       8138     12103 Anchor        Primary Anchor
#> 13 524005  813  3771     3771       6694     10465 Traced Prelim Roughly traced
#> 14  13501 1001  5644     5644      10177     15821 Anchor        Primary Anchor
#> 15  15753  816  3398     3398       7717     11115 Traced Prelim Roughly traced
#> 16  15699  563  4175     4175       4587      8762 Traced Prelim Roughly traced
#> 17  18186  660  1714     1714       5915      7629 Traced Prelim Roughly traced
#> 18 521738  494  1176     1176       3903      5079 Anchor        Primary Anchor
#> 19  17554  530  1507     1507       4322      5829 Anchor        Primary Anchor
#> 20  19176  490  1310     1310       3924      5234 Anchor        Primary Anchor
#> 21  23242  832   386      386       7799      8185 Anchor        Primary Anchor
#> 22  22047  868   397      397       8283      8680 Anchor        Primary Anchor
#> 23  57240 1708  3131     3131      16351     19482 Traced Prelim Roughly traced
#> 24 513129  550  1626     1626       4704      6330 Traced Prelim Roughly traced
#> 25 522809 1108  3799     3799       9172     12971 Anchor        Primary Anchor
#> 26 523919 1439  3685     3685      16387     20072 Anchor        Primary Anchor
#> 27 529570 1597  4405     4405      18177     22582 Anchor        Primary Anchor
#> 28 574984 1998  2615     2615      18325     20940 Anchor                Anchor
#> 29 555309 2116  2658     2658      19575     22233 Anchor        Primary Anchor
#>    cropped     name                  type somaSide                 class
#> 1    FALSE  10133_R                  <NA>        R                  <NA>
#> 2       NA  10133_L                  <NA>     <NA>                  <NA>
#> 3    FALSE  13161_R                  <NA>        R                  <NA>
#> 4    FALSE  13161_L                  <NA>        L                  <NA>
#> 5    FALSE  11211_R                  <NA>        R                  <NA>
#> 6    FALSE  11211_L                  <NA>        L                  <NA>
#> 7    FALSE  27083_L                  <NA>        L                  <NA>
#> 8    FALSE  16251_R                  <NA>        R                  <NA>
#> 9    FALSE  16251_L                  <NA>        L                  <NA>
#> 10   FALSE  16251_L                  <NA>        L                  <NA>
#> 11   FALSE  13837_R                  <NA>        R                  <NA>
#> 12   FALSE  13837_L                  <NA>        L                  <NA>
#> 13   FALSE 513691_L                  <NA>        L                  <NA>
#> 14   FALSE  13501_R                  <NA>        R                  <NA>
#> 15   FALSE  13095_L                  <NA>        L                  <NA>
#> 16   FALSE  10962_R                  <NA>        R                  <NA>
#> 17   FALSE  13160_L                  <NA>        L                  <NA>
#> 18   FALSE  11568_R Ascending Interneuron     <NA> Ascending Interneuron
#> 19   FALSE  11568_L Ascending Interneuron     <NA> Ascending Interneuron
#> 20   FALSE  11568_L Ascending Interneuron     <NA> Ascending Interneuron
#> 21   FALSE  22047_R Ascending Interneuron     <NA> Ascending Interneuron
#> 22   FALSE  22047_L Ascending Interneuron     <NA> Ascending Interneuron
#> 23   FALSE  57240_R                  <NA>     <NA>                  <NA>
#> 24   FALSE  30391_L        P1_L candidate        L               central
#> 25   FALSE 575975_R                  <NA>        R                  <NA>
#> 26   FALSE  13357_R                  <NA>        R                  <NA>
#> 27   FALSE  13357_L                  <NA>        L                  <NA>
#> 28   FALSE  12229_L                  <NA>        L                  <NA>
#> 29   FALSE  12229_L                  <NA>        L                  <NA>
#>         somaLocation nucleusVoxels     voxels halfbrainBodyId  group
#> 1  36916,39560,22678     133710248 4207313634      4999824097  10333
#> 2  60844,37386,20894     132889584 3710867625              NA  10333
#> 3  33335,18589,36094      98358120 1538548394              NA  10983
#> 4  64646,18676,35154     113251784 1752970686              NA  10983
#> 5  45404,48928,28412     103840368 2286457771              NA  11803
#> 6  52290,47120,31620     112753856 1915409289              NA  11803
#> 7  55314,20366,38448     106673824  991224730              NA  12724
#> 8  22596,26628,20862      82156824  914373442              NA  12980
#> 9  72396,20028,18998      66791336 1015480220              NA  12980
#> 10 73100,21418,18478      64357224 1073210280              NA  12980
#> 11 47814,18265,35894     104159464 1706719281              NA  13165
#> 12 52000,18403,35249      90686240 1401551880              NA  13165
#> 13 56828,15422,13200      84447920 1436832278              NA  13240
#> 14 34240,38982,23488     116267800 2038019649              NA  13289
#> 15 72434,19340,20556      80977056 1062035690              NA  14893
#> 16 36706,15734,14854      81391648 1079221910              NA  15699
#> 17 63162,16752,34430      79477112  801247330              NA  16767
#> 18                NA            NA  731940166              NA  17554
#> 19                NA            NA  850842573              NA  17554
#> 20                NA            NA  723243083              NA  17554
#> 21                NA            NA  574930629              NA  21958
#> 22                NA            NA  585137398              NA  21958
#> 23                NA            NA 1295461283              NA  24136
#> 24 55388,14978,35062      82900288  629491986              NA  25542
#> 25 45910,12914,29880      91371200 1260804377              NA 383941
#> 26 33460,38390,22566     103695616 1585448215              NA 523919
#> 27 62019,35233,17014      93284040 1537129301              NA 523919
#> 28 65336,38724,24508     101708160 2229159941              NA 555309
#> 29 64894,39068,25218     102306560 1973140010              NA 555309
#>           neckLocation cellBodyFiber  soma instance_group type_group
#> 1                   NA          <NA>  TRUE          10133         NA
#> 2                   NA          <NA>  TRUE          10133         NA
#> 3                   NA          <NA>  TRUE          13161         NA
#> 4                   NA          <NA>  TRUE          13161         NA
#> 5                   NA          <NA>  TRUE          11211         NA
#> 6                   NA          <NA>  TRUE          11211         NA
#> 7                   NA          <NA>  TRUE          27083         NA
#> 8                   NA          <NA>  TRUE          16251         NA
#> 9                   NA          <NA>  TRUE          16251         NA
#> 10                  NA          <NA>  TRUE          16251         NA
#> 11                  NA          <NA>  TRUE          13837         NA
#> 12                  NA          <NA>  TRUE          13837         NA
#> 13                  NA          <NA>  TRUE         513691         NA
#> 14                  NA          <NA>  TRUE          13501         NA
#> 15                  NA          <NA>  TRUE          13095         NA
#> 16                  NA          <NA>  TRUE          10962         NA
#> 17                  NA          <NA>  TRUE          13160         NA
#> 18 48899, 47171, 43456          <NA> FALSE          11568         NA
#> 19 48961, 47253, 43456          <NA> FALSE          11568         NA
#> 20 48980, 47381, 43456          <NA> FALSE          11568         NA
#> 21 48394, 45011, 43456          <NA> FALSE          22047         NA
#> 22 49244, 45755, 43456          <NA> FALSE          22047         NA
#> 23                  NA          <NA> FALSE          57240         NA
#> 24                  NA          <NA>  TRUE          30391         NA
#> 25                  NA          <NA>  TRUE         575975         NA
#> 26                  NA          <NA>  TRUE          13357         NA
#> 27                  NA          <NA>  TRUE          13357         NA
#> 28                  NA          <NA>  TRUE          12229         NA
#> 29                  NA          <NA>  TRUE          12229         NA
# }
if (FALSE) {
# neurons where the recorded group and type group disagree
type_group_mismatch <- allg %>% filter(!is.na(group) & !is.na(type_group) & group!=type_group)
allg %>%
  filter(group %in% type_group_mismatch$group | type_group %in% type_group_mismatch$type_group) %>%
 select(bodyid, type, name, group, type_group, instance_group) %>%
 arrange(type, group) %>% View
}