Predict the cell type of male cns neurons from type and name/instance fields

Predict the cell type of male cns neurons from type and name/instance fields

Usage

mcns_predict_type(
  ids,
  method = c("auto", "instance", "type", "foreign_type", "all"),
  prefer.foreign = FALSE
)

Arguments

ids

Body ids in any form understandable by mcns_ids

method

The prediction method to use (type, instance or auto, which uses type when available, instance otherwise). The special value of all adds 3 columns to the metadata data.frame type_t, type_i, type_a containing the results of each of the other methods.

prefer.foreign

Whether to prefer a foreign type (flywire, manc, hemibrain) when available. This may be useful for integration across datasets.

Value

A data.frame when type='all', a character vector otherwise.

Details

For the time being when the instance is of the form 12345_L then 12345 will be returned as the type. This obviously isn't very useful for most purposes (use mcns_predict_group if you want the group).

See also

Examples

# \donttest{
mnm.ti <- mcns_neuprint_meta('where:exists(n.type) OR exists(n.instance)')
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
# Descending neurons with the wrong superclass
mnm.ti %>%
  mutate(ptype=mcns_predict_type(.)) %>%
  filter(grepl("DN[abdgpx]", ptype)) %>%
  filter(superclass!='Descending' | is.na(superclass))
#> Error: object 'mnm.ti' not found

# report all the different types available for these two VPNs
mcns_predict_type("LoVP106", method = 'all') |>
  select(matches('.ype'))
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
# report just one type, preferring malecns type (the default)
mcns_predict_type("LoVP106")
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
# ... or preferring the foreign type (flywire in this case)
mcns_predict_type("LoVP106", prefer.foreign=TRUE)
#> Error in clio_auth(): Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
# }