Evaluate an expression after temporarily setting malevnc options

malecns is a thin wrapper around malevnc which operates by changing the server/dataset options of that package. This function temporarily changes those options, runs the required expression and then sets back again.

codechoose_mcns swaps out the male vnc dataset for the male cns. This means that all functions from the malevnc package should target the male cns instead. It is recommended that you use the with_mcns function to do this temporarily unless you have no intention of using the male vnc dataset.

Usage

with_mcns(expr, dataset = getOption("malecns.dataset"))

choose_mcns(dataset = getOption("malecns.dataset"))

Arguments

expr

An expression to evaluate with a default autosegmentation

dataset

The name of the dataset as reported in Clio e.g. CNS, VNC etc

Examples

if (FALSE) { # \dontrun{
with_mcns(malevnc::manc_dvid_node(type = 'clio'))
} # }